Compare cluster memberships among multiple bioregionalizations

This function computes pairwise comparisons for several bioregionalizations, usually outputs from netclu_, hclu_, or nhclu_ functions. It also provides the confusion matrix from pairwise comparisons, enabling the user to compute additional comparison metrics.

Usage

compare_bioregionalizations(
  cluster_object,
  indices = c("rand", "jaccard"),
  cor_frequency = FALSE,
  store_pairwise_membership = TRUE,
  store_confusion_matrix = TRUE
)

Arguments

cluster_object

A data.frame object where each row corresponds to a site, and each column to a bioregionalization.

indices

NULL or character. Indices to compute for the pairwise comparison of bioregionalizations. Currently available metrics are "rand" and "jaccard".

cor_frequency

A boolean. If TRUE, computes the correlation between each bioregionalization and the total frequency of co-membership of items across all bioregionalizations. This is useful for identifying which bioregionalization(s) is(are) most representative of all computed bioregionalizations.

store_pairwise_membership

A boolean. If TRUE, stores the pairwise membership of items in the output object.

store_confusion_matrix

A boolean. If TRUE, stores the confusion matrices of pairwise bioregionalization comparisons in the output object.

Value

A list containing 4 to 7 elements:

args: A list of user-provided arguments.

inputs: A list containing information on the input bioregionalizations, such as the number of items clustered.

pairwise_membership (optional): If store_pairwise_membership = TRUE, a boolean matrix where TRUE indicates two items are in the same cluster, and FALSE indicates they are not.

freq_item_pw_membership: A numeric vector containing the number of times each item pair is clustered together, corresponding to the sum of rows in pairwise_membership.

bioregionalization_freq_cor (optional): If cor_frequency = TRUE, a numeric vector of correlations between individual bioregionalizations and the total frequency of pairwise membership.

confusion_matrix (optional): If store_confusion_matrix = TRUE, a list of confusion matrices for each pair of bioregionalizations.

bioregionalization_comparison: A data.frame containing comparison results, where the first column indicates the bioregionalizations compared, and the remaining columns contain the requested indices.

Details

This function operates in two main steps:

Within each bioregionalization, the function compares all pairs of items and documents whether they are clustered together (TRUE) or separately (FALSE). For example, if site 1 and site 2 are clustered in the same cluster in bioregionalization 1, their pairwise membership site1_site2 will be TRUE. This output is stored in the pairwise_membership slot if store_pairwise_membership = TRUE.

Across all bioregionalizations, the function compares their pairwise memberships to determine similarity. For each pair of bioregionalizations, it computes a confusion matrix with the following elements:

a: Number of item pairs grouped in both bioregionalizations.

b: Number of item pairs grouped in the first but not in the second bioregionalization.

c: Number of item pairs grouped in the second but not in the first bioregionalization.

d: Number of item pairs not grouped in either bioregionalization.

The confusion matrix is stored in confusion_matrix if store_confusion_matrix = TRUE.

Based on these confusion matrices, various indices can be computed to measure agreement among bioregionalizations. The currently implemented indices are:

Rand index: (a + d) / (a + b + c + d) Measures agreement by considering both grouped and ungrouped item pairs.

Jaccard index: a / (a + b + c) Measures agreement based only on grouped item pairs.

These indices are complementary: the Jaccard index evaluates clustering similarity, while the Rand index considers both clustering and separation. For example, if two bioregionalizations never group the same pairs, their Jaccard index will be 0, but their Rand index may be > 0 due to ungrouped pairs.

Users can compute additional indices manually using the list of confusion matrices.

To identify which bioregionalization is most representative of the others, the function can compute the correlation between the pairwise membership of each bioregionalization and the total frequency of pairwise membership across all bioregionalizations. This is enabled by setting cor_frequency = TRUE.

See also

Author

Boris Leroy (leroy.boris@gmail.com)
Maxime Lenormand (maxime.lenormand@inrae.fr)
Pierre Denelle (pierre.denelle@gmail.com)

Examples

# A simple case with four bioregionalizations of four items
bioregionalizations <- data.frame(matrix(nr = 4, 
                                         nc = 4, 
                                         c(1,2,1,1,1,2,2,1,2,1,3,1,2,1,4,2),
                                         byrow = TRUE))
bioregionalizations
#>   X1 X2 X3 X4
#> 1  1  2  1  1
#> 2  1  2  2  1
#> 3  2  1  3  1
#> 4  2  1  4  2

compare_bioregionalizations(bioregionalizations)
#> 2025-01-16 14:31:46.221678 - Computing pairwise membership comparisons for each bioregionalization...
#> 2025-01-16 14:31:46.222209 - Comparing memberships among bioregionalizations...
#> 2025-01-16 14:31:46.222858 - Computing Rand index...
#> 2025-01-16 14:31:46.223228 - Computing Jaccard index...
#> $args
#> $args$indices
#> [1] "rand"    "jaccard"
#> 
#> $args$cor_frequency
#> [1] FALSE
#> 
#> $args$store_pairwise_membership
#> [1] TRUE
#> 
#> $args$store_confusion_matrix
#> [1] TRUE
#> 
#> 
#> $inputs
#>               number_items number_bioregionalizations 
#>                          4                          4 
#> 
#> $pairwise_membership
#>        X1    X2    X3    X4
#> 1_2  TRUE  TRUE FALSE  TRUE
#> 1_3 FALSE FALSE FALSE  TRUE
#> 1_4 FALSE FALSE FALSE FALSE
#> 2_3 FALSE FALSE FALSE  TRUE
#> 2_4 FALSE FALSE FALSE FALSE
#> 3_4  TRUE  TRUE FALSE FALSE
#> 
#> $freq_item_pw_membership
#> 1_2 1_3 1_4 2_3 2_4 3_4 
#>   3   1   0   1   0   2 
#> 
#> $confusion_matrix
#> $confusion_matrix$`X1%X2`
#> a b c d 
#> 2 0 0 4 
#> 
#> $confusion_matrix$`X1%X3`
#> a b c d 
#> 0 2 0 4 
#> 
#> $confusion_matrix$`X1%X4`
#> a b c d 
#> 1 1 2 2 
#> 
#> $confusion_matrix$`X2%X3`
#> a b c d 
#> 0 2 0 4 
#> 
#> $confusion_matrix$`X2%X4`
#> a b c d 
#> 1 1 2 2 
#> 
#> $confusion_matrix$`X3%X4`
#> a b c d 
#> 0 0 3 3 
#> 
#> 
#> $bioregionalization_comparison
#>   bioregionalization_comparison      rand jaccard
#> 1                         X1%X2 1.0000000    1.00
#> 2                         X1%X3 0.6666667    0.00
#> 3                         X1%X4 0.5000000    0.25
#> 4                         X2%X3 0.6666667    0.00
#> 5                         X2%X4 0.5000000    0.25
#> 6                         X3%X4 0.5000000    0.00
#> 
#> attr(,"class")
#> [1] "bioregion.bioregionalization.comparison"
#> [2] "list"                                   

# Find out which bioregionalizations are most representative
compare_bioregionalizations(bioregionalizations,
                            cor_frequency = TRUE)
#> 2025-01-16 14:31:46.224988 - Computing pairwise membership comparisons for each bioregionalization...
#> 2025-01-16 14:31:46.225458 - Comparing memberships among bioregionalizations...
#> 2025-01-16 14:31:46.22604 - Computing Rand index...
#> 2025-01-16 14:31:46.226365 - Computing Jaccard index...
#> 2025-01-16 14:31:46.226691 - Computing the correlation between each bioregionalization and the vector of frequency of pairwise membership...
#> $args
#> $args$indices
#> [1] "rand"    "jaccard"
#> 
#> $args$cor_frequency
#> [1] TRUE
#> 
#> $args$store_pairwise_membership
#> [1] TRUE
#> 
#> $args$store_confusion_matrix
#> [1] TRUE
#> 
#> 
#> $inputs
#>               number_items number_bioregionalizations 
#>                          4                          4 
#> 
#> $pairwise_membership
#>        X1    X2    X3    X4
#> 1_2  TRUE  TRUE FALSE  TRUE
#> 1_3 FALSE FALSE FALSE  TRUE
#> 1_4 FALSE FALSE FALSE FALSE
#> 2_3 FALSE FALSE FALSE  TRUE
#> 2_4 FALSE FALSE FALSE FALSE
#> 3_4  TRUE  TRUE FALSE FALSE
#> 
#> $freq_item_pw_membership
#> 1_2 1_3 1_4 2_3 2_4 3_4 
#>   3   1   0   1   0   2 
#> 
#> $bioregionalization_freq_cor
#>        X1        X2        X3        X4 
#> 0.8834522 0.8834522 0.0000000 0.4685213 
#> 
#> $confusion_matrix
#> $confusion_matrix$`X1%X2`
#> a b c d 
#> 2 0 0 4 
#> 
#> $confusion_matrix$`X1%X3`
#> a b c d 
#> 0 2 0 4 
#> 
#> $confusion_matrix$`X1%X4`
#> a b c d 
#> 1 1 2 2 
#> 
#> $confusion_matrix$`X2%X3`
#> a b c d 
#> 0 2 0 4 
#> 
#> $confusion_matrix$`X2%X4`
#> a b c d 
#> 1 1 2 2 
#> 
#> $confusion_matrix$`X3%X4`
#> a b c d 
#> 0 0 3 3 
#> 
#> 
#> $bioregionalization_comparison
#>   bioregionalization_comparison      rand jaccard
#> 1                         X1%X2 1.0000000    1.00
#> 2                         X1%X3 0.6666667    0.00
#> 3                         X1%X4 0.5000000    0.25
#> 4                         X2%X3 0.6666667    0.00
#> 5                         X2%X4 0.5000000    0.25
#> 6                         X3%X4 0.5000000    0.00
#> 
#> attr(,"class")
#> [1] "bioregion.bioregionalization.comparison"
#> [2] "list"