Calculate contribution metrics of sites and species

Cz metrics

Usage

contribution(cluster_object, comat, indices = c("contribution", "Cz"))

Arguments

cluster_object

a bioregion.clusters object or a data.frame or a list of data.frame containing multiple partitions. At least two partitions are required. If a list of data.frame is provided, they should all have the same number of rows (i.e., same items in the clustering for all partitions).

comat

a co-occurrence matrix with sites as rows and species as columns.

indices

a character specifying the contribution metric to compute. Available options are contribution.

Value

A list of data.frames if multiples indices are selected or a single data.frame with three columns if one index is selected. Each data.frame has three columns: the species, the bioregion, and the contribution statistics.

Details

The contribution metric is derived from Lenormand2019bioregion. Its formula is the following: \((n_ij - ((n_i n_j)/n))/(sqrt(((n - n_j)/(n-1)) (1-(n_j/n)) ((n_i n_j)/n)))\)

with n the number of sites, n_i the number of sites in which species i is present, n_j the number of sites belonging to the bioregion j, n_ij the number of occurrences of species i in sites belonging to the bioregion j.

See also

Author

Pierre Denelle (pierre.denelle@gmail.com), Boris Leroy (leroy.boris@gmail.com) and Maxime Lenormand (maxime.lenormand@inrae.fr)

Examples

comat <- matrix(sample(1000, 50), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

comat <- matrix(sample(0:1000, size = 500, replace = TRUE, prob = 1/1:1001),
                20, 25)
rownames(comat) <- paste0("Site",1:20)
colnames(comat) <- paste0("Species",1:25)

dissim <- dissimilarity(comat, metric = "Simpson")
clust1 <- nhclu_kmeans(dissim, n_clust = 3, index = "Simpson")

net <- similarity(comat, metric = "Simpson")
com <- netclu_greedy(net)

contribution(cluster_object = clust1, comat = comat,
indices = "contribution")
#>    Bioregion   Species Contribution
#> 1          1  Species1    14.784446
#> 2          1  Species2   895.703803
#> 3          1  Species3   483.131789
#> 4          1  Species4   110.343071
#> 5          1  Species5   139.024408
#> 6          1  Species6   188.140523
#> 7          1  Species7  1250.813685
#> 8          1  Species8   452.066063
#> 9          1  Species9    94.212790
#> 10         1 Species10   846.959241
#> 11         1 Species11    81.361595
#> 12         1 Species12   680.112367
#> 13         1 Species13   577.523805
#> 14         1 Species14   785.206656
#> 15         1 Species15   624.122394
#> 16         1 Species16   142.603633
#> 17         1 Species17  1334.127367
#> 18         1 Species18   140.556556
#> 19         1 Species19  1325.490913
#> 20         1 Species20   234.860601
#> 21         1 Species21   958.958438
#> 22         1 Species22   892.442921
#> 23         1 Species23   517.184605
#> 24         1 Species24   879.105881
#> 25         1 Species25   577.269590
#> 26         2  Species1  1463.629119
#> 27         2  Species2   388.500675
#> 28         2  Species3   453.862764
#> 29         2  Species4   123.848681
#> 30         2  Species5  1142.427787
#> 31         2  Species6   243.057196
#> 32         2  Species7   390.350620
#> 33         2  Species8   325.593140
#> 34         2  Species9  1569.061298
#> 35         2 Species10   319.366459
#> 36         2 Species11    63.930518
#> 37         2 Species12  1063.286656
#> 38         2 Species13   358.594548
#> 39         2 Species14  1094.818182
#> 40         2 Species15   253.986310
#> 41         2 Species16   598.944427
#> 42         2 Species17  1776.030303
#> 43         2 Species18  2159.062626
#> 44         2 Species19   678.023274
#> 45         2 Species20   223.835591
#> 46         2 Species21  1176.113184
#> 47         2 Species22   311.595231
#> 48         2 Species23   605.793776
#> 49         2 Species24   787.028620
#> 50         2 Species25   623.004257
#> 51         3  Species1    12.289814
#> 52         3  Species2     7.664615
#> 53         3  Species3   220.506226
#> 54         3  Species4   567.254514
#> 55         3  Species5    97.939011
#> 56         3  Species6   375.037542
#> 57         3  Species7   217.295594
#> 58         3  Species8   225.643238
#> 59         3  Species9   306.551171
#> 60         3 Species10   713.659346
#> 61         3 Species11   142.209078
#> 62         3 Species12   160.822107
#> 63         3 Species13   198.031800
#> 64         3 Species14   618.250000
#> 65         3 Species15   386.816979
#> 66         3 Species16   194.821168
#> 67         3 Species17    40.750000
#> 68         3 Species18   400.227743
#> 69         3 Species19   378.469335
#> 70         3 Species20   279.890597
#> 71         3 Species21   616.484742
#> 72         3 Species22   264.122033
#> 73         3 Species23   624.403768
#> 74         3 Species24   377.680513
#> 75         3 Species25    79.024824

contribution(cluster_object = com, comat = comat,
indices = "contribution")
#>    Bioregion   Species  Contribution
#> 1          1  Species1  7435.4146567
#> 2          1  Species2 10054.7994071
#> 3          1  Species3  8871.3616768
#> 4          1  Species4  5972.2238739
#> 5          1  Species5 10469.0000000
#> 6          1  Species6  6072.3644440
#> 7          1  Species7 14424.5069317
#> 8          1  Species8  6165.4997274
#> 9          1  Species9 14226.5170330
#> 10         1 Species10 14325.5119823
#> 11         1 Species11  2174.0000000
#> 12         1 Species12 14597.0517571
#> 13         1 Species13  8729.9403205
#> 14         1 Species14 18247.4980563
#> 15         1 Species15  9625.6089100
#> 16         1 Species16  6754.7553784
#> 17         1 Species17 22441.2176715
#> 18         1 Species18 20231.9430604
#> 19         1 Species19 16644.8222246
#> 20         1 Species20  5606.6960440
#> 21         1 Species21 20987.7340368
#> 22         1 Species22 11309.0000000
#> 23         1 Species23 13255.4237201
#> 24         1 Species24 15492.4481204
#> 25         1 Species25  9831.6666321
#> 26         2  Species1   883.7537604
#> 27         2  Species2     2.3925840
#> 28         2  Species3    -0.9733285
#> 29         2  Species4     0.1081476
#> 30         2  Species5    -0.9176629
#> 31         2  Species6     1.2797542
#> 32         2  Species7    -0.9733285
#> 33         2  Species8   349.4249411
#> 34         2  Species9   148.2703789
#> 35         2 Species10     3.3525760
#> 36         2 Species11    -0.9176629
#> 37         2 Species12     0.0000000
#> 38         2 Species13    -0.9733285
#> 39         2 Species14   177.9473684
#> 40         2 Species15    14.1673374
#> 41         2 Species16    76.8929536
#> 42         2 Species17   420.0526316
#> 43         2 Species18    36.9352207
#> 44         2 Species19   395.9284152
#> 45         2 Species20    -0.9459053
#> 46         2 Species21     2.0519567
#> 47         2 Species22     7.1118877
#> 48         2 Species23    -0.9733285
#> 49         2 Species24    40.2287961
#> 50         2 Species25     4.6182435