Download, unzip, check permission and test the bioregion's binary files

This function downloads and unzips the 'bin' folder needed to run some functions of bioregion. It also checks if the files have the permissions to be executed as programs. It finally tests if the binary files are running properly.

Usage

install_binaries(
  binpath = "tempdir",
  download_only = FALSE,
  infomap_version = c("2.1.0", "2.6.0", "2.7.1", "2.8.0")
)

Arguments

binpath

a character indicating the path to the folder that will host the 'bin' folder containing the binary files (see Details).

download_only

a boolean indicating whether the function should only download the 'bin.zip' file or execute the entire process (see Details).

infomap_version

a character vector indicating the Infomap version(s) to install.

Value

No return value

Details

By default, the binary files are installed in R's temporary directory (binpath = "tempdir"). In this case the bin folder will be automatically removed at the end of the R session. Alternatively, the binary files can be installed in the bioregion's package folder (binpath = "pkgfolder").

Finally, a path to a folder of your choice can be specified. In this case , and only in this case, download_only can be set to TRUE, but you must ensure that the files have the necessary permissions to be executed as programs.

In any case, PLEASE MAKE SURE to update the binpath and check_install accordingly in netclu_infomap, netclu_louvain and netclu_oslom).

Note

Only the Infomap version 2.1.0, 2.6.0, 2.7.1 and 2.8.0 are available for now.

Author

Maxime Lenormand (maxime.lenormand@inrae.fr), Boris Leroy (leroy.boris@gmail.com) and Pierre Denelle (pierre.denelle@gmail.com)