Create a map of bioregions

This plot function can be used to visualise bioregions based on a bioregion.clusters object combined with a geometry (sf objects).

Usage

map_clusters(clusters, geometry, write_clusters = FALSE, plot = TRUE, ...)

Arguments

clusters

an object of class bioregion.clusters or a data.frame. If a data.frame is used, the first column should represent the sites' ID, and the next column(s) the clusters.

geometry

a spatial object that can be handled by the sf package. The first attribute should correspond to the sites' ID (see Details).

write_clusters

a boolean indicating if the clusters should be added in geometry.

plot

a boolean indicating if the plot should be drawn.

...

further arguments to be passed to sf::plot()

Value

One or several maps of bioregions if plot = TRUE and the geometry with additional clusters' attributes if write_clusters = TRUE.

Details

The clusters and geometry site IDs should correspond. They should have the same type (i.e. character is cluster is a bioregion.clusters object) and the site of clusters should be included in the sites of geometry.

Author

Maxime Lenormand (maxime.lenormand@inrae.fr), Boris Leroy (leroy.boris@gmail.com) and Pierre Denelle (pierre.denelle@gmail.com)

Examples

data(fishmat)
data(fishsf)

net <- similarity(fishmat, metric = "Simpson")
clu <- netclu_greedy(net)
map <- map_clusters(clu, fishsf, write_clusters = TRUE, plot = FALSE)