Calculate contribution metrics for bioregions

This function calculates metrics at the bioregion scale.

This function calculates metrics that assess the contribution of a given species to its bioregion.

Usage

contribution(
  cluster_object,
  comat,
  indices = c("rho", "Cz"),
  bipartite_link = NULL
)

contribution(
  cluster_object,
  comat,
  indices = c("rho", "Cz"),
  bipartite_link = NULL
)

Arguments

cluster_object

a bioregion.clusters object or a data.frame or a list of data.frame containing multiple partitions. At least two partitions are required. If a list of data.frame is provided, they should all have the same number of rows (i.e., same items in the clustering for all partitions).

comat

a co-occurrence matrix with sites as rows and species as columns.

indices

a character specifying the contribution metric to compute. Available options are rho and Cz.

bipartite_link

NULL by default. Needed for Cz indices. A data.frame where each row represents the interaction between two nodes and an optional third column indicating the weight of the interaction.

Value

A list of data.frames if multiples indices are selected or a single data.frame with three columns if one index is selected. Each data.frame has three columns: the species, the bioregion, and the contribution statistics.

A list of data.frames if multiples indices are selected or a single data.frame with three columns if one index is selected. Each data.frame has three columns: the species, the bioregion, and the desired summary statistics.

Details

The contribution metric is derived from Lenormand2019bioregion. Its formula is the following: \((n_ij - ((n_i n_j)/n))/(sqrt(((n - n_j)/(n-1)) (1-(n_j/n)) ((n_i n_j)/n)))\)

with \(n\) the number of sites, \(n_i\) the number of sites in which species \(i\) is present, \(n_j\) the number of sites belonging to the bioregion \(j\), \(n_ij\) the number of occurrences of species \(i\) in sites belonging to the bioregion \(j\).

Cz metrics are derived from Guimera2005bioregion. Their respective formula are: \(C_i = 1 - \sum_{s=1}^{N_M}{{(\frac{k_is}{k_i}})^2}\)

where \(k_{is}\) is the number of links of node (species or site) \(i\) to nodes in bioregion \(s\), and \(k_i\) is the total degree of node \(i\). The participation coefficient of a node is therefore close to 1 if its links are uniformly distributed among all the bioregions and 0 if all its links are within its own bioregion.

And: \(z_i = \frac{k_i - \overline{k_{si}}}{\sigma_{k_{si}}}\)

where \(k_i\) is the number of links of node (species or site) \(i\) to other nodes in its bioregion \(s_i\), \(\overline{k_{si}}\) is the average of \(k\) over all the nodes in \(s_i\), and \(\sigma_{k_{si}}\) is the standard deviation of \(k\) in \(s_i\). The within-bioregion degree z-score measures how well-connected node \(i\) is to other nodes in the bioregion.

The metric is derived from Lenormand2019bioregion. Its formula is the following: \(\rho_{ij} = (n_ij - ((n_i n_j)/n))/(sqrt(((n - n_j)/(n-1)) (1-(n_j/n)) ((n_i n_j)/n)))\)

with \(n\) the number of sites, \(n_i\) the number of sites in which species \(i\) is present, \(n_j\) the number of sites belonging to the bioregion \(j\), \(n_ij\) the number of occurrences of species \(i\) in sites belonging to the bioregion \(j\).

Cz metrics are derived from Guimera2005bioregion. Their respective formula are: \(C_i = 1 - \sum_{s=1}^{N_M}{{(\frac{k_is}{k_i}})^2}\)

where \(k_{is}\) is the number of links of node (species or site) \(i\) to nodes in bioregion \(s\), and \(k_i\) is the total degree of node \(i\). The participation coefficient of a node is therefore close to 1 if its links are uniformly distributed among all the bioregions and 0 if all its links are within its own bioregion.

And: \(z_i = \frac{k_i - \overline{k_{si}}}{\sigma_{k_{si}}}\)

where \(k_i\) is the number of links of node (species or site) \(i\) to other nodes in its bioregion \(s_i\), \(\overline{k_{si}}\) is the average of \(k\) over all the nodes in \(s_i\), and \(\sigma_{k_{si}}\) is the standard deviation of \(k\) in \(s_i\). The within-bioregion degree z-score measures how well-connected node \(i\) is to other nodes in the bioregion.

See also

Author

Pierre Denelle (pierre.denelle@gmail.com), Boris Leroy (leroy.boris@gmail.com) and Maxime Lenormand (maxime.lenormand@inrae.fr)

Examples

comat <- matrix(sample(1000, 50), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

comat <- matrix(sample(0:1000, size = 500, replace = TRUE, prob = 1/1:1001),
                20, 25)
rownames(comat) <- paste0("Site",1:20)
colnames(comat) <- paste0("Species",1:25)

dissim <- dissimilarity(comat, metric = "Simpson")
clust1 <- nhclu_kmeans(dissim, n_clust = 3, index = "Simpson")

net <- similarity(comat, metric = "Simpson")
com <- netclu_greedy(net)

contribution(cluster_object = clust1, comat = comat,
indices = "contribution")
#> Error: Please choose algorithm among the followings values:
#>     rho or Cz.

contribution(cluster_object = com, comat = comat,
indices = "contribution")
#> Error: Please choose algorithm among the followings values:
#>     rho or Cz.

# Cz indices
net_bip <- mat_to_net(comat, weight = TRUE)
clust_bip <- netclu_greedy(net_bip, bipartite = TRUE)
contribution(cluster_object = clust_bip, comat = comat, 
bipartite_link = net_bip, indices = "Cz")
#> Warning: Column 'Node1' has been renamed 'Sites'. If this column does not
#>             correspond to sites, rename it before running the function.
#> Warning: Column 'Node2' has been renamed 'Species'. If this column does not
#>             correspond to species, rename it before running the function.
#>         Node Bioregion Category         C          z
#> 1      Site1         4     site 0.8194444 -0.5720776
#> 2      Site2         4     site 0.8204159  0.8581163
#> 3      Site3         5     site 0.8229167  1.2421180
#> 4      Site4         1     site 0.8090737        NaN
#> 5      Site5         3     site 0.8100000  0.8250645
#> 6      Site6         1     site 0.8181818        NaN
#> 7      Site7         1     site 0.8117914        NaN
#> 8      Site8         2     site 0.8192000        NaN
#> 9      Site9         3     site 0.8159722  1.5001172
#> 10    Site10         4     site 0.8204159  0.8581163
#> 11    Site11         4     site 0.8090737 -0.5720776
#> 12    Site12         5     site 0.8264463  1.2421180
#> 13    Site13         3     site 0.8090737  1.5001172
#> 14    Site14         6     site 0.7850000 -0.2070197
#> 15    Site15         6     site 0.8100000  1.2421180
#> 16    Site16         2     site 0.8223140        NaN
#> 17    Site17         5     site 0.8159722 -0.2070197
#> 18    Site18         1     site 0.8204159        NaN
#> 19    Site19         4     site 0.8199446 -0.5720776
#> 20    Site20         6     site 0.8253968  1.2421180
#> 21  Species1         2  species 0.8144044        NaN
#> 22  Species2         5  species 0.8125000 -1.6561573
#> 23  Species4         3  species 0.8235294 -0.5250410
#> 24  Species5         1  species 0.8200000        NaN
#> 25  Species6         6  species 0.8281250 -1.6561573
#> 26  Species7         4  species 0.8024691  0.8581163
#> 27  Species8         3  species 0.8203125 -0.5250410
#> 28  Species9         5  species 0.8200000 -0.2070197
#> 29 Species10         4  species 0.8209877 -0.5720776
#> 30 Species11         6  species 0.8144044 -0.2070197
#> 31 Species12         5  species 0.8148148 -0.2070197
#> 32 Species13         1  species 0.8200000        NaN
#> 33 Species14         4  species 0.8096886  0.8581163
#> 34 Species15         1  species 0.8200000        NaN
#> 35 Species16         4  species 0.8088643  0.8581163
#> 36 Species17         1  species 0.8209877        NaN
#> 37 Species18         6  species 0.8148148 -0.2070197
#> 38 Species19         3  species 0.8209877 -0.5250410
#> 39 Species20         5  species 0.8235294 -0.2070197
#> 40 Species21         6  species 0.8148148 -0.2070197
#> 41 Species22         2  species 0.8088889        NaN
#> 42 Species23         3  species 0.8144044 -0.5250410
#> 43 Species24         3  species 0.7911111 -1.2000938
#> 44 Species25         3  species 0.8148148 -0.5250410
#> 45  Species3         4  species 0.8235294 -2.0022714

comat <- matrix(sample(1000, 50), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

comat <- matrix(sample(0:1000, size = 500, replace = TRUE, prob = 1/1:1001),
                20, 25)
rownames(comat) <- paste0("Site",1:20)
colnames(comat) <- paste0("Species",1:25)

dissim <- dissimilarity(comat, metric = "Simpson")
clust1 <- nhclu_kmeans(dissim, n_clust = 3, index = "Simpson")

net <- similarity(comat, metric = "Simpson")
com <- netclu_greedy(net)

contribution(cluster_object = clust1, comat = comat, indices = "rho")
#>    Bioregion   Species         rho
#> 1          1  Species1  357.696065
#> 2          1  Species2 1169.140069
#> 3          1  Species3 1347.390766
#> 4          1  Species4 1043.388651
#> 5          1  Species5  465.697017
#> 6          1  Species6  971.845821
#> 7          1  Species7  479.542282
#> 8          1  Species8  437.306173
#> 9          1  Species9  482.247343
#> 10         1 Species10 1072.960810
#> 11         1 Species11  488.571898
#> 12         1 Species12  923.531266
#> 13         1 Species13  576.372709
#> 14         1 Species14  121.355361
#> 15         1 Species15  751.868336
#> 16         1 Species16  168.724543
#> 17         1 Species17 1270.981983
#> 18         1 Species18 1842.026737
#> 19         1 Species19 1290.186963
#> 20         1 Species20 1123.399973
#> 21         1 Species21  888.267749
#> 22         1 Species22 1030.357644
#> 23         1 Species23 1005.415924
#> 24         1 Species24 1399.307865
#> 25         1 Species25  826.916473
#> 26         2  Species1  705.018124
#> 27         2  Species2  586.034926
#> 28         2  Species3  646.537358
#> 29         2  Species4  468.218678
#> 30         2  Species5  721.003182
#> 31         2  Species6  277.211934
#> 32         2  Species7   42.577395
#> 33         2  Species8  206.760731
#> 34         2  Species9  396.082174
#> 35         2 Species10  338.838834
#> 36         2 Species11  648.310642
#> 37         2 Species12 1297.623220
#> 38         2 Species13  313.259537
#> 39         2 Species14  168.155278
#> 40         2 Species15   33.252824
#> 41         2 Species16  434.504052
#> 42         2 Species17  506.334057
#> 43         2 Species18  201.097047
#> 44         2 Species19  567.263431
#> 45         2 Species20  320.994427
#> 46         2 Species21  656.703882
#> 47         2 Species22 1056.552507
#> 48         2 Species23   50.563228
#> 49         2 Species24   27.578172
#> 50         2 Species25  806.476853
#> 51         3  Species1  118.910891
#> 52         3  Species2  594.088083
#> 53         3  Species3   11.696153
#> 54         3  Species4  773.955547
#> 55         3  Species5  341.511234
#> 56         3  Species6  196.522616
#> 57         3  Species7  234.941453
#> 58         3  Species8  619.669569
#> 59         3  Species9  852.240042
#> 60         3 Species10  711.516830
#> 61         3 Species11  541.575664
#> 62         3 Species12   65.628415
#> 63         3 Species13  145.877818
#> 64         3 Species14  130.962545
#> 65         3 Species15  724.155108
#> 66         3 Species16  460.206130
#> 67         3 Species17  576.173238
#> 68         3 Species18  195.134199
#> 69         3 Species19  512.031596
#> 70         3 Species20  339.838229
#> 71         3 Species21 1213.334098
#> 72         3 Species22  572.461721
#> 73         3 Species23  943.846918
#> 74         3 Species24 1385.318648
#> 75         3 Species25   -2.342838

contribution(cluster_object = com, comat = comat, indices = "rho")
#>    Bioregion   Species           rho
#> 1          1  Species1  8934.0000000
#> 2          1  Species2 17566.5614639
#> 3          1  Species3 13909.0000000
#> 4          1  Species4 14294.3221328
#> 5          1  Species5  9414.5053570
#> 6          1  Species6 10699.9441697
#> 7          1  Species7  5613.9564381
#> 8          1  Species8  8635.6594164
#> 9          1  Species9 13067.2907855
#> 10         1 Species10 15811.1029626
#> 11         1 Species11 11309.9662641
#> 12         1 Species12 17099.0000000
#> 13         1 Species13  7642.8803471
#> 14         1 Species14  3186.1905061
#> 15         1 Species15 11286.8803823
#> 16         1 Species16  7978.8171161
#> 17         1 Species17 16347.8373765
#> 18         1 Species18 16375.4656679
#> 19         1 Species19 17504.0000000
#> 20         1 Species20 13239.0000000
#> 21         1 Species21 19678.6130086
#> 22         1 Species22 19919.0000000
#> 23         1 Species23 14668.6059324
#> 24         1 Species24 21014.2518068
#> 25         1 Species25  7902.5363455
#> 26         2  Species1     0.2294157
#> 27         2  Species2    -0.8885233
#> 28         2  Species3   213.5860482
#> 29         2  Species4   650.0595091
#> 30         2  Species5   490.9248522
#> 31         2  Species6    -0.9459053
#> 32         2  Species7     0.1081476
#> 33         2  Species8   205.5886144
#> 34         2  Species9     8.4736842
#> 35         2 Species10    12.6842105
#> 36         2 Species11   319.0000000
#> 37         2 Species12    20.8768318
#> 38         2 Species13    16.4156536
#> 39         2 Species14    -0.9459053
#> 40         2 Species15    -0.9459053
#> 41         2 Species16    29.7985719
#> 42         2 Species17   275.8845635
#> 43         2 Species18     8.4736842
#> 44         2 Species19    36.9359332
#> 45         2 Species20    -0.9176629
#> 46         2 Species21   265.0204093
#> 47         2 Species22    20.8768318
#> 48         2 Species23    63.6106578
#> 49         2 Species24     2.1578947
#> 50         2 Species25   982.8690734

# Cz indices
net_bip <- mat_to_net(comat, weight = TRUE)
clust_bip <- netclu_greedy(net_bip, bipartite = TRUE)
contribution(cluster_object = clust_bip, comat = comat, 
bipartite_link = net_bip, indices = "Cz")
#> Warning: Column 'Node1' has been renamed 'Sites'. If this column does not
#>             correspond to sites, rename it before running the function.
#> Warning: Column 'Node2' has been renamed 'Species'. If this column does not
#>             correspond to species, rename it before running the function.
#>         Node Bioregion Category         C          z
#> 1      Site1         7     site 0.8506616 -0.7302967
#> 2      Site2         6     site 0.8347107 -0.6454972
#> 3      Site3         1     site 0.8181818  1.3333333
#> 4      Site4         2     site 0.8355388 -1.2421180
#> 5      Site5         7     site 0.8305785 -0.7302967
#> 6      Site6         2     site 0.8317580 -1.2421180
#> 7      Site7         6     site 0.8512397  1.2909944
#> 8      Site8         3     site 0.8355388  1.7888544
#> 9      Site9         5     site 0.8480726  1.4142136
#> 10    Site10         2     site 0.8468809  0.2070197
#> 11    Site11         5     site 0.7959184  0.0000000
#> 12    Site12         4     site 0.8299320        NaN
#> 13    Site13         2     site 0.8429752  0.2070197
#> 14    Site14         4     site 0.8429752        NaN
#> 15    Site15         4     site 0.8347107        NaN
#> 16    Site16         4     site 0.8300000        NaN
#> 17    Site17         1     site 0.8355388  1.3333333
#> 18    Site18         7     site 0.8271605 -0.7302967
#> 19    Site19         6     site 0.8305785 -0.6454972
#> 20    Site20         1     site 0.8208617  1.3333333
#> 21  Species1         6  species 0.8281250 -0.6454972
#> 22  Species2         3  species 0.8355556 -0.4472136
#> 23  Species3         4  species 0.8125000        NaN
#> 24  Species4         2  species 0.8373702  0.2070197
#> 25  Species5         4  species 0.8235294        NaN
#> 26  Species6         5  species 0.8304498 -1.4142136
#> 27  Species7         3  species 0.8333333 -0.4472136
#> 28  Species8         3  species 0.8333333 -0.4472136
#> 29  Species9         1  species 0.8310249 -0.6666667
#> 30 Species10         2  species 0.8421053  0.2070197
#> 31 Species11         1  species 0.8421053 -0.6666667
#> 32 Species12         6  species 0.8125000  1.2909944
#> 33 Species13         1  species 0.8400000 -0.6666667
#> 34 Species14         1  species 0.8442907 -0.6666667
#> 35 Species15         3  species 0.8304498 -0.4472136
#> 36 Species17         4  species 0.8271605        NaN
#> 37 Species18         5  species 0.8421053  0.0000000
#> 38 Species19         6  species 0.8281250 -0.6454972
#> 39 Species20         5  species 0.8437500  0.0000000
#> 40 Species21         7  species 0.8304498  1.0954451
#> 41 Species23         2  species 0.8400000  1.6561573
#> 42 Species24         7  species 0.8365651  1.0954451
#> 43 Species25         4  species 0.7692308        NaN
#> 44 Species16         1  species 0.8367347 -0.6666667
#> 45 Species22         1  species 0.8046875 -0.6666667