Convert dissimilarity metrics to similarity metrics

This function converts a data.frame of dissimilarity metrics (beta diversity) between sites to similarity metrics.

Usage

dissimilarity_to_similarity(dissimilarity, include_formula = TRUE)

Arguments

dissimilarity

the output object from dissimilarity() or similarity_to_dissimilarity().

include_formula

a boolean indicating if the metrics based on your own formula(s) should be converted (see Details). This argument is set to TRUE by default.

Value

A data.frame with additional class bioregion.pairwise.metric, providing similarity metric(s) between each pair of sites based on a dissimilarity object.

Note

The behavior of this function changes depending on column names. Columns Site1 and Site2 are copied identically. If there are columns called a, b, c, A, B, C they will also be copied identically. If there are columns based on your own formula (argument formula in dissimilarity()) or not in the original list of dissimilarity metrics (argument metrics in dissimilarity()) and if the argument include_formula is set to FALSE, they will also be copied identically. Otherwise there are going to be converted like they other columns (default behavior).

If a column is called Euclidean, the similarity will be calculated based on the following formula:

Euclidean similarity = 1 / (1 - Euclidean distance)Euclidean similarity = 1 / (1 - Euclidean distance)

Otherwise, all other columns will be transformed into dissimilarity with the following formula:

similarity = 1 - dissimilaritysimilarity = 1 - dissimilarity

See also

Author

Maxime Lenormand (maxime.lenormand@inrae.fr), Boris Leroy (leroy.boris@gmail.com) and Pierre Denelle (pierre.denelle@gmail.com)

Examples

comat <- matrix(sample(0:1000, size = 50, replace = TRUE,
prob = 1 / 1:1001), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

dissimil <- dissimilarity(comat, metric = "all")
dissimil
#> Data.frame of dissimilarity between sites
#>  - Total number of sites:  5 
#>  - Total number of species:  10 
#>  - Number of rows:  10 
#>  - Number of dissimilarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn   Sorensen   Simpson      Bray  Brayturn
#> 2  Site1 Site2 0.3333333   0.2500000 0.20000000 0.1428571 0.8281349 0.7966867
#> 3  Site1 Site3 0.4000000   0.2500000 0.25000000 0.1428571 0.8075571 0.6701807
#> 4  Site1 Site4 0.3333333   0.2500000 0.20000000 0.1428571 0.6144890 0.5512048
#> 5  Site1 Site5 0.3000000   0.0000000 0.17647059 0.0000000 0.7386162 0.6671687
#> 8  Site2 Site3 0.3000000   0.2222222 0.17647059 0.1250000 0.8157999 0.7442117
#> 9  Site2 Site4 0.2222222   0.2222222 0.12500000 0.1250000 0.9083287 0.9070295
#> 10 Site2 Site5 0.2000000   0.0000000 0.11111111 0.0000000 0.7487073 0.7320838
#> 14 Site3 Site4 0.3000000   0.2222222 0.17647059 0.1250000 0.8243785 0.7517007
#> 15 Site3 Site5 0.1000000   0.0000000 0.05263158 0.0000000 0.6271315 0.5209348
#> 20 Site4 Site5 0.2000000   0.0000000 0.11111111 0.0000000 0.4300681 0.3832200
#>    Euclidean a b c   A    B    C
#> 2   640.6114 6 1 2 135  529  772
#> 3   987.6548 6 1 3 219  445 1393
#> 4   462.2315 6 1 2 298  366  584
#> 5   567.9709 7 0 3 221  443  806
#> 8   978.6434 7 1 2 232  675 1380
#> 9   643.2379 7 1 1  82  825  800
#> 10  569.1959 8 0 2 243  664  784
#> 14  950.6156 7 2 1 219 1393  663
#> 15  814.4403 9 0 1 492 1120  535
#> 20  301.2424 8 0 2 544  338  483

similarity <- dissimilarity_to_similarity(dissimil)
similarity
#> Data.frame of similarity between sites
#>  - Total number of sites:  5 
#>  - Total number of species:  10 
#>  - Number of rows:  10 
#>  - Number of similarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn  Sorensen   Simpson       Bray   Brayturn
#> 2  Site1 Site2 0.6666667   0.7500000 0.8000000 0.8571429 0.17186505 0.20331325
#> 3  Site1 Site3 0.6000000   0.7500000 0.7500000 0.8571429 0.19244288 0.32981928
#> 4  Site1 Site4 0.6666667   0.7500000 0.8000000 0.8571429 0.38551100 0.44879518
#> 5  Site1 Site5 0.7000000   1.0000000 0.8235294 1.0000000 0.26138380 0.33283133
#> 8  Site2 Site3 0.7000000   0.7777778 0.8235294 0.8750000 0.18420008 0.25578831
#> 9  Site2 Site4 0.7777778   0.7777778 0.8750000 0.8750000 0.09167132 0.09297052
#> 10 Site2 Site5 0.8000000   1.0000000 0.8888889 1.0000000 0.25129266 0.26791621
#> 14 Site3 Site4 0.7000000   0.7777778 0.8235294 0.8750000 0.17562149 0.24829932
#> 15 Site3 Site5 0.9000000   1.0000000 0.9473684 1.0000000 0.37286851 0.47906524
#> 20 Site4 Site5 0.8000000   1.0000000 0.8888889 1.0000000 0.56993190 0.61678005
#>      Euclidean a b c   A    B    C
#> 2  0.001558576 6 1 2 135  529  772
#> 3  0.001011475 6 1 3 219  445 1393
#> 4  0.002158748 6 1 2 298  366  584
#> 5  0.001757559 7 0 3 221  443  806
#> 8  0.001020780 7 1 2 232  675 1380
#> 9  0.001552222 7 1 1  82  825  800
#> 10 0.001753783 8 0 2 243  664  784
#> 14 0.001050844 7 2 1 219 1393  663
#> 15 0.001226331 9 0 1 492 1120  535
#> 20 0.003308602 8 0 2 544  338  483