Convert dissimilarity indices to similarity indices

This function converts a data.frame of dissimilarity indices (beta diversity) between sites to similarity indices.

Usage

dissimilarity_to_similarity(dissimilaritydata)

Arguments

dissimilaritydata

the output object from similarity_to_dissimilarity() or a data.frame with the first columns called "Site1" and "Site2", and the other columns being the similarity metrics.

Value

A data.frame with additional class bioregion.pairwise.metric, providing one or several dissimilarity metric(s) between each pair of sites. The two first columns represent each pair of sites, and the other column represent dissimilarity metrics. Columns with names "a", "b", "c", "A", "B" and "C" are not altered.

Note

The behaviour of this function changes depending on column names. Columns "Site1" and "Site2" are copied identically. If there are columns called "a", "b", "c", "A", "B", "C", they will also be copied identically.

If a column is called "Euclidean", its similarity will be calculated based on the following formula:

\(Euclidean similarity = (1 - Euclidean distance) / Euclidean distance\)

Otherwise, all other columns will be transformed into similarity with the following formula:

\(similarity = 1 - dissimilarity\)

See also

Author

Boris Leroy (leroy.boris@gmail.com), Maxime Lenormand (maxime.lenormand@inrae.fr) and Pierre Denelle (pierre.denelle@gmail.com)

Examples

comat <- matrix(sample(0:1000, size = 50, replace = TRUE,
prob = 1 / 1:1001), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

simil <- similarity(comat, metric = "all")
simil
#> Data.frame of similarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of similarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn  Sorensen Simpson      Bray  Brayturn
#> 2  Site1 Site2 0.6000000   0.6000000 0.7500000   0.750 0.2299101 0.3192133
#> 3  Site1 Site3 0.8000000   1.0000000 0.8888889   1.000 0.4079900 0.4933575
#> 4  Site1 Site4 0.7777778   0.7777778 0.8750000   0.875 0.5185380 0.6781293
#> 5  Site1 Site5 0.8000000   1.0000000 0.8888889   1.000 0.4079674 0.4675468
#> 8  Site2 Site3 0.8000000   1.0000000 0.8888889   1.000 0.3163197 0.3562784
#> 9  Site2 Site4 0.6000000   0.6000000 0.7500000   0.750 0.1865994 0.1959153
#> 10 Site2 Site5 0.8000000   1.0000000 0.8888889   1.000 0.3140127 0.3729198
#> 14 Site3 Site4 0.8000000   1.0000000 0.8888889   1.000 0.6418006 0.6863824
#> 15 Site3 Site5 1.0000000   1.0000000 1.0000000   1.000 0.3517559 0.3689614
#> 20 Site4 Site5 0.8000000   1.0000000 0.8888889   1.000 0.2212714 0.2489684
#>       Euclidean  a b c   A    B    C
#> 2  0.0009083526  6 2 2 422 1927  900
#> 3  0.0010233199  8 0 2 817 1532  839
#> 4  0.0011318106  7 1 1 986 1363  468
#> 5  0.0009602525  8 0 2 850 1499  968
#> 8  0.0010387087  8 0 2 471  851 1185
#> 9  0.0009793473  6 2 2 259 1063 1195
#> 10 0.0010346890  8 0 2 493  829 1325
#> 14 0.0019517119  8 2 0 998  658  456
#> 15 0.0009165394 10 0 0 611 1045 1207
#> 20 0.0008567734  8 0 2 362 1092 1456

dissimilarity <- similarity_to_dissimilarity(simil)
dissimilarity
#> Data.frame of dissimilarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of dissimilarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn  Sorensen Simpson      Bray  Brayturn
#> 2  Site1 Site2 0.4000000   0.4000000 0.2500000   0.250 0.7700899 0.6807867
#> 3  Site1 Site3 0.2000000   0.0000000 0.1111111   0.000 0.5920100 0.5066425
#> 4  Site1 Site4 0.2222222   0.2222222 0.1250000   0.125 0.4814620 0.3218707
#> 5  Site1 Site5 0.2000000   0.0000000 0.1111111   0.000 0.5920326 0.5324532
#> 8  Site2 Site3 0.2000000   0.0000000 0.1111111   0.000 0.6836803 0.6437216
#> 9  Site2 Site4 0.4000000   0.4000000 0.2500000   0.250 0.8134006 0.8040847
#> 10 Site2 Site5 0.2000000   0.0000000 0.1111111   0.000 0.6859873 0.6270802
#> 14 Site3 Site4 0.2000000   0.0000000 0.1111111   0.000 0.3581994 0.3136176
#> 15 Site3 Site5 0.0000000   0.0000000 0.0000000   0.000 0.6482441 0.6310386
#> 20 Site4 Site5 0.2000000   0.0000000 0.1111111   0.000 0.7787286 0.7510316
#>    Euclidean  a b c   A    B    C
#> 2  1099.8941  6 2 2 422 1927  900
#> 3   976.2116  8 0 2 817 1532  839
#> 4   882.5401  7 1 1 986 1363  468
#> 5  1040.3927  8 0 2 850 1499  968
#> 8   961.7339  8 0 2 471  851 1185
#> 9  1020.0882  6 2 2 259 1063 1195
#> 10  965.4740  8 0 2 493  829 1325
#> 14  511.3707  8 2 0 998  658  456
#> 15 1090.0605 10 0 0 611 1045 1207
#> 20 1166.1698  8 0 2 362 1092 1456

simil <- dissimilarity_to_similarity(dissimilarity)
simil
#> Data.frame of dissimilarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of dissimilarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn  Sorensen Simpson      Bray  Brayturn
#> 2  Site1 Site2 0.6000000   0.6000000 0.7500000   0.750 0.2299101 0.3192133
#> 3  Site1 Site3 0.8000000   1.0000000 0.8888889   1.000 0.4079900 0.4933575
#> 4  Site1 Site4 0.7777778   0.7777778 0.8750000   0.875 0.5185380 0.6781293
#> 5  Site1 Site5 0.8000000   1.0000000 0.8888889   1.000 0.4079674 0.4675468
#> 8  Site2 Site3 0.8000000   1.0000000 0.8888889   1.000 0.3163197 0.3562784
#> 9  Site2 Site4 0.6000000   0.6000000 0.7500000   0.750 0.1865994 0.1959153
#> 10 Site2 Site5 0.8000000   1.0000000 0.8888889   1.000 0.3140127 0.3729198
#> 14 Site3 Site4 0.8000000   1.0000000 0.8888889   1.000 0.6418006 0.6863824
#> 15 Site3 Site5 1.0000000   1.0000000 1.0000000   1.000 0.3517559 0.3689614
#> 20 Site4 Site5 0.8000000   1.0000000 0.8888889   1.000 0.2212714 0.2489684
#>       Euclidean  a b c   A    B    C
#> 2  0.0009083526  6 2 2 422 1927  900
#> 3  0.0010233199  8 0 2 817 1532  839
#> 4  0.0011318106  7 1 1 986 1363  468
#> 5  0.0009602525  8 0 2 850 1499  968
#> 8  0.0010387087  8 0 2 471  851 1185
#> 9  0.0009793473  6 2 2 259 1063 1195
#> 10 0.0010346890  8 0 2 493  829 1325
#> 14 0.0019517119  8 2 0 998  658  456
#> 15 0.0009165394 10 0 0 611 1045 1207
#> 20 0.0008567734  8 0 2 362 1092 1456