Convert dissimilarity indices to similarity indices

This function converts a data.frame of dissimilarity indices (beta diversity) between sites to similarity indices.

Usage

dissimilarity_to_similarity(dissimilaritydata)

Arguments

dissimilaritydata

the output object from similarity_to_dissimilarity() or a data.frame with the first columns called "Site1" and "Site2", and the other columns being the similarity metrics.

Value

A data.frame with additional class bioregion.pairwise.metric, providing one or several dissimilarity metric(s) between each pair of sites. The two first columns represent each pair of sites, and the other column represent dissimilarity metrics. Columns with names "a", "b", "c", "A", "B" and "C" are not altered.

Note

The behaviour of this function changes depending on column names. Columns "Site1" and "Site2" are copied identically. If there are columns called "a", "b", "c", "A", "B", "C", they will also be copied identically.

If a column is called "Euclidean", its similarity will be calculated based on the following formula:

\(Euclidean similarity = (1 - Euclidean distance) / Euclidean distance\)

Otherwise, all other columns will be transformed into similarity with the following formula:

\(similarity = 1 - dissimilarity\)

See also

Author

Boris Leroy (leroy.boris@gmail.com), Maxime Lenormand (maxime.lenormand@inrae.fr) and Pierre Denelle (pierre.denelle@gmail.com)

Examples

comat <- matrix(sample(0:1000, size = 50, replace = TRUE,
prob = 1 / 1:1001), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

simil <- similarity(comat, metric = "all")
simil
#> Data.frame of similarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of similarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn  Sorensen   Simpson       Bray  Brayturn
#> 2  Site1 Site2 0.7777778        1.00 0.8750000 1.0000000 0.24875069 0.6666667
#> 3  Site1 Site3 0.7777778        1.00 0.8750000 1.0000000 0.12439156 0.3422619
#> 4  Site1 Site4 0.6000000        0.75 0.7500000 0.8571429 0.42447419 0.6607143
#> 5  Site1 Site5 0.6000000        0.75 0.7500000 0.8571429 0.09747706 0.2529762
#> 8  Site2 Site3 0.8000000        0.80 0.8888889 0.8888889 0.22431162 0.2279863
#> 9  Site2 Site4 0.8000000        0.80 0.8888889 0.8888889 0.43954023 0.6732394
#> 10 Site2 Site5 0.8000000        0.80 0.8888889 0.8888889 0.32927254 0.3359375
#> 14 Site3 Site4 0.8000000        0.80 0.8888889 0.8888889 0.15384615 0.2408451
#> 15 Site3 Site5 0.8000000        0.80 0.8888889 0.8888889 0.30469017 0.3160511
#> 20 Site4 Site5 1.0000000        1.00 1.0000000 1.0000000 0.25684608 0.3830986
#>       Euclidean a b c   A    B    C
#> 2  0.0013602253 7 0 2 224  112 1241
#> 3  0.0013916984 7 0 2 115  221 1398
#> 4  0.0032829401 6 1 3 222  114  488
#> 5  0.0011372954 6 1 3  85  251 1323
#> 8  0.0010137892 8 1 1 334 1131 1179
#> 9  0.0014226957 8 1 1 478  987  232
#> 10 0.0009987232 8 1 1 473  992  935
#> 14 0.0012550920 8 1 1 171 1342  539
#> 15 0.0010688574 8 1 1 445 1068  963
#> 20 0.0011249064 9 0 0 272  438 1136

dissimilarity <- similarity_to_dissimilarity(simil)
dissimilarity
#> Data.frame of dissimilarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of dissimilarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn  Sorensen   Simpson      Bray  Brayturn
#> 2  Site1 Site2 0.2222222        0.00 0.1250000 0.0000000 0.7512493 0.3333333
#> 3  Site1 Site3 0.2222222        0.00 0.1250000 0.0000000 0.8756084 0.6577381
#> 4  Site1 Site4 0.4000000        0.25 0.2500000 0.1428571 0.5755258 0.3392857
#> 5  Site1 Site5 0.4000000        0.25 0.2500000 0.1428571 0.9025229 0.7470238
#> 8  Site2 Site3 0.2000000        0.20 0.1111111 0.1111111 0.7756884 0.7720137
#> 9  Site2 Site4 0.2000000        0.20 0.1111111 0.1111111 0.5604598 0.3267606
#> 10 Site2 Site5 0.2000000        0.20 0.1111111 0.1111111 0.6707275 0.6640625
#> 14 Site3 Site4 0.2000000        0.20 0.1111111 0.1111111 0.8461538 0.7591549
#> 15 Site3 Site5 0.2000000        0.20 0.1111111 0.1111111 0.6953098 0.6839489
#> 20 Site4 Site5 0.0000000        0.00 0.0000000 0.0000000 0.7431539 0.6169014
#>    Euclidean a b c   A    B    C
#> 2   734.1723 7 0 2 224  112 1241
#> 3   717.5465 7 0 2 115  221 1398
#> 4   303.6050 6 1 3 222  114  488
#> 5   878.2790 6 1 3  85  251 1323
#> 8   985.3984 8 1 1 334 1131 1179
#> 9   701.8910 8 1 1 478  987  232
#> 10 1000.2785 8 1 1 473  992  935
#> 14  795.7544 8 1 1 171 1342  539
#> 15  934.5785 8 1 1 445 1068  963
#> 20  887.9628 9 0 0 272  438 1136

simil <- dissimilarity_to_similarity(dissimilarity)
simil
#> Data.frame of dissimilarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of dissimilarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn  Sorensen   Simpson       Bray  Brayturn
#> 2  Site1 Site2 0.7777778        1.00 0.8750000 1.0000000 0.24875069 0.6666667
#> 3  Site1 Site3 0.7777778        1.00 0.8750000 1.0000000 0.12439156 0.3422619
#> 4  Site1 Site4 0.6000000        0.75 0.7500000 0.8571429 0.42447419 0.6607143
#> 5  Site1 Site5 0.6000000        0.75 0.7500000 0.8571429 0.09747706 0.2529762
#> 8  Site2 Site3 0.8000000        0.80 0.8888889 0.8888889 0.22431162 0.2279863
#> 9  Site2 Site4 0.8000000        0.80 0.8888889 0.8888889 0.43954023 0.6732394
#> 10 Site2 Site5 0.8000000        0.80 0.8888889 0.8888889 0.32927254 0.3359375
#> 14 Site3 Site4 0.8000000        0.80 0.8888889 0.8888889 0.15384615 0.2408451
#> 15 Site3 Site5 0.8000000        0.80 0.8888889 0.8888889 0.30469017 0.3160511
#> 20 Site4 Site5 1.0000000        1.00 1.0000000 1.0000000 0.25684608 0.3830986
#>       Euclidean a b c   A    B    C
#> 2  0.0013602253 7 0 2 224  112 1241
#> 3  0.0013916984 7 0 2 115  221 1398
#> 4  0.0032829401 6 1 3 222  114  488
#> 5  0.0011372954 6 1 3  85  251 1323
#> 8  0.0010137892 8 1 1 334 1131 1179
#> 9  0.0014226957 8 1 1 478  987  232
#> 10 0.0009987232 8 1 1 473  992  935
#> 14 0.0012550920 8 1 1 171 1342  539
#> 15 0.0010688574 8 1 1 445 1068  963
#> 20 0.0011249064 9 0 0 272  438 1136