Convert dissimilarity metrics to similarity metrics

This function converts a data.frame of dissimilarity metrics (beta diversity) between sites to similarity metrics.

Usage

dissimilarity_to_similarity(dissimilarity, include_formula = TRUE)

Arguments

dissimilarity

the output object from dissimilarity() or similarity_to_dissimilarity().

include_formula

a boolean indicating if the metrics based on your own formula should be converted (see Details). This argument is set to TRUE by default.

Value

A data.frame with additional class bioregion.pairwise.metric, providing similarity metric(s) between each pair of sites based on a dissimilarity object.

Note

The behavior of this function changes depending on column names. Columns Site1 and Site2 are copied identically. If there are columns called a, b, c, A, B, C they will also be copied identically. If there are columns based on your own formula (argument formula in dissimilarity()) or not in the original list of dissimilarity metrics (argument metrics in dissimilarity()) and if the argument include_formula is set to FALSE, they will also be copied identically. Otherwise there are going to be converted like they other columns (default behavior).

If a column is called Euclidean, the similarity will be calculated based on the following formula:

\(Euclidean similarity = 1 / (1 - Euclidean distance)\)

Otherwise, all other columns will be transformed into dissimilarity with the following formula:

\(similarity = 1 - dissimilarity\)

See also

Author

Maxime Lenormand (maxime.lenormand@inrae.fr), Boris Leroy (leroy.boris@gmail.com) and Pierre Denelle (pierre.denelle@gmail.com)

Examples

comat <- matrix(sample(0:1000, size = 50, replace = TRUE,
prob = 1 / 1:1001), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

dissimil <- dissimilarity(comat, metric = "all")
dissimil
#> Data.frame of dissimilarity between sites
#>  - Total number of sites:  5 
#>  - Total number of species:  10 
#>  - Number of rows:  10 
#>  - Number of dissimilarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn  Sorensen   Simpson      Bray  Brayturn
#> 2  Site1 Site2 0.0000000   0.0000000 0.0000000 0.0000000 0.8859527 0.8815029
#> 3  Site1 Site3 0.3000000   0.0000000 0.1764706 0.0000000 0.8207033 0.7149606
#> 4  Site1 Site4 0.3000000   0.0000000 0.1764706 0.0000000 0.8773946 0.8181818
#> 5  Site1 Site5 0.2000000   0.0000000 0.1111111 0.0000000 0.6009569 0.4093484
#> 8  Site2 Site3 0.3000000   0.0000000 0.1764706 0.0000000 0.6944053 0.4881890
#> 9  Site2 Site4 0.3000000   0.0000000 0.1764706 0.0000000 0.7040073 0.5383523
#> 10 Site2 Site5 0.2000000   0.0000000 0.1111111 0.0000000 0.8362147 0.7450425
#> 14 Site3 Site4 0.4444444   0.4444444 0.2857143 0.2857143 0.9492158 0.9464567
#> 15 Site3 Site5 0.5000000   0.4444444 0.3333333 0.2857143 0.9582401 0.9559055
#> 20 Site4 Site5 0.5000000   0.4444444 0.3333333 0.2857143 0.7801418 0.7798295
#>    Euclidean  a b c   A    B    C
#> 2  1146.8827 10 0 0 164 1220 1328
#> 3   816.8482  7 3 0 181 1203  454
#> 4   922.3459  7 3 0 128 1256  576
#> 5   848.7508  8 2 0 417  967  289
#> 8   663.4312  7 3 0 325 1167  310
#> 9   792.8897  7 3 0 325 1167  379
#> 10  844.5176  8 2 0 180 1312  526
#> 14  552.7685  5 2 2  34  601  670
#> 15  547.3198  5 2 3  28  607  678
#> 20  523.0239  5 2 3 155  549  551

similarity <- dissimilarity_to_similarity(dissimil)
similarity
#> Data.frame of similarity between sites
#>  - Total number of sites:  5 
#>  - Total number of species:  10 
#>  - Number of rows:  10 
#>  - Number of similarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn  Sorensen   Simpson       Bray   Brayturn
#> 2  Site1 Site2 1.0000000   1.0000000 1.0000000 1.0000000 0.11404729 0.11849711
#> 3  Site1 Site3 0.7000000   1.0000000 0.8235294 1.0000000 0.17929668 0.28503937
#> 4  Site1 Site4 0.7000000   1.0000000 0.8235294 1.0000000 0.12260536 0.18181818
#> 5  Site1 Site5 0.8000000   1.0000000 0.8888889 1.0000000 0.39904306 0.59065156
#> 8  Site2 Site3 0.7000000   1.0000000 0.8235294 1.0000000 0.30559473 0.51181102
#> 9  Site2 Site4 0.7000000   1.0000000 0.8235294 1.0000000 0.29599271 0.46164773
#> 10 Site2 Site5 0.8000000   1.0000000 0.8888889 1.0000000 0.16378526 0.25495751
#> 14 Site3 Site4 0.5555556   0.5555556 0.7142857 0.7142857 0.05078417 0.05354331
#> 15 Site3 Site5 0.5000000   0.5555556 0.6666667 0.7142857 0.04175988 0.04409449
#> 20 Site4 Site5 0.5000000   0.5555556 0.6666667 0.7142857 0.21985816 0.22017045
#>       Euclidean  a b c   A    B    C
#> 2  0.0008711691 10 0 0 164 1220 1328
#> 3  0.0012227208  7 3 0 181 1203  454
#> 4  0.0010830177  7 3 0 128 1256  576
#> 5  0.0011768155  8 2 0 417  967  289
#> 8  0.0015050466  7 3 0 325 1167  310
#> 9  0.0012596209  7 3 0 325 1167  379
#> 10 0.0011827075  8 2 0 180 1312  526
#> 14 0.0018058088  5 2 2  34  601  670
#> 15 0.0018237531  5 2 3  28  607  678
#> 20 0.0019083099  5 2 3 155  549  551