Convert dissimilarity metrics to similarity metrics

This function converts a data.frame of dissimilarity metrics (beta diversity) between sites to similarity metrics.

Usage

dissimilarity_to_similarity(dissimilarity, include_formula = TRUE)

Arguments

dissimilarity

the output object from dissimilarity() or similarity_to_dissimilarity().

include_formula

a boolean indicating if the metrics based on your own formula should be converted (see Details). This argument is set to TRUE by default.

Value

A data.frame with additional class bioregion.pairwise.metric, providing similarity metric(s) between each pair of sites based on a dissimilarity object.

Note

The behavior of this function changes depending on column names. Columns Site1 and Site2 are copied identically. If there are columns called a, b, c, A, B, C they will also be copied identically. If there are columns based on your own formula (argument formula in dissimilarity()) or not in the original list of dissimilarity metrics (argument metrics in dissimilarity()) and if the argument include_formula is set to FALSE, they will also be copied identically. Otherwise there are going to be converted like they other columns (default behavior).

If a column is called Euclidean, the similarity will be calculated based on the following formula:

\(Euclidean similarity = 1 / (1 - Euclidean distance)\)

Otherwise, all other columns will be transformed into dissimilarity with the following formula:

\(similarity = 1 - dissimilarity\)

See also

Author

Maxime Lenormand (maxime.lenormand@inrae.fr), Boris Leroy (leroy.boris@gmail.com) and Pierre Denelle (pierre.denelle@gmail.com)

Examples

comat <- matrix(sample(0:1000, size = 50, replace = TRUE,
prob = 1 / 1:1001), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

dissimil <- dissimilarity(comat, metric = "all")
dissimil
#> Data.frame of dissimilarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of dissimilarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn   Sorensen   Simpson      Bray   Brayturn
#> 2  Site1 Site2 0.3333333   0.2500000 0.20000000 0.1428571 0.8178694 0.81403509
#> 3  Site1 Site3 0.4000000   0.2500000 0.25000000 0.1428571 0.6477366 0.27946128
#> 4  Site1 Site4 0.3000000   0.0000000 0.17647059 0.0000000 0.9205864 0.78114478
#> 5  Site1 Site5 0.6000000   0.6000000 0.42857143 0.4285714 0.8508899 0.47811448
#> 8  Site2 Site3 0.3000000   0.2222222 0.17647059 0.1250000 0.8836243 0.75438596
#> 9  Site2 Site4 0.2000000   0.0000000 0.11111111 0.0000000 0.8363077 0.53333333
#> 10 Site2 Site5 0.5000000   0.4444444 0.33333333 0.2857143 0.9322690 0.75438596
#> 14 Site3 Site4 0.1000000   0.0000000 0.05263158 0.0000000 0.8148804 0.77233115
#> 15 Site3 Site5 0.4000000   0.2500000 0.25000000 0.1428571 0.3755556 0.08169935
#> 20 Site4 Site5 0.3000000   0.0000000 0.17647059 0.0000000 0.7520820 0.71119403
#>    Euclidean a b c   A   B    C
#> 2   240.6117 6 1 2  53 244  232
#> 3   457.4899 6 1 3 214  83  704
#> 4   810.3808 7 0 3  65 232 1275
#> 5   816.8568 4 3 3 155 142 1627
#> 8   510.1970 7 1 2  70 215  848
#> 9   709.0973 8 0 2 133 152 1207
#> 10  829.7668 5 3 2  70 215 1712
#> 14  882.5135 9 0 1 209 709 1131
#> 15  451.1541 6 3 1 843  75  939
#> 20 1010.8986 7 3 0 387 953 1395

similarity <- dissimilarity_to_similarity(dissimil)
similarity
#> Data.frame of similarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of similarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn  Sorensen   Simpson       Bray  Brayturn
#> 2  Site1 Site2 0.6666667   0.7500000 0.8000000 0.8571429 0.18213058 0.1859649
#> 3  Site1 Site3 0.6000000   0.7500000 0.7500000 0.8571429 0.35226337 0.7205387
#> 4  Site1 Site4 0.7000000   1.0000000 0.8235294 1.0000000 0.07941356 0.2188552
#> 5  Site1 Site5 0.4000000   0.4000000 0.5714286 0.5714286 0.14911015 0.5218855
#> 8  Site2 Site3 0.7000000   0.7777778 0.8235294 0.8750000 0.11637573 0.2456140
#> 9  Site2 Site4 0.8000000   1.0000000 0.8888889 1.0000000 0.16369231 0.4666667
#> 10 Site2 Site5 0.5000000   0.5555556 0.6666667 0.7142857 0.06773101 0.2456140
#> 14 Site3 Site4 0.9000000   1.0000000 0.9473684 1.0000000 0.18511957 0.2276688
#> 15 Site3 Site5 0.6000000   0.7500000 0.7500000 0.8571429 0.62444444 0.9183007
#> 20 Site4 Site5 0.7000000   1.0000000 0.8235294 1.0000000 0.24791800 0.2888060
#>       Euclidean a b c   A   B    C
#> 2  0.0041388721 6 1 2  53 244  232
#> 3  0.0021810732 6 1 3 214  83  704
#> 4  0.0012324670 7 0 3  65 232 1275
#> 5  0.0012227080 4 3 3 155 142 1627
#> 8  0.0019561929 7 1 2  70 215  848
#> 9  0.0014082577 8 0 2 133 152 1207
#> 10 0.0012037072 5 3 2  70 215 1712
#> 14 0.0011318447 9 0 1 209 709 1131
#> 15 0.0022116355 6 3 1 843  75  939
#> 20 0.0009882413 7 3 0 387 953 1395