Convert dissimilarity metrics to similarity metrics

This function converts a data.frame of dissimilarity metrics (beta diversity) between sites to similarity metrics.

Usage

dissimilarity_to_similarity(dissimilarity, include_formula = TRUE)

Arguments

dissimilarity

the output object from dissimilarity() or similarity_to_dissimilarity().

include_formula

a boolean indicating if the metrics based on your own formula should be converted (see Details). This argument is set to TRUE by default.

Value

A data.frame with additional class bioregion.pairwise.metric, providing similarity metric(s) between each pair of sites based on a dissimilarity object.

Note

The behavior of this function changes depending on column names. Columns Site1 and Site2 are copied identically. If there are columns called a, b, c, A, B, C they will also be copied identically. If there are columns based on your own formula (argument formula in dissimilarity()) or not in the original list of dissimilarity metrics (argument metrics in dissimilarity()) and if the argument include_formula is set to FALSE, they will also be copied identically. Otherwise there are going to be converted like they other columns (default behavior).

If a column is called Euclidean, the similarity will be calculated based on the following formula:

\(Euclidean similarity = 1 / (1 - Euclidean distance)\)

Otherwise, all other columns will be transformed into dissimilarity with the following formula:

\(similarity = 1 - dissimilarity\)

See also

Author

Maxime Lenormand (maxime.lenormand@inrae.fr), Boris Leroy (leroy.boris@gmail.com) and Pierre Denelle (pierre.denelle@gmail.com)

Examples

comat <- matrix(sample(0:1000, size = 50, replace = TRUE,
prob = 1 / 1:1001), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

dissimil <- dissimilarity(comat, metric = "all")
dissimil
#> Data.frame of dissimilarity between sites
#>  - Total number of sites:  5 
#>  - Total number of species:  10 
#>  - Number of rows:  10 
#>  - Number of dissimilarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn   Sorensen   Simpson      Bray  Brayturn
#> 2  Site1 Site2 0.1111111   0.0000000 0.05882353 0.0000000 0.9330484 0.8625731
#> 3  Site1 Site3 0.3000000   0.2222222 0.17647059 0.1250000 0.8412492 0.6432749
#> 4  Site1 Site4 0.4000000   0.4000000 0.25000000 0.2500000 0.9420784 0.8508772
#> 5  Site1 Site5 0.3000000   0.2222222 0.17647059 0.1250000 0.9743670 0.8801170
#> 8  Site2 Site3 0.2000000   0.2000000 0.11111111 0.1111111 0.6083296 0.5838041
#> 9  Site2 Site4 0.3000000   0.2222222 0.17647059 0.1250000 0.8863362 0.8672316
#> 10 Site2 Site5 0.2000000   0.2000000 0.11111111 0.1111111 0.8821128 0.7824859
#> 14 Site3 Site4 0.3000000   0.2222222 0.17647059 0.1250000 0.6878347 0.6585774
#> 15 Site3 Site5 0.2000000   0.2000000 0.11111111 0.1111111 0.7196446 0.5246862
#> 20 Site4 Site5 0.3000000   0.2222222 0.17647059 0.1250000 0.5809167 0.3685694
#>    Euclidean a b c   A   B    C
#> 2   590.9721 8 0 1  47 295 1015
#> 3   529.1266 7 1 2 122 220 1073
#> 4  1000.0035 6 2 2  51 291 1368
#> 5  1401.5395 7 1 2  41 301 2816
#> 8   614.4388 8 1 1 442 620  753
#> 9  1097.9495 7 2 1 141 921 1278
#> 10 1435.8033 8 1 1 231 831 2626
#> 14  858.9971 7 2 1 408 787 1011
#> 15 1337.3631 8 1 1 568 627 2289
#> 20 1154.5579 7 1 2 896 523 1961

similarity <- dissimilarity_to_similarity(dissimil)
similarity
#> Data.frame of similarity between sites
#>  - Total number of sites:  5 
#>  - Total number of species:  10 
#>  - Number of rows:  10 
#>  - Number of similarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn  Sorensen   Simpson       Bray  Brayturn
#> 2  Site1 Site2 0.8888889   1.0000000 0.9411765 1.0000000 0.06695157 0.1374269
#> 3  Site1 Site3 0.7000000   0.7777778 0.8235294 0.8750000 0.15875081 0.3567251
#> 4  Site1 Site4 0.6000000   0.6000000 0.7500000 0.7500000 0.05792164 0.1491228
#> 5  Site1 Site5 0.7000000   0.7777778 0.8235294 0.8750000 0.02563301 0.1198830
#> 8  Site2 Site3 0.8000000   0.8000000 0.8888889 0.8888889 0.39167036 0.4161959
#> 9  Site2 Site4 0.7000000   0.7777778 0.8235294 0.8750000 0.11366385 0.1327684
#> 10 Site2 Site5 0.8000000   0.8000000 0.8888889 0.8888889 0.11788722 0.2175141
#> 14 Site3 Site4 0.7000000   0.7777778 0.8235294 0.8750000 0.31216526 0.3414226
#> 15 Site3 Site5 0.8000000   0.8000000 0.8888889 0.8888889 0.28035538 0.4753138
#> 20 Site4 Site5 0.7000000   0.7777778 0.8235294 0.8750000 0.41908326 0.6314306
#>       Euclidean a b c   A   B    C
#> 2  0.0016892689 8 0 1  47 295 1015
#> 3  0.0018863417 7 1 2 122 220 1073
#> 4  0.0009989975 6 2 2  51 291 1368
#> 5  0.0007129924 7 1 2  41 301 2816
#> 8  0.0016248570 8 1 1 442 620  753
#> 9  0.0009099600 7 2 1 141 921 1278
#> 10 0.0006959895 8 1 1 231 831 2626
#> 14 0.0011627946 7 2 1 408 787 1011
#> 15 0.0007471814 8 1 1 568 627 2289
#> 20 0.0008653828 7 1 2 896 523 1961