Convert dissimilarity indices to similarity indices

This function converts a data.frame of dissimilarity indices (beta diversity) between sites to similarity indices.

Usage

dissimilarity_to_similarity(dissimilaritydata)

Arguments

dissimilaritydata

the output object from similarity_to_dissimilarity() or a data.frame with the first columns called "Site1" and "Site2", and the other columns being the similarity metrics.

Value

A data.frame with additional class bioregion.pairwise.metric, providing one or several dissimilarity metric(s) between each pair of sites. The two first columns represent each pair of sites, and the other column represent dissimilarity metrics. Columns with names "a", "b", "c", "A", "B" and "C" are not altered.

Note

The behaviour of this function changes depending on column names. Columns "Site1" and "Site2" are copied identically. If there are columns called "a", "b", "c", "A", "B", "C", they will also be copied identically.

If a column is called "Euclidean", its similarity will be calculated based on the following formula:

\(Euclidean similarity = (1 - Euclidean distance) / Euclidean distance\)

Otherwise, all other columns will be transformed into similarity with the following formula:

\(similarity = 1 - dissimilarity\)

See also

Author

Boris Leroy (leroy.boris@gmail.com), Maxime Lenormand (maxime.lenormand@inrae.fr) and Pierre Denelle (pierre.denelle@gmail.com)

Examples

comat <- matrix(sample(0:1000, size = 50, replace = TRUE,
prob = 1 / 1:1001), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

simil <- similarity(comat, metric = "all")
simil
#> Data.frame of similarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of similarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn  Sorensen   Simpson       Bray  Brayturn
#> 2  Site1 Site2 0.6666667   0.7500000 0.8000000 0.8571429 0.17526618 0.6350148
#> 3  Site1 Site3 0.7000000   1.0000000 0.8235294 1.0000000 0.31029987 0.4942887
#> 4  Site1 Site4 0.6000000   0.7500000 0.7500000 0.8571429 0.35581669 0.3596200
#> 5  Site1 Site5 0.6000000   0.7500000 0.7500000 0.8571429 0.10526316 0.1179941
#> 8  Site2 Site3 0.8000000   1.0000000 0.8888889 1.0000000 0.16923077 0.3264095
#> 9  Site2 Site4 0.7000000   0.7777778 0.8235294 0.8750000 0.07559308 0.2789318
#> 10 Site2 Site5 0.7000000   0.7777778 0.8235294 0.8750000 0.07185039 0.2166172
#> 14 Site3 Site4 0.9000000   1.0000000 0.9473684 1.0000000 0.28589785 0.4620976
#> 15 Site3 Site5 0.9000000   1.0000000 0.9473684 1.0000000 0.17607223 0.2429907
#> 20 Site4 Site5 0.8000000   0.8000000 0.8888889 0.8888889 0.14200260 0.1610619
#>       Euclidean a b c   A    B    C
#> 2  0.0008783969 6 1 2 214 1891  123
#> 3  0.0009016186 7 0 3 476 1629  487
#> 4  0.0007877790 6 1 3 757 1348 1393
#> 5  0.0006820960 6 1 3 200 1905 1495
#> 8  0.0017947411 8 0 2 110  227  853
#> 9  0.0008481749 7 1 2  94  243 2056
#> 10 0.0010027223 7 1 2  73  264 1622
#> 14 0.0009215176 9 1 0 445  518 1705
#> 15 0.0009120368 9 1 0 234  729 1461
#> 20 0.0006922225 8 1 1 273 1877 1422

dissimilarity <- similarity_to_dissimilarity(simil)
dissimilarity
#> Data.frame of dissimilarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of dissimilarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn   Sorensen   Simpson      Bray  Brayturn
#> 2  Site1 Site2 0.3333333   0.2500000 0.20000000 0.1428571 0.8247338 0.3649852
#> 3  Site1 Site3 0.3000000   0.0000000 0.17647059 0.0000000 0.6897001 0.5057113
#> 4  Site1 Site4 0.4000000   0.2500000 0.25000000 0.1428571 0.6441833 0.6403800
#> 5  Site1 Site5 0.4000000   0.2500000 0.25000000 0.1428571 0.8947368 0.8820059
#> 8  Site2 Site3 0.2000000   0.0000000 0.11111111 0.0000000 0.8307692 0.6735905
#> 9  Site2 Site4 0.3000000   0.2222222 0.17647059 0.1250000 0.9244069 0.7210682
#> 10 Site2 Site5 0.3000000   0.2222222 0.17647059 0.1250000 0.9281496 0.7833828
#> 14 Site3 Site4 0.1000000   0.0000000 0.05263158 0.0000000 0.7141022 0.5379024
#> 15 Site3 Site5 0.1000000   0.0000000 0.05263158 0.0000000 0.8239278 0.7570093
#> 20 Site4 Site5 0.2000000   0.2000000 0.11111111 0.1111111 0.8579974 0.8389381
#>    Euclidean a b c   A    B    C
#> 2  1137.4375 6 1 2 214 1891  123
#> 3  1108.1164 7 0 3 476 1629  487
#> 4  1268.3915 6 1 3 757 1348 1393
#> 5  1465.0693 6 1 3 200 1905 1495
#> 8   556.1834 8 0 2 110  227  853
#> 9  1178.0021 7 1 2  94  243 2056
#> 10  996.2851 7 1 2  73  264 1622
#> 14 1084.1665 9 1 0 445  518 1705
#> 15 1095.4469 9 1 0 234  729 1461
#> 20 1443.6222 8 1 1 273 1877 1422

simil <- dissimilarity_to_similarity(dissimilarity)
simil
#> Data.frame of dissimilarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of dissimilarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn  Sorensen   Simpson       Bray  Brayturn
#> 2  Site1 Site2 0.6666667   0.7500000 0.8000000 0.8571429 0.17526618 0.6350148
#> 3  Site1 Site3 0.7000000   1.0000000 0.8235294 1.0000000 0.31029987 0.4942887
#> 4  Site1 Site4 0.6000000   0.7500000 0.7500000 0.8571429 0.35581669 0.3596200
#> 5  Site1 Site5 0.6000000   0.7500000 0.7500000 0.8571429 0.10526316 0.1179941
#> 8  Site2 Site3 0.8000000   1.0000000 0.8888889 1.0000000 0.16923077 0.3264095
#> 9  Site2 Site4 0.7000000   0.7777778 0.8235294 0.8750000 0.07559308 0.2789318
#> 10 Site2 Site5 0.7000000   0.7777778 0.8235294 0.8750000 0.07185039 0.2166172
#> 14 Site3 Site4 0.9000000   1.0000000 0.9473684 1.0000000 0.28589785 0.4620976
#> 15 Site3 Site5 0.9000000   1.0000000 0.9473684 1.0000000 0.17607223 0.2429907
#> 20 Site4 Site5 0.8000000   0.8000000 0.8888889 0.8888889 0.14200260 0.1610619
#>       Euclidean a b c   A    B    C
#> 2  0.0008783969 6 1 2 214 1891  123
#> 3  0.0009016186 7 0 3 476 1629  487
#> 4  0.0007877790 6 1 3 757 1348 1393
#> 5  0.0006820960 6 1 3 200 1905 1495
#> 8  0.0017947411 8 0 2 110  227  853
#> 9  0.0008481749 7 1 2  94  243 2056
#> 10 0.0010027223 7 1 2  73  264 1622
#> 14 0.0009215176 9 1 0 445  518 1705
#> 15 0.0009120368 9 1 0 234  729 1461
#> 20 0.0006922225 8 1 1 273 1877 1422