Convert dissimilarity metrics to similarity metrics

This function converts a data.frame of dissimilarity metrics (beta diversity) between sites to similarity metrics.

Usage

dissimilarity_to_similarity(dissimilarity, include_formula = TRUE)

Arguments

dissimilarity

the output object from dissimilarity() or similarity_to_dissimilarity().

include_formula

a boolean indicating if the metrics based on your own formula(s) should be converted (see Details). This argument is set to TRUE by default.

Value

A data.frame with additional class bioregion.pairwise.metric, providing similarity metric(s) between each pair of sites based on a dissimilarity object.

Note

The behavior of this function changes depending on column names. Columns Site1 and Site2 are copied identically. If there are columns called a, b, c, A, B, C they will also be copied identically. If there are columns based on your own formula (argument formula in dissimilarity()) or not in the original list of dissimilarity metrics (argument metrics in dissimilarity()) and if the argument include_formula is set to FALSE, they will also be copied identically. Otherwise there are going to be converted like they other columns (default behavior).

If a column is called Euclidean, the similarity will be calculated based on the following formula:

Euclidean similarity = 1 / (1 - Euclidean distance)Euclidean similarity = 1 / (1 - Euclidean distance)

Otherwise, all other columns will be transformed into dissimilarity with the following formula:

similarity = 1 - dissimilaritysimilarity = 1 - dissimilarity

See also

Author

Maxime Lenormand (maxime.lenormand@inrae.fr), Boris Leroy (leroy.boris@gmail.com) and Pierre Denelle (pierre.denelle@gmail.com)

Examples

comat <- matrix(sample(0:1000, size = 50, replace = TRUE,
prob = 1 / 1:1001), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

dissimil <- dissimilarity(comat, metric = "all")
dissimil
#> Data.frame of dissimilarity between sites
#>  - Total number of sites:  5 
#>  - Total number of species:  10 
#>  - Number of rows:  10 
#>  - Number of dissimilarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn  Sorensen   Simpson      Bray  Brayturn
#> 2  Site1 Site2 0.3000000   0.2222222 0.1764706 0.1250000 0.9052069 0.8955114
#> 3  Site1 Site3 0.4000000   0.2500000 0.2500000 0.1428571 0.7005841 0.6986386
#> 4  Site1 Site4 0.2000000   0.2000000 0.1111111 0.1111111 0.6574949 0.4774436
#> 5  Site1 Site5 0.3000000   0.2222222 0.1764706 0.1250000 0.6449585 0.5687233
#> 8  Site2 Site3 0.3333333   0.2500000 0.2000000 0.1428571 0.6739496 0.6431199
#> 9  Site2 Site4 0.3000000   0.2222222 0.1764706 0.1250000 0.8562819 0.8057644
#> 10 Site2 Site5 0.2222222   0.2222222 0.1250000 0.1250000 0.8399567 0.7821928
#> 14 Site3 Site4 0.4000000   0.2500000 0.2500000 0.1428571 0.5923778 0.3834586
#> 15 Site3 Site5 0.3333333   0.2500000 0.2000000 0.1428571 0.4752275 0.3576733
#> 20 Site4 Site5 0.3000000   0.2222222 0.1764706 0.1250000 0.5455704 0.1065163
#>    Euclidean a b c    A    B    C
#> 2  1243.8738 7 2 1  142 1495 1217
#> 3   918.6800 6 3 1  487 1150 1129
#> 4   713.8312 8 1 1  417 1220  381
#> 5  1120.8671 7 2 1  706  931 1634
#> 8   945.0878 6 2 1  485  874 1131
#> 9   989.4751 7 1 2  155 1204  643
#> 10 1344.1395 7 1 1  296 1063 2044
#> 14  585.3546 6 1 3  492 1124  306
#> 15  776.1701 6 1 2 1038  578 1302
#> 20  882.2585 7 2 1  713   85 1627

similarity <- dissimilarity_to_similarity(dissimil)
similarity
#> Data.frame of similarity between sites
#>  - Total number of sites:  5 
#>  - Total number of species:  10 
#>  - Number of rows:  10 
#>  - Number of similarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn  Sorensen   Simpson       Bray  Brayturn
#> 2  Site1 Site2 0.7000000   0.7777778 0.8235294 0.8750000 0.09479306 0.1044886
#> 3  Site1 Site3 0.6000000   0.7500000 0.7500000 0.8571429 0.29941592 0.3013614
#> 4  Site1 Site4 0.8000000   0.8000000 0.8888889 0.8888889 0.34250513 0.5225564
#> 5  Site1 Site5 0.7000000   0.7777778 0.8235294 0.8750000 0.35504149 0.4312767
#> 8  Site2 Site3 0.6666667   0.7500000 0.8000000 0.8571429 0.32605042 0.3568801
#> 9  Site2 Site4 0.7000000   0.7777778 0.8235294 0.8750000 0.14371813 0.1942356
#> 10 Site2 Site5 0.7777778   0.7777778 0.8750000 0.8750000 0.16004325 0.2178072
#> 14 Site3 Site4 0.6000000   0.7500000 0.7500000 0.8571429 0.40762220 0.6165414
#> 15 Site3 Site5 0.6666667   0.7500000 0.8000000 0.8571429 0.52477250 0.6423267
#> 20 Site4 Site5 0.7000000   0.7777778 0.8235294 0.8750000 0.45442957 0.8934837
#>       Euclidean a b c    A    B    C
#> 2  0.0008032943 7 2 1  142 1495 1217
#> 3  0.0010873347 6 3 1  487 1150 1129
#> 4  0.0013989316 8 1 1  417 1220  381
#> 5  0.0008913712 7 2 1  706  931 1634
#> 8  0.0010569843 6 2 1  485  874 1131
#> 9  0.0010096165 7 1 2  155 1204  643
#> 10 0.0007434173 7 1 1  296 1063 2044
#> 14 0.0017054527 6 1 3  492 1124  306
#> 15 0.0012867196 6 1 2 1038  578 1302
#> 20 0.0011321714 7 2 1  713   85 1627