Convert dissimilarity indices to similarity indices

This function converts a data.frame of dissimilarity indices (beta diversity) between sites to similarity indices.

Usage

dissimilarity_to_similarity(dissimilaritydata)

Arguments

dissimilaritydata

the output object from similarity_to_dissimilarity() or a data.frame with the first columns called "Site1" and "Site2", and the other columns being the similarity metrics.

Value

A data.frame with additional class bioregion.pairwise.metric, providing one or several dissimilarity metric(s) between each pair of sites. The two first columns represent each pair of sites, and the other column represent dissimilarity metrics. Columns with names "a", "b", "c", "A", "B" and "C" are not altered.

Note

The behaviour of this function changes depending on column names. Columns "Site1" and "Site2" are copied identically. If there are columns called "a", "b", "c", "A", "B", "C", they will also be copied identically.

If a column is called "Euclidean", its similarity will be calculated based on the following formula:

\(Euclidean similarity = (1 - Euclidean distance) / Euclidean distance\)

Otherwise, all other columns will be transformed into similarity with the following formula:

\(similarity = 1 - dissimilarity\)

See also

Author

Boris Leroy (leroy.boris@gmail.com), Maxime Lenormand (maxime.lenormand@inrae.fr) and Pierre Denelle (pierre.denelle@gmail.com)

Examples

# \donttest{
comat <- matrix(sample(0:1000, size = 50, replace = TRUE,
prob = 1 / 1:1001), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

simil <- similarity(comat, metric = "all")
simil
#> Data.frame of similarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of similarity metrics:  7 
#> 
#> 
#>    Site1 Site2 Jaccard Jaccardturn  Sorensen   Simpson       Bray  Brayturn
#> 2  Site1 Site2     0.9        1.00 0.9473684 1.0000000 0.35018051 0.6092965
#> 3  Site1 Site3     0.8        0.80 0.8888889 0.8888889 0.10000000 0.1469849
#> 4  Site1 Site4     0.9        1.00 0.9473684 1.0000000 0.06445837 0.1121495
#> 5  Site1 Site5     0.6        0.75 0.7500000 0.8571429 0.26452064 0.2985782
#> 8  Site2 Site3     0.9        1.00 0.9473684 1.0000000 0.14667425 0.1670984
#> 9  Site2 Site4     1.0        1.00 1.0000000 1.0000000 0.20740741 0.7414330
#> 10 Site2 Site5     0.7        1.00 0.8235294 1.0000000 0.40122746 0.8262243
#> 14 Site3 Site4     0.9        1.00 0.9473684 1.0000000 0.11903485 0.3457944
#> 15 Site3 Site5     0.6        0.75 0.7500000 0.8571429 0.15066605 0.2590837
#> 20 Site4 Site5     0.7        1.00 0.8235294 1.0000000 0.50524109 0.7507788
#>       Euclidean  a b c   A    B    C
#> 2  0.0013972445  9 0 1 485  311 1489
#> 3  0.0010353772  8 1 1 117  679 1427
#> 4  0.0018802318  9 0 1  36  760  285
#> 5  0.0019558855  6 3 1 189  607  444
#> 8  0.0008395322  9 1 0 258 1716 1286
#> 9  0.0011393014 10 0 0 238 1736   83
#> 10 0.0013419708  7 3 0 523 1451  110
#> 14 0.0011686012  9 0 1 111 1433  210
#> 15 0.0011275055  6 3 1 164 1380  469
#> 20 0.0046274878  7 3 0 241   80  392

dissimilarity <- similarity_to_dissimilarity(simil)
dissimilarity
#> Data.frame of dissimilarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of dissimilarity metrics:  7 
#> 
#> 
#>    Site1 Site2 Jaccard Jaccardturn   Sorensen   Simpson      Bray  Brayturn
#> 2  Site1 Site2     0.1        0.00 0.05263158 0.0000000 0.6498195 0.3907035
#> 3  Site1 Site3     0.2        0.20 0.11111111 0.1111111 0.9000000 0.8530151
#> 4  Site1 Site4     0.1        0.00 0.05263158 0.0000000 0.9355416 0.8878505
#> 5  Site1 Site5     0.4        0.25 0.25000000 0.1428571 0.7354794 0.7014218
#> 8  Site2 Site3     0.1        0.00 0.05263158 0.0000000 0.8533258 0.8329016
#> 9  Site2 Site4     0.0        0.00 0.00000000 0.0000000 0.7925926 0.2585670
#> 10 Site2 Site5     0.3        0.00 0.17647059 0.0000000 0.5987725 0.1737757
#> 14 Site3 Site4     0.1        0.00 0.05263158 0.0000000 0.8809651 0.6542056
#> 15 Site3 Site5     0.4        0.25 0.25000000 0.1428571 0.8493339 0.7409163
#> 20 Site4 Site5     0.3        0.00 0.17647059 0.0000000 0.4947589 0.2492212
#>    Euclidean  a b c   A    B    C
#> 2   714.6943  9 0 1 485  311 1489
#> 3   964.8316  8 1 1 117  679 1427
#> 4   530.8493  9 0 1  36  760  285
#> 5   510.2774  6 3 1 189  607  444
#> 8  1190.1395  9 1 0 258 1716 1286
#> 9   876.7309 10 0 0 238 1736   83
#> 10  744.1727  7 3 0 523 1451  110
#> 14  854.7239  9 0 1 111 1433  210
#> 15  885.9137  6 3 1 164 1380  469
#> 20  215.1000  7 3 0 241   80  392

simil <- dissimilarity_to_similarity(dissimilarity)
simil
#> Data.frame of dissimilarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of dissimilarity metrics:  7 
#> 
#> 
#>    Site1 Site2 Jaccard Jaccardturn  Sorensen   Simpson       Bray  Brayturn
#> 2  Site1 Site2     0.9        1.00 0.9473684 1.0000000 0.35018051 0.6092965
#> 3  Site1 Site3     0.8        0.80 0.8888889 0.8888889 0.10000000 0.1469849
#> 4  Site1 Site4     0.9        1.00 0.9473684 1.0000000 0.06445837 0.1121495
#> 5  Site1 Site5     0.6        0.75 0.7500000 0.8571429 0.26452064 0.2985782
#> 8  Site2 Site3     0.9        1.00 0.9473684 1.0000000 0.14667425 0.1670984
#> 9  Site2 Site4     1.0        1.00 1.0000000 1.0000000 0.20740741 0.7414330
#> 10 Site2 Site5     0.7        1.00 0.8235294 1.0000000 0.40122746 0.8262243
#> 14 Site3 Site4     0.9        1.00 0.9473684 1.0000000 0.11903485 0.3457944
#> 15 Site3 Site5     0.6        0.75 0.7500000 0.8571429 0.15066605 0.2590837
#> 20 Site4 Site5     0.7        1.00 0.8235294 1.0000000 0.50524109 0.7507788
#>       Euclidean  a b c   A    B    C
#> 2  0.0013972445  9 0 1 485  311 1489
#> 3  0.0010353772  8 1 1 117  679 1427
#> 4  0.0018802318  9 0 1  36  760  285
#> 5  0.0019558855  6 3 1 189  607  444
#> 8  0.0008395322  9 1 0 258 1716 1286
#> 9  0.0011393014 10 0 0 238 1736   83
#> 10 0.0013419708  7 3 0 523 1451  110
#> 14 0.0011686012  9 0 1 111 1433  210
#> 15 0.0011275055  6 3 1 164 1380  469
#> 20 0.0046274878  7 3 0 241   80  392
# }