Convert dissimilarity indices to similarity indices

This function converts a data.frame of dissimilarity indices (beta diversity) between sites to similarity indices.

Usage

dissimilarity_to_similarity(dissimilaritydata)

Arguments

dissimilaritydata

the output object from similarity_to_dissimilarity() or a data.frame with the first columns called "Site1" and "Site2", and the other columns being the similarity metrics.

Value

A data.frame with additional class bioregion.pairwise.metric, providing one or several dissimilarity metric(s) between each pair of sites. The two first columns represent each pair of sites, and the other column represent dissimilarity metrics. Columns with names "a", "b", "c", "A", "B" and "C" are not altered.

Note

The behaviour of this function changes depending on column names. Columns "Site1" and "Site2" are copied identically. If there are columns called "a", "b", "c", "A", "B", "C", they will also be copied identically.

If a column is called "Euclidean", its similarity will be calculated based on the following formula:

\(Euclidean similarity = (1 - Euclidean distance) / Euclidean distance\)

Otherwise, all other columns will be transformed into similarity with the following formula:

\(similarity = 1 - dissimilarity\)

See also

Author

Boris Leroy (leroy.boris@gmail.com), Maxime Lenormand (maxime.lenormand@inrae.fr) and Pierre Denelle (pierre.denelle@gmail.com)

Examples

comat <- matrix(sample(0:1000, size = 50, replace = TRUE,
prob = 1 / 1:1001), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

simil <- similarity(comat, metric = "all")
simil
#> Data.frame of similarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of similarity metrics:  7 
#> 
#> 
#>    Site1 Site2 Jaccard Jaccardturn  Sorensen   Simpson       Bray   Brayturn
#> 2  Site1 Site2     0.7        1.00 0.8235294 1.0000000 0.42363776 0.49909584
#> 3  Site1 Site3     1.0        1.00 1.0000000 1.0000000 0.12627581 0.17733333
#> 4  Site1 Site4     1.0        1.00 1.0000000 1.0000000 0.21930974 0.31812421
#> 5  Site1 Site5     0.9        1.00 0.9473684 1.0000000 0.32065964 0.51244510
#> 8  Site2 Site3     0.7        1.00 0.8235294 1.0000000 0.05655931 0.09764919
#> 9  Site2 Site4     0.7        1.00 0.8235294 1.0000000 0.17097625 0.20532319
#> 10 Site2 Site5     0.6        0.75 0.7500000 0.8571429 0.23476803 0.30746706
#> 14 Site3 Site4     1.0        1.00 1.0000000 1.0000000 0.32609937 0.72370089
#> 15 Site3 Site5     0.9        1.00 0.9473684 1.0000000 0.26442874 0.65739385
#> 20 Site4 Site5     0.9        1.00 0.9473684 1.0000000 0.34646739 0.37335286
#>       Euclidean  a b c   A    B    C
#> 2  0.0012408717  7 3 0 552  948  554
#> 3  0.0007281858 10 0 0 266 1234 2447
#> 4  0.0011601209 10 0 0 251 1249  538
#> 5  0.0013630440  9 1 0 350 1150  333
#> 8  0.0006993592  7 0 3 108  998 2605
#> 9  0.0012064545  7 0 3 162  944  627
#> 10 0.0012475063  6 1 3 210  896  473
#> 14 0.0009048921 10 0 0 571 2142  218
#> 15 0.0009538747  9 1 0 449 2264  234
#> 20 0.0017790237  9 1 0 255  534  428

dissimilarity <- similarity_to_dissimilarity(simil)
dissimilarity
#> Data.frame of dissimilarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of dissimilarity metrics:  7 
#> 
#> 
#>    Site1 Site2 Jaccard Jaccardturn   Sorensen   Simpson      Bray  Brayturn
#> 2  Site1 Site2     0.3        0.00 0.17647059 0.0000000 0.5763622 0.5009042
#> 3  Site1 Site3     0.0        0.00 0.00000000 0.0000000 0.8737242 0.8226667
#> 4  Site1 Site4     0.0        0.00 0.00000000 0.0000000 0.7806903 0.6818758
#> 5  Site1 Site5     0.1        0.00 0.05263158 0.0000000 0.6793404 0.4875549
#> 8  Site2 Site3     0.3        0.00 0.17647059 0.0000000 0.9434407 0.9023508
#> 9  Site2 Site4     0.3        0.00 0.17647059 0.0000000 0.8290237 0.7946768
#> 10 Site2 Site5     0.4        0.25 0.25000000 0.1428571 0.7652320 0.6925329
#> 14 Site3 Site4     0.0        0.00 0.00000000 0.0000000 0.6739006 0.2762991
#> 15 Site3 Site5     0.1        0.00 0.05263158 0.0000000 0.7355713 0.3426061
#> 20 Site4 Site5     0.1        0.00 0.05263158 0.0000000 0.6535326 0.6266471
#>    Euclidean  a b c   A    B    C
#> 2   804.8851  7 3 0 552  948  554
#> 3  1372.2758 10 0 0 266 1234 2447
#> 4   860.9791 10 0 0 251 1249  538
#> 5   732.6520  9 1 0 350 1150  333
#> 8  1428.8803  7 0 3 108  998 2605
#> 9   827.8750  7 0 3 162  944  627
#> 10  800.5992  6 1 3 210  896  473
#> 14 1104.1042 10 0 0 571 2142  218
#> 15 1047.3557  9 1 0 449 2264  234
#> 20  561.1061  9 1 0 255  534  428

simil <- dissimilarity_to_similarity(dissimilarity)
simil
#> Data.frame of dissimilarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of dissimilarity metrics:  7 
#> 
#> 
#>    Site1 Site2 Jaccard Jaccardturn  Sorensen   Simpson       Bray   Brayturn
#> 2  Site1 Site2     0.7        1.00 0.8235294 1.0000000 0.42363776 0.49909584
#> 3  Site1 Site3     1.0        1.00 1.0000000 1.0000000 0.12627581 0.17733333
#> 4  Site1 Site4     1.0        1.00 1.0000000 1.0000000 0.21930974 0.31812421
#> 5  Site1 Site5     0.9        1.00 0.9473684 1.0000000 0.32065964 0.51244510
#> 8  Site2 Site3     0.7        1.00 0.8235294 1.0000000 0.05655931 0.09764919
#> 9  Site2 Site4     0.7        1.00 0.8235294 1.0000000 0.17097625 0.20532319
#> 10 Site2 Site5     0.6        0.75 0.7500000 0.8571429 0.23476803 0.30746706
#> 14 Site3 Site4     1.0        1.00 1.0000000 1.0000000 0.32609937 0.72370089
#> 15 Site3 Site5     0.9        1.00 0.9473684 1.0000000 0.26442874 0.65739385
#> 20 Site4 Site5     0.9        1.00 0.9473684 1.0000000 0.34646739 0.37335286
#>       Euclidean  a b c   A    B    C
#> 2  0.0012408717  7 3 0 552  948  554
#> 3  0.0007281858 10 0 0 266 1234 2447
#> 4  0.0011601209 10 0 0 251 1249  538
#> 5  0.0013630440  9 1 0 350 1150  333
#> 8  0.0006993592  7 0 3 108  998 2605
#> 9  0.0012064545  7 0 3 162  944  627
#> 10 0.0012475063  6 1 3 210  896  473
#> 14 0.0009048921 10 0 0 571 2142  218
#> 15 0.0009538747  9 1 0 449 2264  234
#> 20 0.0017790237  9 1 0 255  534  428