Convert dissimilarity indices to similarity indices

This function converts a data.frame of dissimilarity indices (beta diversity) between sites to similarity indices.

Usage

dissimilarity_to_similarity(dissimilaritydata)

Arguments

dissimilaritydata

the output object from similarity_to_dissimilarity() or a data.frame with the first columns called "Site1" and "Site2", and the other columns being the similarity metrics.

Value

A data.frame with additional class bioregion.pairwise.metric, providing one or several dissimilarity metric(s) between each pair of sites. The two first columns represent each pair of sites, and the other column represent dissimilarity metrics. Columns with names "a", "b", "c", "A", "B" and "C" are not altered.

Note

The behaviour of this function changes depending on column names. Columns "Site1" and "Site2" are copied identically. If there are columns called "a", "b", "c", "A", "B", "C", they will also be copied identically.

If a column is called "Euclidean", its similarity will be calculated based on the following formula:

\(Euclidean similarity = (1 - Euclidean distance) / Euclidean distance\)

Otherwise, all other columns will be transformed into similarity with the following formula:

\(similarity = 1 - dissimilarity\)

See also

Author

Boris Leroy (leroy.boris@gmail.com), Maxime Lenormand (maxime.lenormand@inrae.fr) and Pierre Denelle (pierre.denelle@gmail.com)

Examples

comat <- matrix(sample(0:1000, size = 50, replace = TRUE,
prob = 1 / 1:1001), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

simil <- similarity(comat, metric = "all")
simil
#> Data.frame of similarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of similarity metrics:  7 
#> 
#> 
#>    Site1 Site2 Jaccard Jaccardturn  Sorensen   Simpson       Bray  Brayturn
#> 2  Site1 Site2     0.6        0.75 0.7500000 0.8571429 0.20532003 0.3110831
#> 3  Site1 Site3     0.7        1.00 0.8235294 1.0000000 0.16638274 0.1817618
#> 4  Site1 Site4     0.6        0.75 0.7500000 0.8571429 0.12663237 0.1748634
#> 5  Site1 Site5     0.6        0.75 0.7500000 0.8571429 0.12841328 0.1619107
#> 8  Site2 Site3     0.9        1.00 0.9473684 1.0000000 0.23742604 0.4042821
#> 9  Site2 Site4     1.0        1.00 1.0000000 1.0000000 0.22703335 0.2443325
#> 10 Site2 Site5     0.8        0.80 0.8888889 0.8888889 0.05605174 0.1146096
#> 14 Site3 Site4     0.9        1.00 0.9473684 1.0000000 0.26973451 0.4163934
#> 15 Site3 Site5     0.9        1.00 0.9473684 1.0000000 0.30575292 0.3492147
#> 20 Site4 Site5     0.8        0.80 0.8888889 0.8888889 0.07660333 0.1409836
#>       Euclidean a b c   A    B    C
#> 2  0.0010639353 6 1 3 247 1365  547
#> 3  0.0007560550 7 0 3 293 1319 1617
#> 4  0.0008871552 6 1 3 160 1452  755
#> 5  0.0006431077 6 1 3 261 1351 2192
#> 8  0.0008813730 9 0 1 321  473 1589
#> 9  0.0013748426 9 0 0 194  600  721
#> 10 0.0006922112 8 1 1  91  703 2362
#> 14 0.0009103962 9 1 0 381 1529  534
#> 15 0.0006854524 9 1 0 667 1243 1786
#> 20 0.0006765728 8 1 1 129  786 2324

dissimilarity <- similarity_to_dissimilarity(simil)
dissimilarity
#> Data.frame of dissimilarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of dissimilarity metrics:  7 
#> 
#> 
#>    Site1 Site2 Jaccard Jaccardturn   Sorensen   Simpson      Bray  Brayturn
#> 2  Site1 Site2     0.4        0.25 0.25000000 0.1428571 0.7946800 0.6889169
#> 3  Site1 Site3     0.3        0.00 0.17647059 0.0000000 0.8336173 0.8182382
#> 4  Site1 Site4     0.4        0.25 0.25000000 0.1428571 0.8733676 0.8251366
#> 5  Site1 Site5     0.4        0.25 0.25000000 0.1428571 0.8715867 0.8380893
#> 8  Site2 Site3     0.1        0.00 0.05263158 0.0000000 0.7625740 0.5957179
#> 9  Site2 Site4     0.0        0.00 0.00000000 0.0000000 0.7729666 0.7556675
#> 10 Site2 Site5     0.2        0.20 0.11111111 0.1111111 0.9439483 0.8853904
#> 14 Site3 Site4     0.1        0.00 0.05263158 0.0000000 0.7302655 0.5836066
#> 15 Site3 Site5     0.1        0.00 0.05263158 0.0000000 0.6942471 0.6507853
#> 20 Site4 Site5     0.2        0.20 0.11111111 0.1111111 0.9233967 0.8590164
#>    Euclidean a b c   A    B    C
#> 2   938.9068 6 1 3 247 1365  547
#> 3  1321.6550 7 0 3 293 1319 1617
#> 4  1126.1985 6 1 3 160 1452  755
#> 5  1553.9495 6 1 3 261 1351 2192
#> 8  1133.5934 9 0 1 321  473 1589
#> 9   726.3560 9 0 0 194  600  721
#> 10 1443.6457 8 1 1  91  703 2362
#> 14 1097.4229 9 1 0 381 1529  534
#> 15 1457.8906 9 1 0 667 1243 1786
#> 20 1477.0376 8 1 1 129  786 2324

simil <- dissimilarity_to_similarity(dissimilarity)
simil
#> Data.frame of dissimilarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of dissimilarity metrics:  7 
#> 
#> 
#>    Site1 Site2 Jaccard Jaccardturn  Sorensen   Simpson       Bray  Brayturn
#> 2  Site1 Site2     0.6        0.75 0.7500000 0.8571429 0.20532003 0.3110831
#> 3  Site1 Site3     0.7        1.00 0.8235294 1.0000000 0.16638274 0.1817618
#> 4  Site1 Site4     0.6        0.75 0.7500000 0.8571429 0.12663237 0.1748634
#> 5  Site1 Site5     0.6        0.75 0.7500000 0.8571429 0.12841328 0.1619107
#> 8  Site2 Site3     0.9        1.00 0.9473684 1.0000000 0.23742604 0.4042821
#> 9  Site2 Site4     1.0        1.00 1.0000000 1.0000000 0.22703335 0.2443325
#> 10 Site2 Site5     0.8        0.80 0.8888889 0.8888889 0.05605174 0.1146096
#> 14 Site3 Site4     0.9        1.00 0.9473684 1.0000000 0.26973451 0.4163934
#> 15 Site3 Site5     0.9        1.00 0.9473684 1.0000000 0.30575292 0.3492147
#> 20 Site4 Site5     0.8        0.80 0.8888889 0.8888889 0.07660333 0.1409836
#>       Euclidean a b c   A    B    C
#> 2  0.0010639353 6 1 3 247 1365  547
#> 3  0.0007560550 7 0 3 293 1319 1617
#> 4  0.0008871552 6 1 3 160 1452  755
#> 5  0.0006431077 6 1 3 261 1351 2192
#> 8  0.0008813730 9 0 1 321  473 1589
#> 9  0.0013748426 9 0 0 194  600  721
#> 10 0.0006922112 8 1 1  91  703 2362
#> 14 0.0009103962 9 1 0 381 1529  534
#> 15 0.0006854524 9 1 0 667 1243 1786
#> 20 0.0006765728 8 1 1 129  786 2324