Community structure detection in weighted bipartite network via modularity optimization

This function takes a bipartite weighted graph and computes modules by applying Newman’s modularity measure in a bipartite weighted version to it.

Usage

netclu_beckett(
  net,
  weight = TRUE,
  index = names(net)[3],
  forceLPA = FALSE,
  site_col = 1,
  species_col = 2,
  return_node_type = "both",
  algorithm_in_output = TRUE
)

Arguments

net

a data.frame representing a bipartite network with the two first columns as undirected links between pair of nodes and and the next column(s) are the weight of the links.

weight

a boolean indicating if the weights should be considered if there are more than two columns (see Note).

index

name or number of the column to use as weight. By default, the third column name of net is used.

forceLPA

a boolean indicating if the even faster pure LPA-algorithm of Beckett should be used? DIRT-LPA, the default, is less likely to get trapped in a local minimum, but is slightly slower. Defaults to FALSE.

site_col

name or number for the column of site nodes (i.e. primary nodes).

species_col

name or number for the column of species nodes (i.e. feature nodes).

return_node_type

a character indicating what types of nodes ("sites", "species" or "both") should be returned in the output (return_node_type = "both" by default).

algorithm_in_output

a boolean indicating if the original output of computeModules should be returned in the output (TRUE by default, see Value).

Value

A list of class bioregion.clusters with five slots:

name: character containing the name of the algorithm

args: list of input arguments as provided by the user

inputs: list of characteristics of the clustering process

algorithm: list of all objects associated with the clustering procedure, such as original cluster objects (only if algorithm_in_output = TRUE)

clusters: data.frame containing the clustering results

In the algorithm slot, if algorithm_in_output = TRUE, users can find the output of computeModules.

Details

This function is based on the modularity optimization algorithm provided by Stephen Beckett (Beckett 2016) as implemented in the bipartite package (computeModules).

Note

Beckett has been designed to deal with weighted bipartite networks. Note that if weight = FALSE, a weight of 1 will be assigned to each pair of nodes. Do not forget to indicate which of the first two columns is dedicated to the site nodes (i.e. primary nodes) and species nodes (i.e. feature nodes) using the arguments site_col and species_col. The type of nodes returned in the output can be chosen with the argument return_node_type equal to "both" to keep both types of nodes,"sites" to preserve only the sites nodes and "species" to preserve only the species nodes.

References

Beckett SJ (2016). “Improved community detection in weighted bipartite networks.” Royal Society Open Science, 3(1), 140536.

See also

Author

Maxime Lenormand (maxime.lenormand@inrae.fr), Pierre Denelle (pierre.denelle@gmail.com) and Boris Leroy (leroy.boris@gmail.com)

Examples

net <- data.frame(
  Site = c(rep("A", 2), rep("B", 3), rep("C", 2)),
  Species = c("a", "b", "a", "c", "d", "b", "d"),
  Weight = c(10, 100, 1, 20, 50, 10, 20))

com <- netclu_beckett(net)