Convert similarity metrics to dissimilarity metrics

This function converts a data.frame of similarity metrics between sites to dissimilarity metrics (beta diversity).

Usage

similarity_to_dissimilarity(similarity, include_formula = TRUE)

Arguments

similarity

the output object from similarity() or dissimilarity_to_similarity().

include_formula

a boolean indicating if the metrics based on your own formula(s) should be converted (see Details). This argument is set to TRUE by default.

Value

A data.frame with additional class bioregion.pairwise.metric, providing dissimilarity metric(s) between each pair of sites based on a similarity object.

Note

The behavior of this function changes depending on column names. Columns Site1 and Site2 are copied identically. If there are columns called a, b, c, A, B, C they will also be copied identically. If there are columns based on your own formula (argument formula in similarity()) or not in the original list of similarity metrics (argument metrics in similarity()) and if the argument include_formula is set to FALSE, they will also be copied identically. Otherwise there are going to be converted like they other columns (default behavior).

If a column is called Euclidean, its distance will be calculated based on the following formula:

Euclidean distance = (1 - Euclidean similarity) / Euclidean similarityEuclidean distance = (1 - Euclidean similarity) / Euclidean similarity

Otherwise, all other columns will be transformed into dissimilarity with the following formula:

dissimilarity = 1 - similaritydissimilarity = 1 - similarity

See also

Author

Maxime Lenormand (maxime.lenormand@inrae.fr), Boris Leroy (leroy.boris@gmail.com) and Pierre Denelle (pierre.denelle@gmail.com)

Examples

comat <- matrix(sample(0:1000, size = 50, replace = TRUE,
prob = 1 / 1:1001), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

simil <- similarity(comat, metric = "all")
simil
#> Data.frame of similarity between sites
#>  - Total number of sites:  5 
#>  - Total number of species:  10 
#>  - Number of rows:  10 
#>  - Number of similarity metrics:  7 
#> 
#> 
#>    Site1 Site2 Jaccard Jaccardturn  Sorensen   Simpson       Bray  Brayturn
#> 2  Site1 Site2     0.9   1.0000000 0.9473684 1.0000000 0.29383886 0.4092409
#> 3  Site1 Site3     0.9   1.0000000 0.9473684 1.0000000 0.18653846 0.2235023
#> 4  Site1 Site4     1.0   1.0000000 1.0000000 1.0000000 0.31585256 0.8696370
#> 5  Site1 Site5     0.8   1.0000000 0.8888889 1.0000000 0.16043847 0.2656766
#> 8  Site2 Site3     0.8   0.8000000 0.8888889 0.8888889 0.19525066 0.3410138
#> 9  Site2 Site4     0.9   1.0000000 0.9473684 1.0000000 0.13690008 0.2412200
#> 10 Site2 Site5     0.7   0.7777778 0.8235294 0.8750000 0.34716069 0.3983364
#> 14 Site3 Site4     0.9   1.0000000 0.9473684 1.0000000 0.09162717 0.3341014
#> 15 Site3 Site5     0.7   0.7777778 0.8235294 0.8750000 0.31825613 0.6728111
#> 20 Site4 Site5     0.8   1.0000000 0.8888889 1.0000000 0.41142304 0.6067095
#>       Euclidean  a b c   A    B    C
#> 2  0.0018313203  9 1 0 248  358  834
#> 3  0.0023847158  9 1 0  97  509  337
#> 4  0.0008225182 10 0 0 527   79 2204
#> 5  0.0012693631  8 2 0 161  445 1240
#> 8  0.0017895170  8 1 1 148  934  286
#> 9  0.0007393107  9 0 1 261  821 2470
#> 10 0.0012224847  7 2 1 431  651  970
#> 14 0.0007304974  9 0 1 145  289 2586
#> 15 0.0013313747  7 2 1 292  142 1109
#> 20 0.0008358470  8 2 0 850 1881  551

dissimilarity <- similarity_to_dissimilarity(simil)
dissimilarity
#> Data.frame of dissimilarity between sites
#>  - Total number of sites:  5 
#>  - Total number of species:  10 
#>  - Number of rows:  10 
#>  - Number of dissimilarity metrics:  7 
#> 
#> 
#>    Site1 Site2 Jaccard Jaccardturn   Sorensen   Simpson      Bray  Brayturn
#> 2  Site1 Site2     0.1   0.0000000 0.05263158 0.0000000 0.7061611 0.5907591
#> 3  Site1 Site3     0.1   0.0000000 0.05263158 0.0000000 0.8134615 0.7764977
#> 4  Site1 Site4     0.0   0.0000000 0.00000000 0.0000000 0.6841474 0.1303630
#> 5  Site1 Site5     0.2   0.0000000 0.11111111 0.0000000 0.8395615 0.7343234
#> 8  Site2 Site3     0.2   0.2000000 0.11111111 0.1111111 0.8047493 0.6589862
#> 9  Site2 Site4     0.1   0.0000000 0.05263158 0.0000000 0.8630999 0.7587800
#> 10 Site2 Site5     0.3   0.2222222 0.17647059 0.1250000 0.6528393 0.6016636
#> 14 Site3 Site4     0.1   0.0000000 0.05263158 0.0000000 0.9083728 0.6658986
#> 15 Site3 Site5     0.3   0.2222222 0.17647059 0.1250000 0.6817439 0.3271889
#> 20 Site4 Site5     0.2   0.0000000 0.11111111 0.0000000 0.5885770 0.3932905
#>    Euclidean  a b c   A    B    C
#> 2   545.0541  9 1 0 248  358  834
#> 3   418.3372  9 1 0  97  509  337
#> 4  1214.7786 10 0 0 527   79 2204
#> 5   786.7967  8 2 0 161  445 1240
#> 8   557.8100  8 1 1 148  934  286
#> 9  1351.6113  9 0 1 261  821 2470
#> 10  817.0061  7 2 1 431  651  970
#> 14 1367.9302  9 0 1 145  289 2586
#> 15  750.1033  7 2 1 292  142 1109
#> 20 1195.3911  8 2 0 850 1881  551