Convert similarity metrics to dissimilarity metrics

This function converts a data.frame of similarity metrics between sites to dissimilarity metrics (beta diversity).

Usage

similarity_to_dissimilarity(similarity, include_formula = TRUE)

Arguments

similarity

the output object from similarity() or dissimilarity_to_similarity().

include_formula

a boolean indicating if the metrics based on your own formula(s) should be converted (see Details). This argument is set to TRUE by default.

Value

A data.frame with additional class bioregion.pairwise.metric, providing dissimilarity metric(s) between each pair of sites based on a similarity object.

Note

The behavior of this function changes depending on column names. Columns Site1 and Site2 are copied identically. If there are columns called a, b, c, A, B, C they will also be copied identically. If there are columns based on your own formula (argument formula in similarity()) or not in the original list of similarity metrics (argument metrics in similarity()) and if the argument include_formula is set to FALSE, they will also be copied identically. Otherwise there are going to be converted like they other columns (default behavior).

If a column is called Euclidean, its distance will be calculated based on the following formula:

Euclidean distance = (1 - Euclidean similarity) / Euclidean similarityEuclidean distance = (1 - Euclidean similarity) / Euclidean similarity

Otherwise, all other columns will be transformed into dissimilarity with the following formula:

dissimilarity = 1 - similaritydissimilarity = 1 - similarity

See also

Author

Maxime Lenormand (maxime.lenormand@inrae.fr), Boris Leroy (leroy.boris@gmail.com) and Pierre Denelle (pierre.denelle@gmail.com)

Examples

comat <- matrix(sample(0:1000, size = 50, replace = TRUE,
prob = 1 / 1:1001), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

simil <- similarity(comat, metric = "all")
simil
#> Data.frame of similarity between sites
#>  - Total number of sites:  5 
#>  - Total number of species:  10 
#>  - Number of rows:  10 
#>  - Number of similarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn  Sorensen   Simpson       Bray   Brayturn
#> 2  Site1 Site2 0.4000000   0.6666667 0.5714286 0.8000000 0.03498191 0.10320285
#> 3  Site1 Site3 0.5000000   1.0000000 0.6666667 1.0000000 0.04893617 0.16370107
#> 4  Site1 Site4 0.3333333   0.4285714 0.5000000 0.6000000 0.04769001 0.05693950
#> 5  Site1 Site5 0.4000000   0.6666667 0.5714286 0.8000000 0.07166124 0.07829181
#> 8  Site2 Site3 0.9000000   1.0000000 0.9473684 1.0000000 0.47715054 0.51561365
#> 9  Site2 Site4 0.6000000   0.7500000 0.7500000 0.8571429 0.19807583 0.44871795
#> 10 Site2 Site5 1.0000000   1.0000000 1.0000000 1.0000000 0.35204678 0.90390390
#> 14 Site3 Site4 0.7000000   1.0000000 0.8235294 1.0000000 0.14278532 0.36410256
#> 15 Site3 Site5 0.9000000   1.0000000 0.9473684 1.0000000 0.08281573 0.24024024
#> 20 Site4 Site5 0.6000000   0.7500000 0.7500000 0.8571429 0.16320885 0.17717718
#>       Euclidean a b c   A    B    C
#> 2  0.0014046151 4 1 5  29  252 1348
#> 3  0.0009391180 5 0 5  46  235 1553
#> 4  0.0029375342 3 2 4  16  265  374
#> 5  0.0032945745 4 1 5  22  259  311
#> 8  0.0011952848 9 0 1 710  667  889
#> 9  0.0015083936 6 3 1 175 1202  215
#> 10 0.0017499806 9 0 0 301 1076   32
#> 14 0.0009748313 7 3 0 142 1457  248
#> 15 0.0009587939 9 1 0  80 1519  253
#> 20 0.0035183700 6 1 3  59  331  274

dissimilarity <- similarity_to_dissimilarity(simil)
dissimilarity
#> Data.frame of dissimilarity between sites
#>  - Total number of sites:  5 
#>  - Total number of species:  10 
#>  - Number of rows:  10 
#>  - Number of dissimilarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn   Sorensen   Simpson      Bray  Brayturn
#> 2  Site1 Site2 0.6000000   0.3333333 0.42857143 0.2000000 0.9650181 0.8967972
#> 3  Site1 Site3 0.5000000   0.0000000 0.33333333 0.0000000 0.9510638 0.8362989
#> 4  Site1 Site4 0.6666667   0.5714286 0.50000000 0.4000000 0.9523100 0.9430605
#> 5  Site1 Site5 0.6000000   0.3333333 0.42857143 0.2000000 0.9283388 0.9217082
#> 8  Site2 Site3 0.1000000   0.0000000 0.05263158 0.0000000 0.5228495 0.4843863
#> 9  Site2 Site4 0.4000000   0.2500000 0.25000000 0.1428571 0.8019242 0.5512821
#> 10 Site2 Site5 0.0000000   0.0000000 0.00000000 0.0000000 0.6479532 0.0960961
#> 14 Site3 Site4 0.3000000   0.0000000 0.17647059 0.0000000 0.8572147 0.6358974
#> 15 Site3 Site5 0.1000000   0.0000000 0.05263158 0.0000000 0.9171843 0.7597598
#> 20 Site4 Site5 0.4000000   0.2500000 0.25000000 0.1428571 0.8367911 0.8228228
#>    Euclidean a b c   A    B    C
#> 2   710.9388 4 1 5  29  252 1348
#> 3  1063.8289 5 0 5  46  235 1553
#> 4   339.4216 3 2 4  16  265  374
#> 5   302.5293 4 1 5  22  259  311
#> 8   835.6207 9 0 1 710  667  889
#> 9   661.9569 6 3 1 175 1202  215
#> 10  570.4349 9 0 0 301 1076   32
#> 14 1024.8185 7 3 0 142 1457  248
#> 15 1041.9770 9 1 0  80 1519  253
#> 20  283.2225 6 1 3  59  331  274