Convert similarity indices to dissimilarity metrics

This function converts a data.frame of similarity metrics between sites to dissimilarity metrics (= beta diversity).

Usage

similarity_to_dissimilarity(similarity)

Arguments

similarity

the output object from similarity() or a data.frame with the first columns called "Site1" and "Site2", and the other columns being the similarity metrics.

Value

A data.frame with additional class bioregion.pairwise.metric, providing one or several similarity metric(s) between each pair of sites. The two first columns represent each pair of sites, and the other column represent similarity metrics. Columns with names "a", "b", "c", "A", "B" and "C" are not altered.

Note

The behavior of this function changes depending on column names. Columns "Site1" and "Site2" are copied identically. If there are columns called "a", "b", "c", "A", "B", "C", they will also be copied identically.

If a column is called "Euclidean", its distance will be calculated based on the following formula:

\(Euclidean distance = (1 - Euclidean similarity) / Euclidean similarity\)

Otherwise, all other columns will be transformed into dissimilarity with the following formula:

\(dissimilarity = 1 - similarity\)

See also

Author

Boris Leroy (leroy.boris@gmail.com), Maxime Lenormand (maxime.lenormand@inrae.fr) and Pierre Denelle (pierre.denelle@gmail.com)

Examples

# \donttest{
comat <- matrix(sample(0:1000, size = 50, replace = TRUE,
prob = 1 / 1:1001), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

simil <- similarity(comat, metric = "all")
simil
#> Data.frame of similarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of similarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn  Sorensen   Simpson       Bray   Brayturn
#> 2  Site1 Site2 0.7777778   1.0000000 0.8750000 1.0000000 0.09878419 0.15081206
#> 3  Site1 Site3 0.6666667   0.7500000 0.8000000 0.8571429 0.03375068 0.03502825
#> 4  Site1 Site4 0.4444444   0.5000000 0.6153846 0.6666667 0.38879159 0.86381323
#> 5  Site1 Site5 0.7500000   0.7500000 0.8571429 0.8571429 0.03043868 0.07327586
#> 8  Site2 Site3 0.7000000   0.7777778 0.8235294 0.8750000 0.38611714 0.61948956
#> 9  Site2 Site4 0.5000000   0.7142857 0.6666667 0.8333333 0.17151163 0.22957198
#> 10 Site2 Site5 0.7777778   1.0000000 0.8750000 1.0000000 0.08446456 0.12068966
#> 14 Site3 Site4 0.5555556   0.7142857 0.7142857 0.8333333 0.07113317 0.16731518
#> 15 Site3 Site5 0.6666667   0.7500000 0.8000000 0.8571429 0.03547297 0.09051724
#> 20 Site4 Site5 0.3000000   0.3333333 0.4615385 0.5000000 0.01226994 0.01293103
#>       Euclidean a b c   A   B   C
#> 2  0.0015776350 7 0 2  65 820 366
#> 3  0.0009892006 6 1 2  31 854 921
#> 4  0.0018601241 4 3 2 222 663  35
#> 5  0.0015391841 6 1 1  17 868 215
#> 8  0.0014993263 7 2 1 267 164 685
#> 9  0.0034077158 5 4 1  59 372 198
#> 10 0.0032747529 7 2 0  28 403 204
#> 14 0.0011962699 5 3 1  43 909 214
#> 15 0.0011935684 6 2 1  21 931 211
#> 20 0.0033697451 3 3 4   3 254 229

dissimilarity <- similarity_to_dissimilarity(simil)
dissimilarity
#> Data.frame of dissimilarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of dissimilarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn  Sorensen   Simpson      Bray  Brayturn
#> 2  Site1 Site2 0.2222222   0.0000000 0.1250000 0.0000000 0.9012158 0.8491879
#> 3  Site1 Site3 0.3333333   0.2500000 0.2000000 0.1428571 0.9662493 0.9649718
#> 4  Site1 Site4 0.5555556   0.5000000 0.3846154 0.3333333 0.6112084 0.1361868
#> 5  Site1 Site5 0.2500000   0.2500000 0.1428571 0.1428571 0.9695613 0.9267241
#> 8  Site2 Site3 0.3000000   0.2222222 0.1764706 0.1250000 0.6138829 0.3805104
#> 9  Site2 Site4 0.5000000   0.2857143 0.3333333 0.1666667 0.8284884 0.7704280
#> 10 Site2 Site5 0.2222222   0.0000000 0.1250000 0.0000000 0.9155354 0.8793103
#> 14 Site3 Site4 0.4444444   0.2857143 0.2857143 0.1666667 0.9288668 0.8326848
#> 15 Site3 Site5 0.3333333   0.2500000 0.2000000 0.1428571 0.9645270 0.9094828
#> 20 Site4 Site5 0.7000000   0.6666667 0.5384615 0.5000000 0.9877301 0.9870690
#>    Euclidean a b c   A   B   C
#> 2   632.8602 7 0 2  65 820 366
#> 3  1009.9173 6 1 2  31 854 921
#> 4   536.5985 4 3 2 222 663  35
#> 5   648.6948 6 1 1  17 868 215
#> 8   665.9662 7 2 1 267 164 685
#> 9   292.4517 5 4 1  59 372 198
#> 10  304.3666 7 2 0  28 403 204
#> 14  834.9317 5 3 1  43 909 214
#> 15  836.8238 6 2 1  21 931 211
#> 20  295.7583 3 3 4   3 254 229
# }