Convert similarity metrics to dissimilarity metrics

This function converts a data.frame of similarity metrics between sites to dissimilarity metrics (beta diversity).

Usage

similarity_to_dissimilarity(similarity, include_formula = TRUE)

Arguments

similarity

the output object from similarity() or dissimilarity_to_similarity().

include_formula

a boolean indicating if the metrics based on your own formula should be converted (see Details). This argument is set to TRUE by default.

Value

A data.frame with additional class bioregion.pairwise.metric, providing dissimilarity metric(s) between each pair of sites based on a similarity object.

Note

The behavior of this function changes depending on column names. Columns Site1 and Site2 are copied identically. If there are columns called a, b, c, A, B, C they will also be copied identically. If there are columns based on your own formula (argument formula in similarity()) or not in the original list of similarity metrics (argument metrics in similarity()) and if the argument include_formula is set to FALSE, they will also be copied identically. Otherwise there are going to be converted like they other columns (default behavior).

If a column is called Euclidean, its distance will be calculated based on the following formula:

\(Euclidean distance = (1 - Euclidean similarity) / Euclidean similarity\)

Otherwise, all other columns will be transformed into dissimilarity with the following formula:

\(dissimilarity = 1 - similarity\)

See also

Author

Maxime Lenormand (maxime.lenormand@inrae.fr), Boris Leroy (leroy.boris@gmail.com) and Pierre Denelle (pierre.denelle@gmail.com)

Examples

comat <- matrix(sample(0:1000, size = 50, replace = TRUE,
prob = 1 / 1:1001), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

simil <- similarity(comat, metric = "all")
simil
#> Data.frame of similarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of similarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn  Sorensen   Simpson       Bray  Brayturn
#> 2  Site1 Site2 0.7000000   0.7777778 0.8235294 0.8750000 0.22608696 0.2488038
#> 3  Site1 Site3 0.6666667   0.7500000 0.8000000 0.8571429 0.06472847 0.0850757
#> 4  Site1 Site4 0.7000000   0.7777778 0.8235294 0.8750000 0.03228022 0.0719755
#> 5  Site1 Site5 0.8000000   1.0000000 0.8888889 1.0000000 0.10250818 0.1334280
#> 8  Site2 Site3 0.7777778   1.0000000 0.8750000 1.0000000 0.45654835 0.5378515
#> 9  Site2 Site4 0.8000000   0.8000000 0.8888889 0.8888889 0.12233623 0.2373660
#> 10 Site2 Site5 0.9000000   1.0000000 0.9473684 1.0000000 0.33252280 0.3882186
#> 14 Site3 Site4 0.6000000   0.7500000 0.7500000 0.8571429 0.14117647 0.2205207
#> 15 Site3 Site5 0.7000000   1.0000000 0.8235294 1.0000000 0.43133047 0.4347513
#> 20 Site4 Site5 0.9000000   1.0000000 0.9473684 1.0000000 0.31134821 0.4915773
#>       Euclidean a b c   A    B    C
#> 2  0.0007644511 7 1 2 468 1791 1413
#> 3  0.0006681295 6 2 1 118 2141 1269
#> 4  0.0007864328 7 1 2  47 2212  606
#> 5  0.0007290523 8 0 2 188 2071 1221
#> 8  0.0011930955 7 2 0 746 1135  641
#> 9  0.0010570334 8 1 1 155 1726  498
#> 10 0.0010062702 9 0 1 547 1334  862
#> 14 0.0010556010 6 1 3 144 1243  509
#> 15 0.0011426674 7 0 3 603  784  806
#> 20 0.0013765908 9 0 1 321  332 1088

dissimilarity <- similarity_to_dissimilarity(simil)
dissimilarity
#> Data.frame of dissimilarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of dissimilarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn   Sorensen   Simpson      Bray  Brayturn
#> 2  Site1 Site2 0.3000000   0.2222222 0.17647059 0.1250000 0.7739130 0.7511962
#> 3  Site1 Site3 0.3333333   0.2500000 0.20000000 0.1428571 0.9352715 0.9149243
#> 4  Site1 Site4 0.3000000   0.2222222 0.17647059 0.1250000 0.9677198 0.9280245
#> 5  Site1 Site5 0.2000000   0.0000000 0.11111111 0.0000000 0.8974918 0.8665720
#> 8  Site2 Site3 0.2222222   0.0000000 0.12500000 0.0000000 0.5434517 0.4621485
#> 9  Site2 Site4 0.2000000   0.2000000 0.11111111 0.1111111 0.8776638 0.7626340
#> 10 Site2 Site5 0.1000000   0.0000000 0.05263158 0.0000000 0.6674772 0.6117814
#> 14 Site3 Site4 0.4000000   0.2500000 0.25000000 0.1428571 0.8588235 0.7794793
#> 15 Site3 Site5 0.3000000   0.0000000 0.17647059 0.0000000 0.5686695 0.5652487
#> 20 Site4 Site5 0.1000000   0.0000000 0.05263158 0.0000000 0.6886518 0.5084227
#>    Euclidean a b c   A    B    C
#> 2  1307.1281 7 1 2 468 1791 1413
#> 3  1495.7159 6 2 1 118 2141 1269
#> 4  1270.5644 7 1 2  47 2212  606
#> 5  1370.6436 8 0 2 188 2071 1221
#> 8   837.1559 7 2 0 746 1135  641
#> 9   945.0439 8 1 1 155 1726  498
#> 10  992.7689 9 0 1 547 1334  862
#> 14  946.3276 6 1 3 144 1243  509
#> 15  874.1453 7 0 3 603  784  806
#> 20  725.4323 9 0 1 321  332 1088