Convert similarity metrics to dissimilarity metrics

This function converts a data.frame of similarity metrics between sites to dissimilarity metrics (beta diversity).

Usage

similarity_to_dissimilarity(similarity, include_formula = TRUE)

Arguments

similarity

the output object from similarity() or dissimilarity_to_similarity().

include_formula

a boolean indicating if the metrics based on your own formula should be converted (see Details). This argument is set to TRUE by default.

Value

A data.frame with additional class bioregion.pairwise.metric, providing dissimilarity metric(s) between each pair of sites based on a similarity object.

Note

The behavior of this function changes depending on column names. Columns Site1 and Site2 are copied identically. If there are columns called a, b, c, A, B, C they will also be copied identically. If there are columns based on your own formula (argument formula in similarity()) or not in the original list of similarity metrics (argument metrics in similarity()) and if the argument include_formula is set to FALSE, they will also be copied identically. Otherwise there are going to be converted like they other columns (default behavior).

If a column is called Euclidean, its distance will be calculated based on the following formula:

\(Euclidean distance = (1 - Euclidean similarity) / Euclidean similarity\)

Otherwise, all other columns will be transformed into dissimilarity with the following formula:

\(dissimilarity = 1 - similarity\)

See also

Author

Maxime Lenormand (maxime.lenormand@inrae.fr), Boris Leroy (leroy.boris@gmail.com) and Pierre Denelle (pierre.denelle@gmail.com)

Examples

comat <- matrix(sample(0:1000, size = 50, replace = TRUE,
prob = 1 / 1:1001), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

simil <- similarity(comat, metric = "all")
simil
#> Data.frame of similarity between sites
#>  - Total number of sites:  5 
#>  - Total number of species:  10 
#>  - Number of rows:  10 
#>  - Number of similarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn  Sorensen   Simpson       Bray  Brayturn
#> 2  Site1 Site2 0.9000000        1.00 0.9473684 1.0000000 0.19003260 0.6986301
#> 3  Site1 Site3 0.9000000        1.00 0.9473684 1.0000000 0.28255208 0.7431507
#> 4  Site1 Site4 0.7000000        1.00 0.8235294 1.0000000 0.32200358 0.3370787
#> 5  Site1 Site5 1.0000000        1.00 1.0000000 1.0000000 0.32536765 0.6061644
#> 8  Site2 Site3 0.8000000        0.80 0.8888889 0.8888889 0.47563730 0.5924437
#> 9  Site2 Site4 0.7777778        1.00 0.8750000 1.0000000 0.08765316 0.3483146
#> 10 Site2 Site5 0.9000000        1.00 0.9473684 1.0000000 0.15918521 0.2650754
#> 14 Site3 Site4 0.6000000        0.75 0.7500000 0.8571429 0.23957644 0.6779026
#> 15 Site3 Site5 0.9000000        1.00 0.9473684 1.0000000 0.13823529 0.1771357
#> 20 Site4 Site5 0.7000000        1.00 0.8235294 1.0000000 0.11853246 0.2359551
#>       Euclidean  a b c   A    B    C
#> 2  0.0009748777  9 1 0 204   88 1651
#> 3  0.0014700041  9 1 0 217   75 1027
#> 4  0.0053226323  7 3 0  90  202  177
#> 5  0.0025397200 10 0 0 177  115  619
#> 8  0.0010851854  8 1 1 737 1118  507
#> 9  0.0009096490  7 2 0  93 1762  174
#> 10 0.0008857288  9 0 1 211 1644  585
#> 14 0.0013022514  6 3 1 181 1063   86
#> 15 0.0012042340  9 0 1 141 1103  655
#> 20 0.0021905133  7 0 3  63  204  733

dissimilarity <- similarity_to_dissimilarity(simil)
dissimilarity
#> Data.frame of dissimilarity between sites
#>  - Total number of sites:  5 
#>  - Total number of species:  10 
#>  - Number of rows:  10 
#>  - Number of dissimilarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn   Sorensen   Simpson      Bray  Brayturn
#> 2  Site1 Site2 0.1000000        0.00 0.05263158 0.0000000 0.8099674 0.3013699
#> 3  Site1 Site3 0.1000000        0.00 0.05263158 0.0000000 0.7174479 0.2568493
#> 4  Site1 Site4 0.3000000        0.00 0.17647059 0.0000000 0.6779964 0.6629213
#> 5  Site1 Site5 0.0000000        0.00 0.00000000 0.0000000 0.6746324 0.3938356
#> 8  Site2 Site3 0.2000000        0.20 0.11111111 0.1111111 0.5243627 0.4075563
#> 9  Site2 Site4 0.2222222        0.00 0.12500000 0.0000000 0.9123468 0.6516854
#> 10 Site2 Site5 0.1000000        0.00 0.05263158 0.0000000 0.8408148 0.7349246
#> 14 Site3 Site4 0.4000000        0.25 0.25000000 0.1428571 0.7604236 0.3220974
#> 15 Site3 Site5 0.1000000        0.00 0.05263158 0.0000000 0.8617647 0.8228643
#> 20 Site4 Site5 0.3000000        0.00 0.17647059 0.0000000 0.8814675 0.7640449
#>    Euclidean  a b c   A    B    C
#> 2  1024.7697  9 1 0 204   88 1651
#> 3   679.2702  9 1 0 217   75 1027
#> 4   186.8770  7 3 0  90  202  177
#> 5   392.7442 10 0 0 177  115  619
#> 8   920.5015  8 1 1 737 1118  507
#> 9  1098.3251  7 2 0  93 1762  174
#> 10 1128.0137  9 0 1 211 1644  585
#> 14  766.9009  6 3 1 181 1063   86
#> 15  829.4034  9 0 1 141 1103  655
#> 20  455.5140  7 0 3  63  204  733