Convert similarity metrics to dissimilarity metrics

This function converts a data.frame of similarity metrics between sites into dissimilarity metrics (beta diversity).

Usage

similarity_to_dissimilarity(similarity, include_formula = TRUE)

Arguments

similarity

The output object from similarity() or dissimilarity_to_similarity().

include_formula

A boolean indicating whether metrics based on custom formula(s) should also be converted (see Details). The default is TRUE.

Value

A data.frame with additional class bioregion.pairwise.metric, providing dissimilarity metric(s) between each pair of sites based on a similarity object.

Note

The behavior of this function changes depending on column names. Columns Site1 and Site2 are copied identically. If there are columns called a, b, c, A, B, C they will also be copied identically. If there are columns based on your own formula (argument formula in similarity()) or not in the original list of similarity metrics (argument metrics in similarity()) and if the argument include_formula is set to FALSE, they will also be copied identically. Otherwise there are going to be converted like they other columns (default behavior).

If a column is called Euclidean, its distance will be calculated based on the following formula:

Euclidean distance = (1 - Euclidean similarity) / Euclidean similarity

Otherwise, all other columns will be transformed into dissimilarity with the following formula:

dissimilarity = 1 - similarity

See also

Author

Maxime Lenormand (maxime.lenormand@inrae.fr)
Boris Leroy (leroy.boris@gmail.com)
Pierre Denelle (pierre.denelle@gmail.com)

Examples

comat <- matrix(sample(0:1000, size = 50, replace = TRUE,
prob = 1 / 1:1001), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

simil <- similarity(comat, metric = "all")
simil
#> Data.frame of similarity between sites
#>  - Total number of sites:  5 
#>  - Total number of species:  10 
#>  - Number of rows:  10 
#>  - Number of similarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn  Sorensen   Simpson       Bray   Brayturn
#> 2  Site1 Site2 0.7000000   1.0000000 0.8235294 1.0000000 0.21861777 0.32908705
#> 3  Site1 Site3 1.0000000   1.0000000 1.0000000 1.0000000 0.24496322 0.31062449
#> 4  Site1 Site4 0.8000000   1.0000000 0.8888889 1.0000000 0.18330606 0.18961625
#> 5  Site1 Site5 0.8000000   1.0000000 0.8888889 1.0000000 0.51054650 0.56230201
#> 8  Site2 Site3 0.7000000   1.0000000 0.8235294 1.0000000 0.20873563 0.24097665
#> 9  Site2 Site4 0.5000000   0.5555556 0.6666667 0.7142857 0.04642594 0.06687898
#> 10 Site2 Site5 0.6666667   0.7500000 0.8000000 0.8571429 0.14596273 0.24946921
#> 14 Site3 Site4 0.8000000   1.0000000 0.8888889 1.0000000 0.10116473 0.12327656
#> 15 Site3 Site5 0.8000000   1.0000000 0.8888889 1.0000000 0.10082598 0.14355231
#> 20 Site4 Site5 0.6000000   0.6000000 0.7500000 0.7500000 0.45629630 0.52144470
#>       Euclidean  a b c    A    B    C
#> 2  0.0012216232  7 3 0  310 1584  632
#> 3  0.0009476402 10 0 0  383 1511  850
#> 4  0.0007660859  8 2 0  336 1558 1436
#> 5  0.0010691062  8 2 0 1065  829 1213
#> 8  0.0010053133  7 0 3  227  715 1006
#> 9  0.0007723967  5 2 3   63  879 1709
#> 10 0.0008873387  6 1 2  235  707 2043
#> 14 0.0006810961  8 2 0  152 1081 1620
#> 15 0.0007248472  8 2 0  177 1056 2101
#> 20 0.0010108979  6 2 2  924  848 1354

dissimilarity <- similarity_to_dissimilarity(simil)
dissimilarity
#> Data.frame of dissimilarity between sites
#>  - Total number of sites:  5 
#>  - Total number of species:  10 
#>  - Number of rows:  10 
#>  - Number of dissimilarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn  Sorensen   Simpson      Bray  Brayturn
#> 2  Site1 Site2 0.3000000   0.0000000 0.1764706 0.0000000 0.7813822 0.6709130
#> 3  Site1 Site3 0.0000000   0.0000000 0.0000000 0.0000000 0.7550368 0.6893755
#> 4  Site1 Site4 0.2000000   0.0000000 0.1111111 0.0000000 0.8166939 0.8103837
#> 5  Site1 Site5 0.2000000   0.0000000 0.1111111 0.0000000 0.4894535 0.4376980
#> 8  Site2 Site3 0.3000000   0.0000000 0.1764706 0.0000000 0.7912644 0.7590234
#> 9  Site2 Site4 0.5000000   0.4444444 0.3333333 0.2857143 0.9535741 0.9331210
#> 10 Site2 Site5 0.3333333   0.2500000 0.2000000 0.1428571 0.8540373 0.7505308
#> 14 Site3 Site4 0.2000000   0.0000000 0.1111111 0.0000000 0.8988353 0.8767234
#> 15 Site3 Site5 0.2000000   0.0000000 0.1111111 0.0000000 0.8991740 0.8564477
#> 20 Site4 Site5 0.4000000   0.4000000 0.2500000 0.2500000 0.5437037 0.4785553
#>    Euclidean  a b c    A    B    C
#> 2   817.5830  7 3 0  310 1584  632
#> 3  1054.2528 10 0 0  383 1511  850
#> 4  1304.3366  8 2 0  336 1558 1436
#> 5   934.3607  8 2 0 1065  829 1213
#> 8   993.7147  7 0 3  227  715 1006
#> 9  1293.6715  5 2 3   63  879 1709
#> 10 1125.9654  6 1 2  235  707 2043
#> 14 1467.2215  8 2 0  152 1081 1620
#> 15 1378.6011  8 2 0  177 1056 2101
#> 20  988.2196  6 2 2  924  848 1354