Convert similarity metrics to dissimilarity metrics

This function converts a data.frame of similarity metrics between sites to dissimilarity metrics (beta diversity).

Usage

similarity_to_dissimilarity(similarity, include_formula = TRUE)

Arguments

similarity

the output object from similarity() or dissimilarity_to_similarity().

include_formula

a boolean indicating if the metrics based on your own formula should be converted (see Details). This argument is set to TRUE by default.

Value

A data.frame with additional class bioregion.pairwise.metric, providing dissimilarity metric(s) between each pair of sites based on a similarity object.

Note

The behavior of this function changes depending on column names. Columns Site1 and Site2 are copied identically. If there are columns called a, b, c, A, B, C they will also be copied identically. If there are columns based on your own formula (argument formula in similarity()) or not in the original list of similarity metrics (argument metrics in similarity()) and if the argument include_formula is set to FALSE, they will also be copied identically. Otherwise there are going to be converted like they other columns (default behavior).

If a column is called Euclidean, its distance will be calculated based on the following formula:

\(Euclidean distance = (1 - Euclidean similarity) / Euclidean similarity\)

Otherwise, all other columns will be transformed into dissimilarity with the following formula:

\(dissimilarity = 1 - similarity\)

See also

Author

Maxime Lenormand (maxime.lenormand@inrae.fr), Boris Leroy (leroy.boris@gmail.com) and Pierre Denelle (pierre.denelle@gmail.com)

Examples

comat <- matrix(sample(0:1000, size = 50, replace = TRUE,
prob = 1 / 1:1001), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

simil <- similarity(comat, metric = "all")
simil
#> Data.frame of similarity between sites
#>  - Total number of sites:  5 
#>  - Total number of species:  10 
#>  - Number of rows:  10 
#>  - Number of similarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn  Sorensen   Simpson       Bray  Brayturn
#> 2  Site1 Site2 0.7000000   0.7777778 0.8235294 0.8750000 0.14110032 0.3096591
#> 3  Site1 Site3 0.7000000   0.7777778 0.8235294 0.8750000 0.27358848 0.4249559
#> 4  Site1 Site4 0.7000000   0.7777778 0.8235294 0.8750000 0.40523428 0.8283003
#> 5  Site1 Site5 0.8888889   1.0000000 0.9411765 1.0000000 0.28012253 0.3582934
#> 8  Site2 Site3 0.8000000   0.8000000 0.8888889 0.8888889 0.17132486 0.2234848
#> 9  Site2 Site4 0.8000000   0.8000000 0.8888889 0.8888889 0.28081264 0.2945076
#> 10 Site2 Site5 0.8000000   0.8000000 0.8888889 0.8888889 0.22487325 0.3570076
#> 14 Site3 Site4 0.8000000   0.8000000 0.8888889 0.8888889 0.14695591 0.1811907
#> 15 Site3 Site5 0.8000000   0.8000000 0.8888889 0.8888889 0.09109109 0.1071218
#> 20 Site4 Site5 0.8000000   0.8000000 0.8888889 0.8888889 0.21585648 0.3218292
#>       Euclidean a b c   A    B    C
#> 2  0.0006499738 7 1 2 327 3252  729
#> 3  0.0006018593 7 1 2 722 2857  977
#> 4  0.0007924331 7 1 2 960 2619  199
#> 5  0.0006041435 8 0 1 823 2756 1474
#> 8  0.0008457226 8 1 1 236  820 1463
#> 9  0.0013653040 8 1 1 311  745  848
#> 10 0.0007698501 8 1 1 377  679 1920
#> 14 0.0008356688 8 1 1 210 1489  949
#> 15 0.0006009373 8 1 1 182 1517 2115
#> 20 0.0007820249 8 1 1 373  786 1924

dissimilarity <- similarity_to_dissimilarity(simil)
dissimilarity
#> Data.frame of dissimilarity between sites
#>  - Total number of sites:  5 
#>  - Total number of species:  10 
#>  - Number of rows:  10 
#>  - Number of dissimilarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn   Sorensen   Simpson      Bray  Brayturn
#> 2  Site1 Site2 0.3000000   0.2222222 0.17647059 0.1250000 0.8588997 0.6903409
#> 3  Site1 Site3 0.3000000   0.2222222 0.17647059 0.1250000 0.7264115 0.5750441
#> 4  Site1 Site4 0.3000000   0.2222222 0.17647059 0.1250000 0.5947657 0.1716997
#> 5  Site1 Site5 0.1111111   0.0000000 0.05882353 0.0000000 0.7198775 0.6417066
#> 8  Site2 Site3 0.2000000   0.2000000 0.11111111 0.1111111 0.8286751 0.7765152
#> 9  Site2 Site4 0.2000000   0.2000000 0.11111111 0.1111111 0.7191874 0.7054924
#> 10 Site2 Site5 0.2000000   0.2000000 0.11111111 0.1111111 0.7751268 0.6429924
#> 14 Site3 Site4 0.2000000   0.2000000 0.11111111 0.1111111 0.8530441 0.8188093
#> 15 Site3 Site5 0.2000000   0.2000000 0.11111111 0.1111111 0.9089089 0.8928782
#> 20 Site4 Site5 0.2000000   0.2000000 0.11111111 0.1111111 0.7841435 0.6781708
#>    Euclidean a b c   A    B    C
#> 2  1537.5237 7 1 2 327 3252  729
#> 3  1660.5180 7 1 2 722 2857  977
#> 4  1260.9362 7 1 2 960 2619  199
#> 5  1654.2358 8 0 1 823 2756 1474
#> 8  1181.4208 8 1 1 236  820 1463
#> 9   731.4376 8 1 1 311  745  848
#> 10 1297.9542 8 1 1 377  679 1920
#> 14 1195.6463 8 1 1 210 1489  949
#> 15 1663.0670 8 1 1 182 1517 2115
#> 20 1277.7316 8 1 1 373  786 1924