Convert similarity indices to dissimilarity metrics

This function converts a data.frame of similarity metrics between sites to dissimilarity metrics (= beta diversity).

Usage

similarity_to_dissimilarity(similarity)

Arguments

similarity

the output object from similarity() or a data.frame with the first columns called "Site1" and "Site2", and the other columns being the similarity metrics.

Value

A data.frame with additional class bioregion.pairwise.metric, providing one or several similarity metric(s) between each pair of sites. The two first columns represent each pair of sites, and the other column represent similarity metrics. Columns with names "a", "b", "c", "A", "B" and "C" are not altered.

Note

The behavior of this function changes depending on column names. Columns "Site1" and "Site2" are copied identically. If there are columns called "a", "b", "c", "A", "B", "C", they will also be copied identically.

If a column is called "Euclidean", its distance will be calculated based on the following formula:

\(Euclidean distance = (1 - Euclidean similarity) / Euclidean similarity\)

Otherwise, all other columns will be transformed into dissimilarity with the following formula:

\(dissimilarity = 1 - similarity\)

See also

Author

Boris Leroy (leroy.boris@gmail.com), Maxime Lenormand (maxime.lenormand@inrae.fr) and Pierre Denelle (pierre.denelle@gmail.com)

Examples

comat <- matrix(sample(0:1000, size = 50, replace = TRUE,
prob = 1 / 1:1001), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

simil <- similarity(comat, metric = "all")
simil
#> Data.frame of similarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of similarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn  Sorensen   Simpson      Bray   Brayturn
#> 2  Site1 Site2 0.4444444   0.6666667 0.6153846 0.8000000 0.2527316 0.26814516
#> 3  Site1 Site3 0.6666667   0.7500000 0.8000000 0.8571429 0.3079300 0.31473684
#> 4  Site1 Site4 0.8000000   1.0000000 0.8888889 1.0000000 0.2573466 0.45463710
#> 5  Site1 Site5 0.7000000   0.7777778 0.8235294 0.8750000 0.1780105 0.34274194
#> 8  Site2 Site3 0.5000000   0.6666667 0.6666667 0.8000000 0.1134270 0.12315789
#> 9  Site2 Site4 0.5000000   1.0000000 0.6666667 1.0000000 0.5609487 0.91374663
#> 10 Site2 Site5 0.4000000   0.6666667 0.5714286 0.8000000 0.0416138 0.07367475
#> 14 Site3 Site4 0.7000000   1.0000000 0.8235294 1.0000000 0.4522091 0.82421053
#> 15 Site3 Site5 0.6000000   0.7500000 0.7500000 0.8571429 0.4139756 0.82315789
#> 20 Site4 Site5 0.9000000   1.0000000 0.9473684 1.0000000 0.2935780 0.31197772
#>       Euclidean a b c    A    B    C
#> 2  0.0013421281 4 4 1  266  726  847
#> 3  0.0015248783 6 2 1  299  693  651
#> 4  0.0009199613 8 0 2  451  541 2062
#> 5  0.0007651863 7 1 2  340  652 2488
#> 8  0.0011446434 4 1 3  117  996  833
#> 9  0.0011632403 5 0 5 1017   96 1496
#> 10 0.0006638422 4 1 5   82 1031 2746
#> 14 0.0010532638 7 0 3  783  167 1730
#> 15 0.0009349723 6 1 3  782  168 2046
#> 20 0.0007119096 9 1 0  784 1729 2044

dissimilarity <- similarity_to_dissimilarity(simil)
dissimilarity
#> Data.frame of dissimilarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of dissimilarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn   Sorensen   Simpson      Bray   Brayturn
#> 2  Site1 Site2 0.5555556   0.3333333 0.38461538 0.2000000 0.7472684 0.73185484
#> 3  Site1 Site3 0.3333333   0.2500000 0.20000000 0.1428571 0.6920700 0.68526316
#> 4  Site1 Site4 0.2000000   0.0000000 0.11111111 0.0000000 0.7426534 0.54536290
#> 5  Site1 Site5 0.3000000   0.2222222 0.17647059 0.1250000 0.8219895 0.65725806
#> 8  Site2 Site3 0.5000000   0.3333333 0.33333333 0.2000000 0.8865730 0.87684211
#> 9  Site2 Site4 0.5000000   0.0000000 0.33333333 0.0000000 0.4390513 0.08625337
#> 10 Site2 Site5 0.6000000   0.3333333 0.42857143 0.2000000 0.9583862 0.92632525
#> 14 Site3 Site4 0.3000000   0.0000000 0.17647059 0.0000000 0.5477909 0.17578947
#> 15 Site3 Site5 0.4000000   0.2500000 0.25000000 0.1428571 0.5860244 0.17684211
#> 20 Site4 Site5 0.1000000   0.0000000 0.05263158 0.0000000 0.7064220 0.68802228
#>    Euclidean a b c    A    B    C
#> 2   744.0853 4 4 1  266  726  847
#> 3   654.7900 6 2 1  299  693  651
#> 4  1086.0023 8 0 2  451  541 2062
#> 5  1305.8714 7 1 2  340  652 2488
#> 8   872.6345 4 1 3  117  996  833
#> 9   858.6676 5 0 5 1017   96 1496
#> 10 1505.3820 4 1 5   82 1031 2746
#> 14  948.4298 7 0 3  783  167 1730
#> 15 1068.5504 6 1 3  782  168 2046
#> 20 1403.6727 9 1 0  784 1729 2044