Convert similarity indices to dissimilarity metrics

This function converts a data.frame of similarity metrics between sites to dissimilarity metrics (= beta diversity).

Usage

similarity_to_dissimilarity(similarity)

Arguments

similarity

the output object from similarity() or a data.frame with the first columns called "Site1" and "Site2", and the other columns being the similarity metrics.

Value

A data.frame with additional class bioregion.pairwise.metric, providing one or several similarity metric(s) between each pair of sites. The two first columns represent each pair of sites, and the other column represent similarity metrics. Columns with names "a", "b", "c", "A", "B" and "C" are not altered.

Note

The behavior of this function changes depending on column names. Columns "Site1" and "Site2" are copied identically. If there are columns called "a", "b", "c", "A", "B", "C", they will also be copied identically.

If a column is called "Euclidean", its distance will be calculated based on the following formula:

\(Euclidean distance = (1 - Euclidean similarity) / Euclidean similarity\)

Otherwise, all other columns will be transformed into dissimilarity with the following formula:

\(dissimilarity = 1 - similarity\)

See also

Author

Boris Leroy (leroy.boris@gmail.com), Maxime Lenormand (maxime.lenormand@inrae.fr) and Pierre Denelle (pierre.denelle@gmail.com)

Examples

comat <- matrix(sample(0:1000, size = 50, replace = TRUE,
prob = 1 / 1:1001), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

simil <- similarity(comat, metric = "all")
simil
#> Data.frame of similarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of similarity metrics:  7 
#> 
#> 
#>    Site1 Site2 Jaccard Jaccardturn  Sorensen   Simpson      Bray  Brayturn
#> 2  Site1 Site2     1.0         1.0 1.0000000 1.0000000 0.3756806 0.4254882
#> 3  Site1 Site3     0.8         0.8 0.8888889 0.8888889 0.2725865 0.3739377
#> 4  Site1 Site4     0.8         0.8 0.8888889 0.8888889 0.3263229 0.3907392
#> 5  Site1 Site5     0.9         1.0 0.9473684 1.0000000 0.4664778 0.5569335
#> 8  Site2 Site3     0.8         0.8 0.8888889 0.8888889 0.4776653 0.5679887
#> 9  Site2 Site4     0.8         0.8 0.8888889 0.8888889 0.3256506 0.4501542
#> 10 Site2 Site5     0.9         1.0 0.9473684 1.0000000 0.3021505 0.3167982
#> 14 Site3 Site4     0.8         0.8 0.8888889 0.8888889 0.2748659 0.4716714
#> 15 Site3 Site5     0.9         1.0 0.9473684 1.0000000 0.2096673 0.2365439
#> 20 Site4 Site5     0.9         1.0 0.9473684 1.0000000 0.3471582 0.5095829
#>      Euclidean a b c   A    B    C
#> 2  0.001729723 9 0 0 414  817  559
#> 3  0.001570830 8 1 1 264  967  442
#> 4  0.001282031 8 1 1 481  750 1236
#> 5  0.001840127 9 0 1 494  737  393
#> 8  0.002960244 8 1 1 401  572  305
#> 9  0.001277002 8 1 1 438  535 1279
#> 10 0.001891457 9 0 1 281  692  606
#> 14 0.001408037 8 1 1 333  373 1384
#> 15 0.001815400 9 0 1 167  539  720
#> 20 0.001506382 9 0 1 452 1265  435

dissimilarity <- similarity_to_dissimilarity(simil)
dissimilarity
#> Data.frame of dissimilarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of dissimilarity metrics:  7 
#> 
#> 
#>    Site1 Site2 Jaccard Jaccardturn   Sorensen   Simpson      Bray  Brayturn
#> 2  Site1 Site2     0.0         0.0 0.00000000 0.0000000 0.6243194 0.5745118
#> 3  Site1 Site3     0.2         0.2 0.11111111 0.1111111 0.7274135 0.6260623
#> 4  Site1 Site4     0.2         0.2 0.11111111 0.1111111 0.6736771 0.6092608
#> 5  Site1 Site5     0.1         0.0 0.05263158 0.0000000 0.5335222 0.4430665
#> 8  Site2 Site3     0.2         0.2 0.11111111 0.1111111 0.5223347 0.4320113
#> 9  Site2 Site4     0.2         0.2 0.11111111 0.1111111 0.6743494 0.5498458
#> 10 Site2 Site5     0.1         0.0 0.05263158 0.0000000 0.6978495 0.6832018
#> 14 Site3 Site4     0.2         0.2 0.11111111 0.1111111 0.7251341 0.5283286
#> 15 Site3 Site5     0.1         0.0 0.05263158 0.0000000 0.7903327 0.7634561
#> 20 Site4 Site5     0.1         0.0 0.05263158 0.0000000 0.6528418 0.4904171
#>    Euclidean a b c   A    B    C
#> 2   577.1274 9 0 0 414  817  559
#> 3   635.6060 8 1 1 264  967  442
#> 4   779.0122 8 1 1 481  750 1236
#> 5   542.4408 9 0 1 494  737  393
#> 8   336.8100 8 1 1 401  572  305
#> 9   782.0844 8 1 1 438  535 1279
#> 10  527.6931 9 0 1 281  692  606
#> 14  709.2087 8 1 1 333  373 1384
#> 15  549.8427 9 0 1 167  539  720
#> 20  662.8424 9 0 1 452 1265  435