Convert similarity metrics to dissimilarity metrics

This function converts a data.frame of similarity metrics between sites to dissimilarity metrics (beta diversity).

Usage

similarity_to_dissimilarity(similarity, include_formula = TRUE)

Arguments

similarity

the output object from similarity() or dissimilarity_to_similarity().

include_formula

a boolean indicating if the metrics based on your own formula(s) should be converted (see Details). This argument is set to TRUE by default.

Value

A data.frame with additional class bioregion.pairwise.metric, providing dissimilarity metric(s) between each pair of sites based on a similarity object.

Note

The behavior of this function changes depending on column names. Columns Site1 and Site2 are copied identically. If there are columns called a, b, c, A, B, C they will also be copied identically. If there are columns based on your own formula (argument formula in similarity()) or not in the original list of similarity metrics (argument metrics in similarity()) and if the argument include_formula is set to FALSE, they will also be copied identically. Otherwise there are going to be converted like they other columns (default behavior).

If a column is called Euclidean, its distance will be calculated based on the following formula:

Euclidean distance = (1 - Euclidean similarity) / Euclidean similarityEuclidean distance = (1 - Euclidean similarity) / Euclidean similarity

Otherwise, all other columns will be transformed into dissimilarity with the following formula:

dissimilarity = 1 - similaritydissimilarity = 1 - similarity

See also

Author

Maxime Lenormand (maxime.lenormand@inrae.fr), Boris Leroy (leroy.boris@gmail.com) and Pierre Denelle (pierre.denelle@gmail.com)

Examples

comat <- matrix(sample(0:1000, size = 50, replace = TRUE,
prob = 1 / 1:1001), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

simil <- similarity(comat, metric = "all")
simil
#> Data.frame of similarity between sites
#>  - Total number of sites:  5 
#>  - Total number of species:  10 
#>  - Number of rows:  10 
#>  - Number of similarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn  Sorensen   Simpson      Bray  Brayturn
#> 2  Site1 Site2 0.7000000   0.7777778 0.8235294 0.8750000 0.2112971 0.2429345
#> 3  Site1 Site3 0.6000000   0.7500000 0.7500000 0.8571429 0.2171875 0.6967419
#> 4  Site1 Site4 0.6000000   0.7500000 0.7500000 0.8571429 0.4324912 0.5236842
#> 5  Site1 Site5 0.6000000   0.7500000 0.7500000 0.8571429 0.2543609 0.4042339
#> 8  Site2 Site3 0.5000000   0.5555556 0.6666667 0.7142857 0.2240543 0.5789474
#> 9  Site2 Site4 0.5000000   0.5555556 0.6666667 0.7142857 0.4693685 0.4914474
#> 10 Site2 Site5 0.5000000   0.5555556 0.6666667 0.7142857 0.4903955 0.6562500
#> 14 Site3 Site4 0.4000000   0.4000000 0.5714286 0.5714286 0.3147473 0.7568922
#> 15 Site3 Site5 0.5555556   0.5555556 0.7142857 0.7142857 0.3134436 0.5463659
#> 20 Site4 Site5 0.7500000   0.7500000 0.8571429 0.8571429 0.6560510 0.8306452
#>       Euclidean a b c   A    B    C
#> 2  0.0008071799 7 2 1 404 1757 1259
#> 3  0.0009817253 6 3 1 278 1883  121
#> 4  0.0010416916 6 3 1 796 1365  724
#> 5  0.0009168120 6 3 1 401 1760  591
#> 8  0.0012629055 5 3 2 231 1432  168
#> 9  0.0012338564 5 3 2 747  916  773
#> 10 0.0014155918 5 3 2 651 1012  341
#> 14 0.0013049242 4 3 3 302   97 1218
#> 15 0.0016812218 5 2 2 218  181  774
#> 20 0.0022075271 6 1 1 824  696  168

dissimilarity <- similarity_to_dissimilarity(simil)
dissimilarity
#> Data.frame of dissimilarity between sites
#>  - Total number of sites:  5 
#>  - Total number of species:  10 
#>  - Number of rows:  10 
#>  - Number of dissimilarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn  Sorensen   Simpson      Bray  Brayturn
#> 2  Site1 Site2 0.3000000   0.2222222 0.1764706 0.1250000 0.7887029 0.7570655
#> 3  Site1 Site3 0.4000000   0.2500000 0.2500000 0.1428571 0.7828125 0.3032581
#> 4  Site1 Site4 0.4000000   0.2500000 0.2500000 0.1428571 0.5675088 0.4763158
#> 5  Site1 Site5 0.4000000   0.2500000 0.2500000 0.1428571 0.7456391 0.5957661
#> 8  Site2 Site3 0.5000000   0.4444444 0.3333333 0.2857143 0.7759457 0.4210526
#> 9  Site2 Site4 0.5000000   0.4444444 0.3333333 0.2857143 0.5306315 0.5085526
#> 10 Site2 Site5 0.5000000   0.4444444 0.3333333 0.2857143 0.5096045 0.3437500
#> 14 Site3 Site4 0.6000000   0.6000000 0.4285714 0.4285714 0.6852527 0.2431078
#> 15 Site3 Site5 0.4444444   0.4444444 0.2857143 0.2857143 0.6865564 0.4536341
#> 20 Site4 Site5 0.2500000   0.2500000 0.1428571 0.1428571 0.3439490 0.1693548
#>    Euclidean a b c   A    B    C
#> 2  1237.8813 7 2 1 404 1757 1259
#> 3  1017.6149 6 3 1 278 1883  121
#> 4   958.9771 6 3 1 796 1365  724
#> 5  1089.7362 6 3 1 401 1760  591
#> 8   790.8249 5 3 2 231 1432  168
#> 9   809.4671 5 3 2 747  916  773
#> 10  705.4183 5 3 2 651 1012  341
#> 14  765.3280 4 3 3 302   97 1218
#> 15  593.8055 5 2 2 218  181  774
#> 20  451.9956 6 1 1 824  696  168