Convert similarity metrics to dissimilarity metrics

This function converts a data.frame of similarity metrics between sites to dissimilarity metrics (beta diversity).

Usage

similarity_to_dissimilarity(similarity, include_formula = TRUE)

Arguments

similarity

the output object from similarity() or dissimilarity_to_similarity().

include_formula

a boolean indicating if the metrics based on your own formula(s) should be converted (see Details). This argument is set to TRUE by default.

Value

A data.frame with additional class bioregion.pairwise.metric, providing dissimilarity metric(s) between each pair of sites based on a similarity object.

Note

The behavior of this function changes depending on column names. Columns Site1 and Site2 are copied identically. If there are columns called a, b, c, A, B, C they will also be copied identically. If there are columns based on your own formula (argument formula in similarity()) or not in the original list of similarity metrics (argument metrics in similarity()) and if the argument include_formula is set to FALSE, they will also be copied identically. Otherwise there are going to be converted like they other columns (default behavior).

If a column is called Euclidean, its distance will be calculated based on the following formula:

Euclidean distance = (1 - Euclidean similarity) / Euclidean similarityEuclidean distance = (1 - Euclidean similarity) / Euclidean similarity

Otherwise, all other columns will be transformed into dissimilarity with the following formula:

dissimilarity = 1 - similaritydissimilarity = 1 - similarity

See also

Author

Maxime Lenormand (maxime.lenormand@inrae.fr), Boris Leroy (leroy.boris@gmail.com) and Pierre Denelle (pierre.denelle@gmail.com)

Examples

comat <- matrix(sample(0:1000, size = 50, replace = TRUE,
prob = 1 / 1:1001), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

simil <- similarity(comat, metric = "all")
simil
#> Data.frame of similarity between sites
#>  - Total number of sites:  5 
#>  - Total number of species:  10 
#>  - Number of rows:  10 
#>  - Number of similarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn  Sorensen   Simpson      Bray  Brayturn
#> 2  Site1 Site2 0.9000000   1.0000000 0.9473684 1.0000000 0.1884945 0.1959288
#> 3  Site1 Site3 0.6000000   1.0000000 0.7500000 1.0000000 0.1182090 0.2519084
#> 4  Site1 Site4 0.8000000   1.0000000 0.8888889 1.0000000 0.1710145 0.4503817
#> 5  Site1 Site5 0.8000000   1.0000000 0.8888889 1.0000000 0.1466855 0.7938931
#> 8  Site2 Site3 0.6666667   1.0000000 0.8000000 1.0000000 0.1465416 0.2948113
#> 9  Site2 Site4 0.8888889   1.0000000 0.9411765 1.0000000 0.3493574 0.8655660
#> 10 Site2 Site5 0.7000000   0.7777778 0.8235294 0.8750000 0.1908985 0.9646226
#> 14 Site3 Site4 0.5555556   0.7142857 0.7142857 0.8333333 0.5718148 0.6599064
#> 15 Site3 Site5 0.4000000   0.5000000 0.5714286 0.6666667 0.2819366 0.5655226
#> 20 Site4 Site5 0.7777778   0.7777778 0.8750000 0.8750000 0.5084868 0.8395945
#>       Euclidean a b c    A   B    C
#> 2  0.0034922217 9 1 0   77 316  347
#> 3  0.0013412188 6 4 0   99 294 1183
#> 4  0.0012277640 8 2 0  177 216 1500
#> 5  0.0006889304 8 2 0  312  81 3549
#> 8  0.0013613621 6 3 0  125 299 1157
#> 9  0.0015606236 8 1 0  367  57 1310
#> 10 0.0006891188 7 2 1  409  15 3452
#> 14 0.0015465556 5 1 3  846 436  831
#> 15 0.0006592137 4 2 4  725 557 3136
#> 20 0.0007726832 7 1 1 1408 269 2453

dissimilarity <- similarity_to_dissimilarity(simil)
dissimilarity
#> Data.frame of dissimilarity between sites
#>  - Total number of sites:  5 
#>  - Total number of species:  10 
#>  - Number of rows:  10 
#>  - Number of dissimilarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn   Sorensen   Simpson      Bray   Brayturn
#> 2  Site1 Site2 0.1000000   0.0000000 0.05263158 0.0000000 0.8115055 0.80407125
#> 3  Site1 Site3 0.4000000   0.0000000 0.25000000 0.0000000 0.8817910 0.74809160
#> 4  Site1 Site4 0.2000000   0.0000000 0.11111111 0.0000000 0.8289855 0.54961832
#> 5  Site1 Site5 0.2000000   0.0000000 0.11111111 0.0000000 0.8533145 0.20610687
#> 8  Site2 Site3 0.3333333   0.0000000 0.20000000 0.0000000 0.8534584 0.70518868
#> 9  Site2 Site4 0.1111111   0.0000000 0.05882353 0.0000000 0.6506426 0.13443396
#> 10 Site2 Site5 0.3000000   0.2222222 0.17647059 0.1250000 0.8091015 0.03537736
#> 14 Site3 Site4 0.4444444   0.2857143 0.28571429 0.1666667 0.4281852 0.34009360
#> 15 Site3 Site5 0.6000000   0.5000000 0.42857143 0.3333333 0.7180634 0.43447738
#> 20 Site4 Site5 0.2222222   0.2222222 0.12500000 0.1250000 0.4915132 0.16040549
#>    Euclidean a b c    A   B    C
#> 2   285.3507 9 1 0   77 316  347
#> 3   744.5905 6 4 0   99 294 1183
#> 4   813.4888 8 2 0  177 216 1500
#> 5  1450.5254 8 2 0  312  81 3549
#> 8   733.5585 6 3 0  125 299 1157
#> 9   639.7695 8 1 0  367  57 1310
#> 10 1450.1286 7 2 1  409  15 3452
#> 14  645.5982 5 1 3  846 436  831
#> 15 1515.9588 4 2 4  725 557 3136
#> 20 1293.1914 7 1 1 1408 269 2453