Convert similarity indices to dissimilarity metrics

This function converts a data.frame of similarity metrics between sites to dissimilarity metrics (= beta diversity).

Usage

similarity_to_dissimilarity(similarity)

Arguments

similarity

the output object from similarity() or a data.frame with the first columns called "Site1" and "Site2", and the other columns being the similarity metrics.

Value

A data.frame with additional class bioregion.pairwise.metric, providing one or several similarity metric(s) between each pair of sites. The two first columns represent each pair of sites, and the other column represent similarity metrics. Columns with names "a", "b", "c", "A", "B" and "C" are not altered.

Note

The behavior of this function changes depending on column names. Columns "Site1" and "Site2" are copied identically. If there are columns called "a", "b", "c", "A", "B", "C", they will also be copied identically.

If a column is called "Euclidean", its distance will be calculated based on the following formula:

\(Euclidean distance = (1 - Euclidean similarity) / Euclidean similarity\)

Otherwise, all other columns will be transformed into dissimilarity with the following formula:

\(dissimilarity = 1 - similarity\)

See also

Author

Boris Leroy (leroy.boris@gmail.com), Maxime Lenormand (maxime.lenormand@inrae.fr) and Pierre Denelle (pierre.denelle@gmail.com)

Examples

# \donttest{
comat <- matrix(sample(0:1000, size = 50, replace = TRUE,
prob = 1 / 1:1001), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

simil <- similarity(comat, metric = "all")
simil
#> Data.frame of similarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of similarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn  Sorensen   Simpson      Bray  Brayturn
#> 2  Site1 Site2 0.7777778        1.00 0.8750000 1.0000000 0.1116202 0.1232751
#> 3  Site1 Site3 0.9000000        1.00 0.9473684 1.0000000 0.1546311 0.1812328
#> 4  Site1 Site4 1.0000000        1.00 1.0000000 1.0000000 0.3472741 0.4277829
#> 5  Site1 Site5 0.8000000        0.80 0.8888889 0.8888889 0.3500838 0.3845446
#> 8  Site2 Site3 0.7000000        1.00 0.8235294 1.0000000 0.5088288 0.5372907
#> 9  Site2 Site4 0.7777778        1.00 0.8750000 1.0000000 0.3194492 0.3531202
#> 10 Site2 Site5 0.6000000        0.75 0.7500000 0.8571429 0.1812620 0.1821676
#> 14 Site3 Site4 0.9000000        1.00 0.9473684 1.0000000 0.3243775 0.3388090
#> 15 Site3 Site5 0.9000000        1.00 0.9473684 1.0000000 0.2049240 0.2175250
#> 20 Site4 Site5 0.8000000        0.80 0.8888889 0.8888889 0.6929461 0.7701768
#>       Euclidean a b c    A    B    C
#> 2  0.0009721790 7 2 0  134  953 1180
#> 3  0.0009464605 9 0 1  197  890 1264
#> 4  0.0011643893 9 0 0  465  622 1126
#> 5  0.0013089937 8 1 1  418  669  883
#> 8  0.0015210804 7 0 3  706  608  755
#> 9  0.0010997616 7 0 2  464  850 1127
#> 10 0.0010631711 6 1 3  237 1077 1064
#> 14 0.0011208288 9 1 0  495  966 1096
#> 15 0.0010044606 9 1 0  283 1178 1018
#> 20 0.0026996297 8 1 1 1002  589  299

dissimilarity <- similarity_to_dissimilarity(simil)
dissimilarity
#> Data.frame of dissimilarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of dissimilarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn   Sorensen   Simpson      Bray  Brayturn
#> 2  Site1 Site2 0.2222222        0.00 0.12500000 0.0000000 0.8883798 0.8767249
#> 3  Site1 Site3 0.1000000        0.00 0.05263158 0.0000000 0.8453689 0.8187672
#> 4  Site1 Site4 0.0000000        0.00 0.00000000 0.0000000 0.6527259 0.5722171
#> 5  Site1 Site5 0.2000000        0.20 0.11111111 0.1111111 0.6499162 0.6154554
#> 8  Site2 Site3 0.3000000        0.00 0.17647059 0.0000000 0.4911712 0.4627093
#> 9  Site2 Site4 0.2222222        0.00 0.12500000 0.0000000 0.6805508 0.6468798
#> 10 Site2 Site5 0.4000000        0.25 0.25000000 0.1428571 0.8187380 0.8178324
#> 14 Site3 Site4 0.1000000        0.00 0.05263158 0.0000000 0.6756225 0.6611910
#> 15 Site3 Site5 0.1000000        0.00 0.05263158 0.0000000 0.7950760 0.7824750
#> 20 Site4 Site5 0.2000000        0.20 0.11111111 0.1111111 0.3070539 0.2298232
#>    Euclidean a b c    A    B    C
#> 2  1027.6171 7 2 0  134  953 1180
#> 3  1055.5681 9 0 1  197  890 1264
#> 4   857.8193 9 0 0  465  622 1126
#> 5   762.9456 8 1 1  418  669  883
#> 8   656.4275 7 0 3  706  608  755
#> 9   908.2879 7 0 2  464  850 1127
#> 10  939.5824 6 1 3  237 1077 1064
#> 14  891.1969 9 1 0  495  966 1096
#> 15  994.5592 9 1 0  283 1178 1018
#> 20  369.4212 8 1 1 1002  589  299
# }