Convert similarity indices to dissimilarity metrics

This function converts a data.frame of similarity metrics between sites to dissimilarity metrics (= beta diversity).

Usage

similarity_to_dissimilarity(similarity)

Arguments

similarity

the output object from similarity() or a data.frame with the first columns called "Site1" and "Site2", and the other columns being the similarity metrics.

Value

A data.frame with additional class bioregion.pairwise.metric, providing one or several similarity metric(s) between each pair of sites. The two first columns represent each pair of sites, and the other column represent similarity metrics. Columns with names "a", "b", "c", "A", "B" and "C" are not altered.

Note

The behavior of this function changes depending on column names. Columns "Site1" and "Site2" are copied identically. If there are columns called "a", "b", "c", "A", "B", "C", they will also be copied identically.

If a column is called "Euclidean", its distance will be calculated based on the following formula:

\(Euclidean distance = (1 - Euclidean similarity) / Euclidean similarity\)

Otherwise, all other columns will be transformed into dissimilarity with the following formula:

\(dissimilarity = 1 - similarity\)

See also

Author

Boris Leroy (leroy.boris@gmail.com), Maxime Lenormand (maxime.lenormand@inrae.fr) and Pierre Denelle (pierre.denelle@gmail.com)

Examples

# \donttest{
comat <- matrix(sample(0:1000, size = 50, replace = TRUE,
prob = 1 / 1:1001), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

simil <- similarity(comat, metric = "all")
simil
#> Data.frame of similarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of similarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn  Sorensen   Simpson       Bray   Brayturn
#> 2  Site1 Site2 0.7000000   0.7777778 0.8235294 0.8750000 0.43721357 0.51904244
#> 3  Site1 Site3 0.7000000   0.7777778 0.8235294 0.8750000 0.19498607 0.80924855
#> 4  Site1 Site4 0.7000000   0.7777778 0.8235294 0.8750000 0.07150675 0.07759303
#> 5  Site1 Site5 0.8000000   0.8000000 0.8888889 0.8888889 0.08709677 0.13567839
#> 8  Site2 Site3 0.6000000   0.6000000 0.7500000 0.7500000 0.13186813 0.41618497
#> 9  Site2 Site4 0.7777778   0.7777778 0.8750000 0.8750000 0.28368794 0.36996736
#> 10 Site2 Site5 0.7000000   0.7777778 0.8235294 0.8750000 0.04485488 0.05695142
#> 14 Site3 Site4 0.6000000   0.6000000 0.7500000 0.7500000 0.09448819 0.45086705
#> 15 Site3 Site5 0.8888889   1.0000000 0.9411765 1.0000000 0.21818182 0.48554913
#> 20 Site4 Site5 0.7000000   0.7777778 0.8235294 0.8750000 0.06843373 0.11892797
#>       Euclidean a b c   A    B    C
#> 2  0.0020743265 7 2 1 477  786  442
#> 3  0.0015379020 7 2 1 140 1123   33
#> 4  0.0008108403 7 2 1  98 1165 1380
#> 5  0.0011854783 8 1 1  81 1182  516
#> 8  0.0022275480 6 2 2  72  847  101
#> 9  0.0009369168 7 1 1 340  579 1138
#> 10 0.0014338840 7 1 2  34  885  563
#> 14 0.0010057835 6 2 2  78   95 1400
#> 15 0.0019570072 8 0 1  84   89  513
#> 20 0.0008948127 7 1 2  71 1407  526

dissimilarity <- similarity_to_dissimilarity(simil)
dissimilarity
#> Data.frame of dissimilarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of dissimilarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn   Sorensen   Simpson      Bray  Brayturn
#> 2  Site1 Site2 0.3000000   0.2222222 0.17647059 0.1250000 0.5627864 0.4809576
#> 3  Site1 Site3 0.3000000   0.2222222 0.17647059 0.1250000 0.8050139 0.1907514
#> 4  Site1 Site4 0.3000000   0.2222222 0.17647059 0.1250000 0.9284933 0.9224070
#> 5  Site1 Site5 0.2000000   0.2000000 0.11111111 0.1111111 0.9129032 0.8643216
#> 8  Site2 Site3 0.4000000   0.4000000 0.25000000 0.2500000 0.8681319 0.5838150
#> 9  Site2 Site4 0.2222222   0.2222222 0.12500000 0.1250000 0.7163121 0.6300326
#> 10 Site2 Site5 0.3000000   0.2222222 0.17647059 0.1250000 0.9551451 0.9430486
#> 14 Site3 Site4 0.4000000   0.4000000 0.25000000 0.2500000 0.9055118 0.5491329
#> 15 Site3 Site5 0.1111111   0.0000000 0.05882353 0.0000000 0.7818182 0.5144509
#> 20 Site4 Site5 0.3000000   0.2222222 0.17647059 0.1250000 0.9315663 0.8810720
#>    Euclidean a b c   A    B    C
#> 2   481.0842 7 2 1 477  786  442
#> 3   649.2365 7 2 1 140 1123   33
#> 4  1232.2885 7 2 1  98 1165 1380
#> 5   842.5414 8 1 1  81 1182  516
#> 8   447.9241 6 2 2  72  847  101
#> 9  1066.3306 7 1 1 340  579 1138
#> 10  696.4065 7 1 2  34  885  563
#> 14  993.2497 6 2 2  78   95 1400
#> 15  509.9843 8 0 1  84   89  513
#> 20 1116.5523 7 1 2  71 1407  526
# }