Convert similarity indices to dissimilarity metrics

This function converts a data.frame of similarity metrics between sites to dissimilarity metrics (= beta diversity).

Usage

similarity_to_dissimilarity(similarity)

Arguments

similarity

the output object from similarity() or a data.frame with the first columns called "Site1" and "Site2", and the other columns being the similarity metrics.

Value

A data.frame with additional class bioregion.pairwise.metric, providing one or several similarity metric(s) between each pair of sites. The two first columns represent each pair of sites, and the other column represent similarity metrics. Columns with names "a", "b", "c", "A", "B" and "C" are not altered.

Note

The behavior of this function changes depending on column names. Columns "Site1" and "Site2" are copied identically. If there are columns called "a", "b", "c", "A", "B", "C", they will also be copied identically.

If a column is called "Euclidean", its distance will be calculated based on the following formula:

\(Euclidean distance = (1 - Euclidean similarity) / Euclidean similarity\)

Otherwise, all other columns will be transformed into dissimilarity with the following formula:

\(dissimilarity = 1 - similarity\)

See also

Author

Boris Leroy (leroy.boris@gmail.com), Maxime Lenormand (maxime.lenormand@inrae.fr) and Pierre Denelle (pierre.denelle@gmail.com)

Examples

comat <- matrix(sample(0:1000, size = 50, replace = TRUE,
prob = 1 / 1:1001), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

simil <- similarity(comat, metric = "all")
simil
#> Data.frame of similarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of similarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn  Sorensen Simpson       Bray   Brayturn
#> 2  Site1 Site2 0.8000000   1.0000000 0.8888889   1.000 0.22833459 0.41336971
#> 3  Site1 Site3 1.0000000   1.0000000 1.0000000   1.000 0.27400768 0.35577722
#> 4  Site1 Site4 0.8000000   1.0000000 0.8888889   1.000 0.40708945 0.43491124
#> 5  Site1 Site5 0.9000000   1.0000000 0.9473684   1.000 0.53529109 0.54086413
#> 8  Site2 Site3 0.8000000   1.0000000 0.8888889   1.000 0.27376033 0.36152797
#> 9  Site2 Site4 0.7777778   0.7777778 0.8750000   0.875 0.17746595 0.29331514
#> 10 Site2 Site5 0.7000000   0.7777778 0.8235294   0.875 0.28433556 0.52251023
#> 14 Site3 Site4 0.8000000   1.0000000 0.8888889   1.000 0.06498445 0.07813799
#> 15 Site3 Site5 0.9000000   1.0000000 0.9473684   1.000 0.32996207 0.43391521
#> 20 Site4 Site5 0.7000000   0.7777778 0.8235294   0.875 0.12873185 0.13905325
#>       Euclidean  a b c    A    B    C
#> 2  0.0011547676  8 2 0  303 1618  430
#> 3  0.0010999474 10 0 0  428 1493  775
#> 4  0.0010570233  8 2 0  735 1186  955
#> 5  0.0012120267  9 1 0 1039  882  922
#> 8  0.0012937586  8 0 2  265  468  938
#> 9  0.0010743838  7 1 1  215  518 1475
#> 10 0.0010010753  7 1 2  383  350 1578
#> 14 0.0008635617  8 2 0   94 1109 1596
#> 15 0.0009749644  9 1 0  522  681 1439
#> 20 0.0007595423  7 1 2  235 1455 1726

dissimilarity <- similarity_to_dissimilarity(simil)
dissimilarity
#> Data.frame of dissimilarity between sites
#>  - Total number of sites:  5 
#>  - Number of rows:  10 
#>  - Number of dissimilarity metrics:  7 
#> 
#> 
#>    Site1 Site2   Jaccard Jaccardturn   Sorensen Simpson      Bray  Brayturn
#> 2  Site1 Site2 0.2000000   0.0000000 0.11111111   0.000 0.7716654 0.5866303
#> 3  Site1 Site3 0.0000000   0.0000000 0.00000000   0.000 0.7259923 0.6442228
#> 4  Site1 Site4 0.2000000   0.0000000 0.11111111   0.000 0.5929106 0.5650888
#> 5  Site1 Site5 0.1000000   0.0000000 0.05263158   0.000 0.4647089 0.4591359
#> 8  Site2 Site3 0.2000000   0.0000000 0.11111111   0.000 0.7262397 0.6384720
#> 9  Site2 Site4 0.2222222   0.2222222 0.12500000   0.125 0.8225340 0.7066849
#> 10 Site2 Site5 0.3000000   0.2222222 0.17647059   0.125 0.7156644 0.4774898
#> 14 Site3 Site4 0.2000000   0.0000000 0.11111111   0.000 0.9350156 0.9218620
#> 15 Site3 Site5 0.1000000   0.0000000 0.05263158   0.000 0.6700379 0.5660848
#> 20 Site4 Site5 0.3000000   0.2222222 0.17647059   0.125 0.8712681 0.8609467
#>    Euclidean  a b c    A    B    C
#> 2   864.9751  8 2 0  303 1618  430
#> 3   908.1344 10 0 0  428 1493  775
#> 4   945.0529  8 2 0  735 1186  955
#> 5   824.0643  9 1 0 1039  882  922
#> 8   771.9417  8 0 2  265  468  938
#> 9   929.7661  7 1 1  215  518 1475
#> 10  997.9258  7 1 2  383  350 1578
#> 14 1156.9948  8 2 0   94 1109 1596
#> 15 1024.6785  9 1 0  522  681 1439
#> 20 1315.5824  7 1 2  235 1455 1726