Calculate contribution metrics of sites and species https://github.com/Farewe/biogeonetworks/blob/master/R/networkmetrics.R

This function calculates metrics that assess the contribution of a given species or site to its bioregion.

Usage

site_species_metrics(
  cluster_object,
  comat,
  indices = c("rho", "Cz"),
  bipartite_link = NULL
)

Arguments

cluster_object

a bioregion.clusters object or a data.frame or a list of data.frame containing multiple partitions. At least two partitions are required. If a list of data.frame is provided, they should all have the same number of rows (i.e., same items in the clustering for all partitions).

comat

a co-occurrence matrix with sites as rows and species as columns.

indices

a character specifying the contribution metric to compute. Available options are rho and Cz.

bipartite_link

NULL by default. Needed for Cz indices. A data.frame where each row represents the interaction between two nodes and an optional third column indicating the weight of the interaction.

Value

A list of data.frames if multiples indices are selected or a single data.frame with three columns if one index is selected. Each data.frame has three columns: the species, the bioregion, and the desired summary statistics.

Details

The \(\rho\) metric is derived from Lenormand2019bioregion. Its formula is the following: \(\rho_{ij} = (n_ij - ((n_i n_j)/n))/(sqrt(((n - n_j)/(n-1)) (1-(n_j/n)) ((n_i n_j)/n)))\)

with \(n\) the number of sites, \(n_i\) the number of sites in which species \(i\) is present, \(n_j\) the number of sites belonging to the bioregion \(j\), \(n_ij\) the number of occurrences of species \(i\) in sites belonging to the bioregion \(j\).

Cz metrics are derived from Guimera2005bioregion. Their respective formula are: \(C_i = 1 - \sum_{s=1}^{N_M}{{(\frac{k_is}{k_i}})^2}\)

where \(k_{is}\) is the number of links of node (species or site) \(i\) to nodes in bioregion \(s\), and \(k_i\) is the total degree of node \(i\). The participation coefficient of a node is therefore close to 1 if its links are uniformly distributed among all the bioregions and 0 if all its links are within its own bioregion.

And: \(z_i = \frac{k_i - \overline{k_{si}}}{\sigma_{k_{si}}}\)

where \(k_i\) is the number of links of node (species or site) \(i\) to other nodes in its bioregion \(s_i\), \(\overline{k_{si}}\) is the average of \(k\) over all the nodes in \(s_i\), and \(\sigma_{k_{si}}\) is the standard deviation of \(k\) in \(s_i\). The within-bioregion degree z-score measures how well-connected node \(i\) is to other nodes in the bioregion.

See also

Author

Pierre Denelle (pierre.denelle@gmail.com), Boris Leroy (leroy.boris@gmail.com) and Maxime Lenormand (maxime.lenormand@inrae.fr)

Examples

comat <- matrix(sample(1000, 50), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

comat <- matrix(sample(0:1000, size = 500, replace = TRUE, prob = 1/1:1001),
                20, 25)
rownames(comat) <- paste0("Site",1:20)
colnames(comat) <- paste0("Species",1:25)

dissim <- dissimilarity(comat, metric = "Simpson")
clust1 <- nhclu_kmeans(dissim, n_clust = 3, index = "Simpson")

net <- similarity(comat, metric = "Simpson")
com <- netclu_greedy(net)

site_species_metrics(cluster_object = clust1, comat = comat,
indices = "rho")
#>    Bioregion   Species         rho
#> 1          3  Species1  533.311410
#> 2          3  Species2  606.021469
#> 3          3  Species3  928.509797
#> 4          3  Species4  697.221218
#> 5          3  Species5  531.843892
#> 6          3  Species6  504.904854
#> 7          3  Species7  528.950878
#> 8          3  Species8    5.579258
#> 9          3  Species9  430.300522
#> 10         3 Species10 1429.043980
#> 11         3 Species11  533.266831
#> 12         3 Species12  132.917613
#> 13         3 Species13  292.765802
#> 14         3 Species14  465.978201
#> 15         3 Species15   62.578631
#> 16         3 Species16  400.021386
#> 17         3 Species17  455.809555
#> 18         3 Species18  737.937946
#> 19         3 Species19  424.751439
#> 20         3 Species20  303.577264
#> 21         3 Species21  256.204963
#> 22         3 Species22  232.247288
#> 23         3 Species23  589.218926
#> 24         3 Species24  481.475394
#> 25         3 Species25  226.255356
#> 26         2  Species1  230.718721
#> 27         2  Species2  151.613266
#> 28         2  Species3   82.448986
#> 29         2  Species4  889.562238
#> 30         2  Species5  338.304486
#> 31         2  Species6  483.325088
#> 32         2  Species7  439.548993
#> 33         2  Species8  733.508309
#> 34         2  Species9  237.315763
#> 35         2 Species10  232.293277
#> 36         2 Species11  688.024577
#> 37         2 Species12  439.448602
#> 38         2 Species13  933.303343
#> 39         2 Species14  490.128735
#> 40         2 Species15  497.071454
#> 41         2 Species16  657.674466
#> 42         2 Species17 1121.220325
#> 43         2 Species18  445.165600
#> 44         2 Species19  756.463262
#> 45         2 Species20 1194.945750
#> 46         2 Species21  489.489700
#> 47         2 Species22  189.105377
#> 48         2 Species23  678.450338
#> 49         2 Species24  359.395579
#> 50         2 Species25  784.104233
#> 51         1  Species1   96.161876
#> 52         1  Species2  949.506743
#> 53         1  Species3  350.407043
#> 54         1  Species4   52.548948
#> 55         1  Species5  563.508769
#> 56         1  Species6  570.127655
#> 57         1  Species7  386.979129
#> 58         1  Species8 1988.229694
#> 59         1  Species9  370.782321
#> 60         1 Species10  464.569155
#> 61         1 Species11  329.667346
#> 62         1 Species12  516.621391
#> 63         1 Species13  908.588269
#> 64         1 Species14  337.501667
#> 65         1 Species15  957.186879
#> 66         1 Species16  513.865753
#> 67         1 Species17  497.391155
#> 68         1 Species18  102.766186
#> 69         1 Species19 1763.084886
#> 70         1 Species20  287.159259
#> 71         1 Species21  329.040356
#> 72         1 Species22  582.276251
#> 73         1 Species23  397.394011
#> 74         1 Species24 1120.819139
#> 75         1 Species25  179.078098

site_species_metrics(cluster_object = com, comat = comat, indices = "rho")
#>    Bioregion   Species           rho
#> 1          1  Species1  6763.5199169
#> 2          1  Species2 10098.4558196
#> 3          1  Species3  8739.3684109
#> 4          1  Species4 12883.0141487
#> 5          1  Species5 10153.5819733
#> 6          1  Species6 12213.3664700
#> 7          1  Species7 10636.8849072
#> 8          1  Species8 21300.1174097
#> 9          1  Species9  8124.2158318
#> 10         1 Species10 13296.7065194
#> 11         1 Species11 12087.2479449
#> 12         1 Species12  7512.2966972
#> 13         1 Species13 16715.5329027
#> 14         1 Species14 10129.0000000
#> 15         1 Species15 11860.5640254
#> 16         1 Species16 12175.9073675
#> 17         1 Species17 16204.0000000
#> 18         1 Species18  9405.8156730
#> 19         1 Species19 22186.4313714
#> 20         1 Species20 13946.6130297
#> 21         1 Species21  8153.2057660
#> 22         1 Species22  6801.8958305
#> 23         1 Species23 11911.0354868
#> 24         1 Species24 15346.9267454
#> 25         1 Species25  8489.5240149
#> 26         2  Species1    -0.8885233
#> 27         2  Species2   746.8756989
#> 28         2  Species3   445.6763177
#> 29         2  Species4     2.2710999
#> 30         2  Species5   247.7661839
#> 31         2  Species6     0.1669245
#> 32         2  Species7    -0.9459053
#> 33         2  Species8    -0.8885233
#> 34         2  Species9     3.5054138
#> 35         2 Species10   777.9473684
#> 36         2 Species11    20.1978603
#> 37         2 Species12   232.4969345
#> 38         2 Species13     1.1896238
#> 39         2 Species14     4.8177304
#> 40         2 Species15    -1.0000000
#> 41         2 Species16    33.6339080
#> 42         2 Species17    17.4355958
#> 43         2 Species18   161.1052632
#> 44         2 Species19   191.5736446
#> 45         2 Species20    18.0666408
#> 46         2 Species21    63.2105263
#> 47         2 Species22   240.1958509
#> 48         2 Species23   265.3157895
#> 49         2 Species24    -0.9459053
#> 50         2 Species25   195.8553291

# Cz indices
net_bip <- mat_to_net(comat, weight = TRUE)
clust_bip <- netclu_greedy(net_bip, bipartite = TRUE)
site_species_metrics(cluster_object = clust_bip, comat = comat, 
bipartite_link = net_bip, indices = "Cz")
#> Warning: Column 'Node1' has been renamed 'Sites'. If this column does not
#>             correspond to sites, rename it before running the function.
#> Warning: Column 'Node2' has been renamed 'Species'. If this column does not
#>             correspond to species, rename it before running the function.
#>         Node Bioregion Category         C          z
#> 1      Site1         5     site 0.7922438 -0.3333333
#> 2      Site2         4     site 0.7777778  0.4743416
#> 3      Site3         1     site 0.7809917  0.0000000
#> 4      Site4         4     site 0.7901701  0.4743416
#> 5      Site5         1     site 0.7939509  1.4142136
#> 6      Site6         3     site 0.7800000        NaN
#> 7      Site7         1     site 0.7800454  0.0000000
#> 8      Site8         4     site 0.7777778  0.4743416
#> 9      Site9         5     site 0.7851240  1.1666667
#> 10    Site10         4     site 0.7664399  0.4743416
#> 11    Site11         2     site 0.7664399  1.3814374
#> 12    Site12         3     site 0.7892562        NaN
#> 13    Site13         2     site 0.7847222  1.3814374
#> 14    Site14         2     site 0.7755102  1.3814374
#> 15    Site15         1     site 0.7916667  1.4142136
#> 16    Site16         2     site 0.7845805  0.7207500
#> 17    Site17         5     site 0.7845805  1.1666667
#> 18    Site18         4     site 0.7716049 -1.8973666
#> 19    Site19         5     site 0.7800000  1.1666667
#> 20    Site20         3     site 0.7845805        NaN
#> 21  Species2         3  species 0.7958478        NaN
#> 22  Species3         2  species 0.7962963 -0.6006250
#> 23  Species5         5  species 0.7839506 -0.3333333
#> 24  Species6         5  species 0.7889273 -0.3333333
#> 25  Species7         5  species 0.7750865 -0.3333333
#> 26  Species9         3  species 0.7958478        NaN
#> 27 Species10         5  species 0.7977839 -0.3333333
#> 28 Species13         1  species 0.7839506  0.0000000
#> 29 Species14         2  species 0.7734375 -0.6006250
#> 30 Species15         4  species 0.7867036  0.4743416
#> 31 Species16         2  species 0.7901235 -0.6006250
#> 32 Species17         2  species 0.7890625 -1.2613124
#> 33 Species18         1  species 0.7922438 -1.4142136
#> 34 Species19         3  species 0.7958478        NaN
#> 35 Species20         4  species 0.7733333  0.4743416
#> 36 Species21         1  species 0.7977839  0.0000000
#> 37 Species22         1  species 0.7901235 -1.4142136
#> 38 Species23         2  species 0.7922438 -0.6006250
#> 39 Species25         4  species 0.7839506  0.4743416
#> 40  Species1         2  species 0.7822222 -0.6006250
#> 41  Species4         2  species 0.7901235 -0.6006250
#> 42  Species8         4  species 0.7644444  0.4743416
#> 43 Species11         1  species 0.7820069  0.0000000
#> 44 Species12         4  species 0.7644444 -1.8973666
#> 45 Species24         5  species 0.7889273 -1.8333333