Calculate contribution metrics of sites and species

This function calculates metrics to assess the contribution of a given species or site to its bioregion.

Usage

site_species_metrics(
  cluster_object,
  comat,
  indices = c("rho"),
  net = NULL,
  site_col = 1,
  species_col = 2
)

Arguments

cluster_object

A bioregion.clusters object.

comat

A co-occurrence matrix with sites as rows and species as columns.

indices

A character specifying the contribution metric to compute. Available options are rho, affinity, fidelity, indicator_value and Cz.

net

NULL by default. Required for Cz indices. A data.frame where each row represents an interaction between two nodes and an optional third column indicating the interaction's weight.

site_col

A number indicating the position of the column containing the sites in net. 1 by default.

species_col

A number indicating the position of the column containing the species in net. 2 by default.

Value

A data.frame with columns Bioregion, Species, and the desired summary statistics, or a list of data.frames if Cz and other indices are selected.

Details

The \(\rho\) metric is derived from Lenormand et al. (2019) with the following formula:

\(\rho_{ij} = \frac{n_{ij} - \frac{n_i n_j}{n}}{\sqrt{\left(\frac{n - n_j}{ n-1}\right) \left(1-\frac{n_j}{n}\right) \frac{n_i n_j}{n}}}\)

where \(n\) is the number of sites, \(n_i\) is the number of sites in which species \(i\) is present, \(n_j\) is the number of sites in bioregion \(j\), and \(n_{ij}\) is the number of occurrences of species \(i\) in sites of bioregion \(j\).

Affinity \(A\), fidelity \(F\), and individual contributions \(IndVal\) describe how species are linked to their bioregions. These metrics are described in Bernardo-Madrid et al. (2019):

Affinity of species to their region: \(A_i = \frac{R_i}{Z}\), where \(R_i\) is the occurrence/range size of species \(i\) in its associated bioregion, and \(Z\) is the total size (number of sites) of the bioregion. High affinity indicates that the species occupies most sites in its bioregion.

Fidelity of species to their region: \(F_i = \frac{R_i}{D_i}\), where \(R_i\) is the occurrence/range size of species \(i\) in its bioregion, and \(D_i\) is its total range size. High fidelity indicates that the species is not present in other regions.

Indicator Value of species: \(IndVal = F_i \cdot A_i\).

Cz metrics are derived from Guimerà & Amaral (2005):

Participation coefficient: \(C_i = 1 - \sum_{s=1}^{N_M}{\left(\frac{k_{is}}{k_i}\right)^2}\), where \(k_{is}\) is the number of links of node \(i\) to nodes in bioregion \(s\), and \(k_i\) is the total degree of node \(i\). A high value means links are uniformly distributed; a low value means links are within the node's bioregion.

Within-bioregion degree z-score: \(z_i = \frac{k_i - \overline{k_{si}}}{\sigma_{k_{si}}}\), where \(k_i\) is the number of links of node \(i\) to nodes in its bioregion \(s_i\), \(\overline{k_{si}}\) is the average degree of nodes in \(s_i\), and \(\sigma_{k_{si}}\) is the standard deviation of degrees in \(s_i\).

References

Bernardo-Madrid R, Calatayud J, González‐Suárez M, Rosvall M, Lucas P, Antonelli A & Revilla E (2019) Human activity is altering the world’s zoogeographical regions. Ecology Letters 22, 1297–1305.

Guimerà R & Amaral LAN (2005) Functional cartography of complex metabolic networks. Nature 433, 895–900.

Lenormand M, Papuga G, Argagnon O, Soubeyrand M, Alleaume S & Luque S (2019) Biogeographical network analysis of plant species distribution in the Mediterranean region. Ecology and Evolution 9, 237–250.

See also

Author

Pierre Denelle (pierre.denelle@gmail.com)
Boris Leroy (leroy.boris@gmail.com)
Maxime Lenormand (maxime.lenormand@inrae.fr)

Examples

comat <- matrix(sample(0:1000, size = 500, replace = TRUE, prob = 1/1:1001),
                20, 25)
rownames(comat) <- paste0("Site",1:20)
colnames(comat) <- paste0("Species",1:25)

dissim <- dissimilarity(comat, metric = "Simpson")
clust1 <- nhclu_kmeans(dissim, n_clust = 3, index = "Simpson")

net <- similarity(comat, metric = "Simpson")
com <- netclu_greedy(net)

site_species_metrics(cluster_object = clust1, comat = comat,
indices = "rho")
#>    Bioregion   Species          rho
#> 1          3  Species1  706.5546173
#> 2          3  Species2  648.2651373
#> 3          3  Species3  365.2772919
#> 4          3  Species4  696.4503570
#> 5          3  Species5  793.5617688
#> 6          3  Species6  891.7795022
#> 7          3  Species7 1152.2304999
#> 8          3  Species8 1047.0133698
#> 9          3  Species9  241.2078538
#> 10         3 Species10  708.9451443
#> 11         3 Species11   83.8617155
#> 12         3 Species12  254.9128455
#> 13         3 Species13  921.2422848
#> 14         3 Species14 1578.0899848
#> 15         3 Species15  637.2210610
#> 16         3 Species16 1547.0317861
#> 17         3 Species17  921.9071266
#> 18         3 Species18 1307.0824353
#> 19         3 Species19  636.6579406
#> 20         3 Species20   87.9611285
#> 21         3 Species21  297.7702243
#> 22         3 Species22  421.7858856
#> 23         3 Species23 1041.8443515
#> 24         3 Species24  296.6507748
#> 25         3 Species25  124.3156824
#> 26         2  Species1   78.1024455
#> 27         2  Species2  226.2421277
#> 28         2  Species3  776.4616190
#> 29         2  Species4   66.7874779
#> 30         2  Species5  773.4768489
#> 31         2  Species6  143.2485353
#> 32         2  Species7  938.7183800
#> 33         2  Species8   31.1718168
#> 34         2  Species9  448.1746391
#> 35         2 Species10  626.1339682
#> 36         2 Species11 1116.7575998
#> 37         2 Species12  591.4085612
#> 38         2 Species13   34.5110638
#> 39         2 Species14  398.1902382
#> 40         2 Species15  768.3436252
#> 41         2 Species16  452.5979193
#> 42         2 Species17  194.5789833
#> 43         2 Species18  588.3348719
#> 44         2 Species19 1135.2870670
#> 45         2 Species20  584.0317069
#> 46         2 Species21 1270.1013836
#> 47         2 Species22   71.1701793
#> 48         2 Species23 1238.9380286
#> 49         2 Species24  944.2680946
#> 50         2 Species25  549.6343294
#> 51         1  Species1  150.5044893
#> 52         1  Species2  446.9094918
#> 53         1  Species3  221.4320663
#> 54         1  Species4  241.0707135
#> 55         1  Species5  408.3141553
#> 56         1  Species6  206.4636099
#> 57         1  Species7  113.4431821
#> 58         1  Species8 1185.0393261
#> 59         1  Species9   79.9013362
#> 60         1 Species10   98.1478477
#> 61         1 Species11  420.4117294
#> 62         1 Species12  212.9126336
#> 63         1 Species13  763.3936075
#> 64         1 Species14 1046.8057126
#> 65         1 Species15  181.4196486
#> 66         1 Species16  613.1298394
#> 67         1 Species17  543.7041452
#> 68         1 Species18  182.0237541
#> 69         1 Species19  741.7292897
#> 70         1 Species20   14.9716508
#> 71         1 Species21  173.3106438
#> 72         1 Species22    0.1677741
#> 73         1 Species23  475.4105781
#> 74         1 Species24 1399.9281452
#> 75         1 Species25  224.8260562

# Contribution metrics
site_species_metrics(cluster_object = com, comat = comat,
indices = c("rho", "affinity", "fidelity", "indicator_value"))
#>    Bioregion   Species           rho  affinity    fidelity       indval
#> 1          1  Species1  7076.4619317  77.63158  86.7647059 6.735681e+03
#> 2          1  Species2 10266.7190583 115.52632 121.9444444 1.408779e+04
#> 3          1  Species3 10107.0272583 119.94737 113.9500000 1.366800e+04
#> 4          1  Species4  7434.0000000  79.05263  93.8750000 7.421066e+03
#> 5          1  Species5 15103.3294416 169.52632 178.9444444 3.033579e+04
#> 6          1  Species6  7226.3662012  83.84211  83.8421053 7.029499e+03
#> 7          1  Species7 13796.8328192 159.21053 159.2105263 2.534799e+04
#> 8          1  Species8 17512.8843998 207.10526 196.7500000 4.074796e+04
#> 9          1  Species9  5834.6795315  64.15789  71.7058824 4.600498e+03
#> 10         1 Species10 10693.3070189 109.73684 139.0000000 1.525342e+04
#> 11         1 Species11 12699.0000000 134.47368 159.6875000 2.147377e+04
#> 12         1 Species12  8293.9907694  90.84211 101.5294118 9.223146e+03
#> 13         1 Species13 12351.5136959 142.63158 142.6315789 2.034377e+04
#> 14         1 Species14 23474.0000000 247.89474 294.3750000 7.297401e+04
#> 15         1 Species15 12333.1604371 142.42105 142.4210526 2.028376e+04
#> 16         1 Species16 20248.0032557 233.21053 233.2105263 5.438715e+04
#> 17         1 Species17 10596.7022229 119.21053 125.8333333 1.500066e+04
#> 18         1 Species18 16224.9057081 176.89474 197.7058824 3.497313e+04
#> 19         1 Species19 19626.9886041 200.78947 254.3333333 5.106746e+04
#> 20         1 Species20  5402.9661189  59.47368  66.4705882 3.953251e+03
#> 21         1 Species21 13678.0861844 140.15789 177.5333333 2.488270e+04
#> 22         1 Species22  2854.3887262  29.84211  37.8000000 1.128032e+03
#> 23         1 Species23 21224.7572424 250.78947 238.2500000 5.975059e+04
#> 24         1 Species24 20537.1891212 223.68421 250.0000000 5.592105e+04
#> 25         1 Species25  6354.0000000  67.68421  80.3750000 5.440118e+03
#> 26         2  Species1    45.7929450  42.00000   2.4705882 1.037647e+02
#> 27         2  Species2     1.1896238   2.00000   0.1111111 2.222222e-01
#> 28         2  Species3   130.2992507 128.00000   6.4000000 8.192000e+02
#> 29         2  Species4    85.1132373  75.00000   4.6875000 3.515625e+02
#> 30         2  Species5    73.6485252  69.00000   3.8333333 2.645000e+02
#> 31         2  Species6   559.0000000 532.00000  28.0000000 1.489600e+04
#> 32         2  Species7   792.6842105 754.00000  39.6842105 2.992189e+04
#> 33         2  Species8     1.0259784   2.00000   0.1000000 2.000000e-01
#> 34         2  Species9    44.6801152  41.00000   2.4117647 9.888235e+01
#> 35         2 Species10   119.9506477 102.00000   6.8000000 6.936000e+02
#> 36         2 Species11     0.2294157   1.00000   0.0625000 6.250000e-02
#> 37         2 Species12    -0.9459053   0.00000   0.0000000 0.000000e+00
#> 38         2 Species13   222.1578947 212.00000  11.1578947 2.365474e+03
#> 39         2 Species14     2.5235731   3.00000   0.1875000 5.625000e-01
#> 40         2 Species15    20.0526316  20.00000   1.0526316 2.105263e+01
#> 41         2 Species16    20.0526316  20.00000   1.0526316 2.105263e+01
#> 42         2 Species17   527.8685044 489.00000  27.1666667 1.328450e+04
#> 43         2 Species18    -0.9459053   0.00000   0.0000000 0.000000e+00
#> 44         2 Species19    -0.8885233   0.00000   0.0000000 0.000000e+00
#> 45         2 Species20     1.2797542   2.00000   0.1176471 2.352941e-01
#> 46         2 Species21    -0.8885233   0.00000   0.0000000 0.000000e+00
#> 47         2 Species22   227.7581435 193.00000  12.8666667 2.483267e+03
#> 48         2 Species23    73.8704414  73.00000   3.6500000 2.664500e+02
#> 49         2 Species24    -0.9459053   0.00000   0.0000000 0.000000e+00
#> 50         2 Species25   156.2321148 137.00000   8.5625000 1.173062e+03

# Cz indices
net_bip <- mat_to_net(comat, weight = TRUE)
clust_bip <- netclu_greedy(net_bip, bipartite = TRUE)
site_species_metrics(cluster_object = clust_bip, comat = comat, 
net = net_bip, indices = "Cz")
#>         Node Bioregion Category         C          z
#> 1      Site1         5     site 0.7936508 -0.8017837
#> 2      Site2         3     site 0.7743056  1.0568458
#> 3      Site3         3     site 0.7922438 -1.4411534
#> 4      Site4         4     site 0.7750473  0.2070197
#> 5      Site5         5     site 0.7755102 -0.8017837
#> 6      Site6         2     site 0.7685950  0.7046643
#> 7      Site7         2     site 0.7845805  1.5854946
#> 8      Site8         2     site 0.7750000  0.7046643
#> 9      Site9         1     site 0.7901701  1.2076147
#> 10    Site10         4     site 0.7922438  1.6561573
#> 11    Site11         1     site 0.7826087  1.2076147
#> 12    Site12         1     site 0.7777778  1.2076147
#> 13    Site13         3     site 0.7561983  1.0568458
#> 14    Site14         3     site 0.7674858  1.0568458
#> 15    Site15         3     site 0.7847222  1.0568458
#> 16    Site16         3     site 0.7600000  1.0568458
#> 17    Site17         5     site 0.7755102 -0.8017837
#> 18    Site18         4     site 0.7851240  0.2070197
#> 19    Site19         5     site 0.7933884 -0.8017837
#> 20    Site20         2     site 0.7845805  0.7046643
#> 21  Species2         3  species 0.7901235 -1.4411534
#> 22  Species3         1  species 0.7850000 -0.7245688
#> 23  Species4         4  species 0.7890625  0.2070197
#> 24  Species6         2  species 0.7756233 -0.1761661
#> 25  Species7         2  species 0.7922438 -0.1761661
#> 26  Species8         5  species 0.7850000  1.0690450
#> 27  Species9         3  species 0.7820069 -1.4411534
#> 28 Species10         4  species 0.7466667 -1.2421180
#> 29 Species11         1  species 0.7890625 -0.7245688
#> 30 Species12         1  species 0.7889273 -0.7245688
#> 31 Species13         4  species 0.7756233 -1.2421180
#> 32 Species14         5  species 0.7968750  1.0690450
#> 33 Species15         3  species 0.7811634 -0.1921538
#> 34 Species16         3  species 0.7756233 -0.1921538
#> 35 Species17         2  species 0.7839506 -0.1761661
#> 36 Species18         2  species 0.7681661 -1.0569964
#> 37 Species19         3  species 0.7466667 -0.1921538
#> 38 Species21         1  species 0.7644444 -0.7245688
#> 39 Species22         2  species 0.7733333 -1.9378267
#> 40 Species23         4  species 0.7850000  0.2070197
#> 41 Species24         5  species 0.7958478  1.0690450
#> 42  Species1         2  species 0.7820069 -0.1761661
#> 43  Species5         3  species 0.7777778 -0.1921538
#> 44 Species20         3  species 0.7612457 -0.1921538
#> 45 Species25         1  species 0.7812500 -0.7245688