Calculate contribution metrics of sites and species

This function calculates metrics to assess the contribution of a given species or site to its bioregion.

Usage

site_species_metrics(
  cluster_object,
  comat,
  indices = c("rho"),
  bipartite_link = NULL
)

Arguments

cluster_object

A bioregion.clusters object, a data.frame, or a list of data.frames containing multiple partitions. At least two partitions are required. If a list of data.frames is provided, they must all have the same number of rows (i.e., the same items in the clustering).

comat

A co-occurrence matrix with sites as rows and species as columns.

indices

A character specifying the contribution metric to compute. Available options are rho and Cz.

bipartite_link

NULL by default. Required for Cz indices. A data.frame where each row represents an interaction between two nodes and an optional third column indicating the interaction's weight.

Value

A data.frame with columns Bioregion, Species, and the desired summary statistics, or a list of data.frames if Cz and other indices are selected.

Details

The \(\rho\) metric is derived from Lenormand et al. (2019) with the following formula:

\(\rho_{ij} = \frac{n_{ij} - \frac{n_i n_j}{n}}{\sqrt{\left(\frac{n - n_j}{ n-1}\right) \left(1-\frac{n_j}{n}\right) \frac{n_i n_j}{n}}}\)

where \(n\) is the number of sites, \(n_i\) is the number of sites in which species \(i\) is present, \(n_j\) is the number of sites in bioregion \(j\), and \(n_{ij}\) is the number of occurrences of species \(i\) in sites of bioregion \(j\).

Affinity, fidelity, and individual contributions describe how species are linked to their bioregions. These metrics are described in Bernardo-Madrid et al. (2019):

Affinity of species to their region: \(A_i = \frac{R_i}{Z}\), where \(R_i\) is the occurrence/range size of species \(i\) in its associated bioregion, and \(Z\) is the total size (number of sites) of the bioregion. High affinity indicates that the species occupies most sites in its bioregion.

Fidelity of species to their region: \(F_i = \frac{R_i}{D_i}\), where \(R_i\) is the occurrence/range size of species \(i\) in its bioregion, and \(D_i\) is its total range size. High fidelity indicates that the species is not present in other regions.

Indicator Value of species: \(IndVal = F_i \cdot A_i\).

Cz metrics are derived from Guimerà & Amaral (2005):

Participation coefficient: \(C_i = 1 - \sum_{s=1}^{N_M}{\left(\frac{k_{is}}{k_i}\right)^2}\), where \(k_{is}\) is the number of links of node \(i\) to nodes in bioregion \(s\), and \(k_i\) is the total degree of node \(i\). A high value means links are uniformly distributed; a low value means links are within the node's bioregion.

Within-bioregion degree z-score: \(z_i = \frac{k_i - \overline{k_{si}}}{\sigma_{k_{si}}}\), where \(k_i\) is the number of links of node \(i\) to nodes in its bioregion \(s_i\), \(\overline{k_{si}}\) is the average degree of nodes in \(s_i\), and \(\sigma_{k_{si}}\) is the standard deviation of degrees in \(s_i\).

References

Bernardo-Madrid R, Calatayud J, González‐Suárez M, Rosvall M, Lucas P, Antonelli A & Revilla E (2019) Human activity is altering the world’s zoogeographical regions. Ecology Letters 22, 1297–1305.

Guimerà R & Amaral LAN (2005) Functional cartography of complex metabolic networks. Nature 433, 895–900.

Lenormand M, Papuga G, Argagnon O, Soubeyrand M, Alleaume S & Luque S (2019) Biogeographical network analysis of plant species distribution in the Mediterranean region. Ecology and Evolution 9, 237–250.

See also

Author

Pierre Denelle (pierre.denelle@gmail.com)
Boris Leroy (leroy.boris@gmail.com)
Maxime Lenormand (maxime.lenormand@inrae.fr)

Examples

comat <- matrix(sample(1000, 50), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

comat <- matrix(sample(0:1000, size = 500, replace = TRUE, prob = 1/1:1001),
                20, 25)
rownames(comat) <- paste0("Site",1:20)
colnames(comat) <- paste0("Species",1:25)

dissim <- dissimilarity(comat, metric = "Simpson")
clust1 <- nhclu_kmeans(dissim, n_clust = 3, index = "Simpson")

net <- similarity(comat, metric = "Simpson")
com <- netclu_greedy(net)

site_species_metrics(cluster_object = clust1, comat = comat,
indices = "rho")
#>    Bioregion   Species         rho
#> 1          3  Species1 1191.229995
#> 2          3  Species2  338.426329
#> 3          3  Species3 1236.049759
#> 4          3  Species4  995.801235
#> 5          3  Species5  393.120705
#> 6          3  Species6 1229.184152
#> 7          3  Species7 1376.897148
#> 8          3  Species8  283.831132
#> 9          3  Species9  392.148314
#> 10         3 Species10  377.235660
#> 11         3 Species11  698.520261
#> 12         3 Species12 2033.028308
#> 13         3 Species13  808.701425
#> 14         3 Species14 1134.941452
#> 15         3 Species15 1690.869865
#> 16         3 Species16 1013.666612
#> 17         3 Species17 1467.296632
#> 18         3 Species18 1426.422338
#> 19         3 Species19  195.044290
#> 20         3 Species20  240.234261
#> 21         3 Species21  267.784143
#> 22         3 Species22 2000.140620
#> 23         3 Species23 1029.668442
#> 24         3 Species24  216.378834
#> 25         3 Species25 1322.873723
#> 26         2  Species1   17.762011
#> 27         2  Species2  284.675902
#> 28         2  Species3  125.179418
#> 29         2  Species4  427.341085
#> 30         2  Species5  437.062360
#> 31         2  Species6  955.319364
#> 32         2  Species7  268.880931
#> 33         2  Species8  164.763439
#> 34         2  Species9  523.085323
#> 35         2 Species10   39.485148
#> 36         2 Species11    1.605051
#> 37         2 Species12  848.467163
#> 38         2 Species13  752.870899
#> 39         2 Species14  108.779583
#> 40         2 Species15  559.112793
#> 41         2 Species16  353.320570
#> 42         2 Species17   31.918653
#> 43         2 Species18   -1.661700
#> 44         2 Species19   84.084431
#> 45         2 Species20   95.365694
#> 46         2 Species21  481.436687
#> 47         2 Species22  258.211417
#> 48         2 Species23   50.776337
#> 49         2 Species24  442.645146
#> 50         2 Species25  132.449843
#> 51         1  Species1  232.215557
#> 52         1  Species2  712.417704
#> 53         1  Species3   58.735849
#> 54         1  Species4  762.528787
#> 55         1  Species5  337.361842
#> 56         1  Species6  737.355158
#> 57         1  Species7  142.603633
#> 58         1  Species8  692.225571
#> 59         1  Species9  237.319552
#> 60         1 Species10 1333.206945
#> 61         1 Species11  441.312542
#> 62         1 Species12   68.818608
#> 63         1 Species13  371.947144
#> 64         1 Species14  746.247241
#> 65         1 Species15  535.750103
#> 66         1 Species16  261.489779
#> 67         1 Species17  780.133085
#> 68         1 Species18  976.890267
#> 69         1 Species19  414.031670
#> 70         1 Species20 1285.807906
#> 71         1 Species21  794.272056
#> 72         1 Species22 1084.670815
#> 73         1 Species23  588.287424
#> 74         1 Species24  412.636487
#> 75         1 Species25  383.012426

# Contribution metrics
site_species_metrics(cluster_object = com, comat = comat,
indices = c("rho", "affinity", "fidelity", "indicator_value"))
#>    Bioregion   Species           rho  affinity     fidelity       indval
#> 1          1  Species1 10589.6008597 122.42105 122.42105263 1.498691e+04
#> 2          1  Species2  9999.6959954 118.68421 112.75000000 1.338164e+04
#> 3          1  Species3  8643.2078343  88.84211 112.53333333 9.997698e+03
#> 4          1  Species4 16118.5200460 185.84211 185.84210526 3.453729e+04
#> 5          1  Species5  8470.6678341  95.47368 100.77777778 9.621626e+03
#> 6          1  Species6 21153.2030671 249.94737 237.45000000 5.935000e+04
#> 7          1  Species7 12934.8686460 145.31579 153.38888889 2.228983e+04
#> 8          1  Species8  7910.7844818  86.68421  96.88235294 8.398170e+03
#> 9          1  Species9  8124.2158318  89.00000  99.47058824 8.852882e+03
#> 10         1 Species10 13527.1604982 133.89474 181.71428571 2.433059e+04
#> 11         1 Species11  8465.9537889  95.42105 100.72222222 9.611020e+03
#> 12         1 Species12 18426.4423287 212.31579 212.31578947 4.507799e+04
#> 13         1 Species13 13862.6087202 151.26316 169.05882353 2.557237e+04
#> 14         1 Species14 12787.4035902 147.63158 147.63157895 2.179508e+04
#> 15         1 Species15 20261.7681997 233.36842 233.36842105 5.446082e+04
#> 16         1 Species16 11761.0713892 132.21053 139.55555556 1.845071e+04
#> 17         1 Species17 16991.3182832 185.21053 207.00000000 3.833858e+04
#> 18         1 Species18 18038.5302616 178.31579 242.00000000 4.315242e+04
#> 19         1 Species19  4534.0000000  48.52632  57.62500000 2.796329e+03
#> 20         1 Species20 12536.8470917 128.52632 162.80000000 2.092408e+04
#> 21         1 Species21 11451.8046073 125.10526 139.82352941 1.749266e+04
#> 22         1 Species22 24808.7880450 285.52632 285.52631579 8.152528e+04
#> 23         1 Species23 12459.0000000 131.94737 156.68750000 2.067450e+04
#> 24         1 Species24  7815.4155502  88.15789  93.05555556 8.203582e+03
#> 25         1 Species25 13178.6582576 143.84211 160.76470588 2.312473e+04
#> 26         2  Species1     1.1052632   2.00000   0.10526316 2.105263e-01
#> 27         2  Species2     0.0000000   1.00000   0.05000000 5.000000e-02
#> 28         2  Species3   374.6606652 317.00000  21.13333333 6.699267e+03
#> 29         2  Species4     2.1578947   3.00000   0.15789474 4.736842e-01
#> 30         2  Species5    -0.9733285   0.00000   0.00000000 0.000000e+00
#> 31         2  Species6     8.2078268   9.00000   0.45000000 4.050000e+00
#> 32         2  Species7    -0.9733285   0.00000   0.00000000 0.000000e+00
#> 33         2  Species8   165.9785599 150.00000   8.82352941 1.323529e+03
#> 34         2  Species9    29.1004984  27.00000   1.58823529 4.288235e+01
#> 35         2 Species10    -0.8583951   0.00000   0.00000000 0.000000e+00
#> 36         2 Species11    21.7376704  21.00000   1.16666667 2.450000e+01
#> 37         2 Species12   595.8421053 567.00000  29.84210526 1.692047e+04
#> 38         2 Species13     5.7310733   6.00000   0.35294118 2.117647e+00
#> 39         2 Species14   474.7894737 452.00000  23.78947368 1.075284e+04
#> 40         2 Species15     2.1578947   3.00000   0.15789474 4.736842e-01
#> 41         2 Species16     9.8414329  10.00000   0.55555556 5.555556e+00
#> 42         2 Species17     5.7310733   6.00000   0.35294118 2.117647e+00
#> 43         2 Species18     1.5941623   2.00000   0.14285714 2.857143e-01
#> 44         2 Species19   164.2616655 144.00000   9.00000000 1.296000e+03
#> 45         2 Species20    -0.8885233   0.00000   0.00000000 0.000000e+00
#> 46         2 Species21     2.3925840   3.00000   0.17647059 5.294118e-01
#> 47         2 Species22     1.1052632   2.00000   0.10526316 2.105263e-01
#> 48         2 Species23    -0.9176629   0.00000   0.00000000 0.000000e+00
#> 49         2 Species24     0.1081476   1.00000   0.05555556 5.555556e-02
#> 50         2 Species25    76.9521785  70.00000   4.11764706 2.882353e+02

# Cz indices
net_bip <- mat_to_net(comat, weight = TRUE)
clust_bip <- netclu_greedy(net_bip, bipartite = TRUE)
site_species_metrics(cluster_object = clust_bip, comat = comat, 
bipartite_link = net_bip, indices = "Cz")
#> Warning: Column 'Node1' has been renamed 'Sites'. If this column does not
#>             correspond to sites, rename it before running the function.
#> Warning: Column 'Node2' has been renamed 'Species'. If this column does not
#>             correspond to species, rename it before running the function.
#>         Node Bioregion Category         C          z
#> 1      Site1         1     site 0.8163265  1.1547005
#> 2      Site2         1     site 0.8199446 -1.1547005
#> 3      Site3         3     site 0.8128544  1.2076147
#> 4      Site4         5     site 0.8144044  1.0954451
#> 5      Site5         3     site 0.8159722  1.2076147
#> 6      Site6         3     site 0.8090737  1.2076147
#> 7      Site7         2     site 0.8140496 -1.0000000
#> 8      Site8         2     site 0.7977316  0.0000000
#> 9      Site9         2     site 0.8159722  0.0000000
#> 10    Site10         4     site 0.7936508  1.0690450
#> 11    Site11         4     site 0.7845805  1.0690450
#> 12    Site12         2     site 0.7809917  0.0000000
#> 13    Site13         5     site 0.8057851  1.0954451
#> 14    Site14         1     site 0.8072562  1.1547005
#> 15    Site15         2     site 0.8117914 -1.0000000
#> 16    Site16         1     site 0.8194444  1.1547005
#> 17    Site17         2     site 0.8200000 -1.0000000
#> 18    Site18         6     site 0.8128544  1.1547005
#> 19    Site19         2     site 0.7981859  0.0000000
#> 20    Site20         4     site 0.8099174  1.0690450
#> 21  Species1         4  species 0.7645429 -0.8017837
#> 22  Species2         3  species 0.7800000 -0.7245688
#> 23  Species3         5  species 0.7733333 -0.7302967
#> 24  Species5         1  species 0.7839506  0.0000000
#> 25  Species6         1  species 0.7800000  0.0000000
#> 26  Species7         4  species 0.8024691 -0.8017837
#> 27  Species8         6  species 0.7958478 -0.5773503
#> 28 Species11         3  species 0.7530864 -0.7245688
#> 29 Species12         2  species 0.7700831  1.0000000
#> 30 Species13         5  species 0.7958478 -0.7302967
#> 31 Species15         4  species 0.7645429 -0.8017837
#> 32 Species16         1  species 0.7777778  0.0000000
#> 33 Species17         1  species 0.7474048 -1.1547005
#> 34 Species18         4  species 0.7244898 -0.8017837
#> 35 Species19         3  species 0.7421875 -0.7245688
#> 36 Species20         3  species 0.7466667 -0.7245688
#> 37 Species21         6  species 0.7820069 -0.5773503
#> 38 Species22         2  species 0.7756233  1.0000000
#> 39 Species23         2  species 0.8125000 -2.0000000
#> 40 Species24         2  species 0.7592593  1.0000000
#> 41 Species25         1  species 0.7820069 -1.1547005
#> 42  Species4         2  species 0.7756233  1.0000000
#> 43  Species9         5  species 0.7820069 -0.7302967
#> 44 Species10         3  species 0.8061224 -0.7245688
#> 45 Species14         2  species 0.7756233  1.0000000