Calculate contribution metrics of sites and species https://github.com/Farewe/biogeonetworks/blob/master/R/networkmetrics.R

This function calculates metrics that assess the contribution of a given species or site to its bioregion.

Usage

site_species_metrics(
  cluster_object,
  comat,
  indices = c("rho", "Cz"),
  bipartite_link = NULL
)

Arguments

cluster_object

a bioregion.clusters object or a data.frame or a list of data.frame containing multiple partitions. At least two partitions are required. If a list of data.frame is provided, they should all have the same number of rows (i.e., same items in the clustering for all partitions).

comat

a co-occurrence matrix with sites as rows and species as columns.

indices

a character specifying the contribution metric to compute. Available options are rho and Cz.

bipartite_link

NULL by default. Needed for Cz indices. A data.frame where each row represents the interaction between two nodes and an optional third column indicating the weight of the interaction.

Value

A list of data.frames if multiples indices are selected or a single data.frame with three columns if one index is selected. Each data.frame has three columns: the species, the bioregion, and the desired summary statistics.

Details

The \(\rho\) metric is derived from Lenormand2019bioregion. Its formula is the following: \(\rho_{ij} = (n_ij - ((n_i n_j)/n))/(sqrt(((n - n_j)/(n-1)) (1-(n_j/n)) ((n_i n_j)/n)))\)

with \(n\) the number of sites, \(n_i\) the number of sites in which species \(i\) is present, \(n_j\) the number of sites belonging to the bioregion \(j\), \(n_ij\) the number of occurrences of species \(i\) in sites belonging to the bioregion \(j\).

Cz metrics are derived from Guimera2005bioregion. Their respective formula are: \(C_i = 1 - \sum_{s=1}^{N_M}{{(\frac{k_is}{k_i}})^2}\)

where \(k_{is}\) is the number of links of node (species or site) \(i\) to nodes in bioregion \(s\), and \(k_i\) is the total degree of node \(i\). The participation coefficient of a node is therefore close to 1 if its links are uniformly distributed among all the bioregions and 0 if all its links are within its own bioregion.

And: \(z_i = \frac{k_i - \overline{k_{si}}}{\sigma_{k_{si}}}\)

where \(k_i\) is the number of links of node (species or site) \(i\) to other nodes in its bioregion \(s_i\), \(\overline{k_{si}}\) is the average of \(k\) over all the nodes in \(s_i\), and \(\sigma_{k_{si}}\) is the standard deviation of \(k\) in \(s_i\). The within-bioregion degree z-score measures how well-connected node \(i\) is to other nodes in the bioregion.

See also

Author

Pierre Denelle (pierre.denelle@gmail.com), Boris Leroy (leroy.boris@gmail.com) and Maxime Lenormand (maxime.lenormand@inrae.fr)

Examples

comat <- matrix(sample(1000, 50), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

comat <- matrix(sample(0:1000, size = 500, replace = TRUE, prob = 1/1:1001),
                20, 25)
rownames(comat) <- paste0("Site",1:20)
colnames(comat) <- paste0("Species",1:25)

dissim <- dissimilarity(comat, metric = "Simpson")
clust1 <- nhclu_kmeans(dissim, n_clust = 3, index = "Simpson")

net <- similarity(comat, metric = "Simpson")
com <- netclu_greedy(net)

site_species_metrics(cluster_object = clust1, comat = comat,
indices = "rho")
#>    Bioregion   Species         rho
#> 1          2  Species1  546.333333
#> 2          2  Species2  201.962318
#> 3          2  Species3  824.350273
#> 4          2  Species4 1581.514659
#> 5          2  Species5 1935.148275
#> 6          2  Species6  115.553883
#> 7          2  Species7 1165.121212
#> 8          2  Species8  343.956753
#> 9          2  Species9 1120.585992
#> 10         2 Species10  697.948652
#> 11         2 Species11 1143.909755
#> 12         2 Species12 1526.803421
#> 13         2 Species13  640.663189
#> 14         2 Species14  622.605814
#> 15         2 Species15  976.657539
#> 16         2 Species16   89.284356
#> 17         2 Species17 1102.682905
#> 18         2 Species18  622.696970
#> 19         2 Species19  875.610336
#> 20         2 Species20 1223.999776
#> 21         2 Species21 1445.773018
#> 22         2 Species22 1101.437569
#> 23         2 Species23 1246.364864
#> 24         2 Species24  746.824686
#> 25         2 Species25  228.757576
#> 26         3  Species1  573.581451
#> 27         3  Species2  325.301264
#> 28         3  Species3  628.892253
#> 29         3  Species4  224.483296
#> 30         3  Species5   98.877698
#> 31         3  Species6  187.633795
#> 32         3  Species7  333.754401
#> 33         3  Species8  952.016688
#> 34         3  Species9  729.405913
#> 35         3 Species10  354.644861
#> 36         3 Species11  147.141774
#> 37         3 Species12   24.143955
#> 38         3 Species13  318.926106
#> 39         3 Species14  691.248724
#> 40         3 Species15 1082.368359
#> 41         3 Species16  335.896887
#> 42         3 Species17 1583.612202
#> 43         3 Species18  212.956992
#> 44         3 Species19   59.460910
#> 45         3 Species20  679.144634
#> 46         3 Species21  296.526209
#> 47         3 Species22  737.259324
#> 48         3 Species23  228.798947
#> 49         3 Species24  137.800338
#> 50         3 Species25 1259.475028
#> 51         1  Species1  726.204753
#> 52         1  Species2  590.367446
#> 53         1  Species3  247.247636
#> 54         1  Species4  464.278442
#> 55         1  Species5   51.773874
#> 56         1  Species6  541.755401
#> 57         1  Species7   15.724304
#> 58         1  Species8   30.051302
#> 59         1  Species9  497.189773
#> 60         1 Species10  110.469378
#> 61         1 Species11  162.609199
#> 62         1 Species12  344.627737
#> 63         1 Species13  194.909017
#> 64         1 Species14  993.247417
#> 65         1 Species15  144.038998
#> 66         1 Species16    3.195456
#> 67         1 Species17   23.238347
#> 68         1 Species18    6.894241
#> 69         1 Species19   -1.776432
#> 70         1 Species20   33.354813
#> 71         1 Species21   11.374927
#> 72         1 Species22  123.030647
#> 73         1 Species23  219.207316
#> 74         1 Species24 1028.553915
#> 75         1 Species25   51.723792

site_species_metrics(cluster_object = com, comat = comat, indices = "rho")
#>    Bioregion   Species           rho
#> 1          1  Species1  1.351236e+04
#> 2          1  Species2  8.054000e+03
#> 3          1  Species3  1.280759e+04
#> 4          1  Species4  1.653640e+04
#> 5          1  Species5  1.533552e+04
#> 6          1  Species6  5.951097e+03
#> 7          1  Species7  1.152562e+04
#> 8          1  Species8  1.022400e+04
#> 9          1  Species9  1.717300e+04
#> 10         1 Species10  7.593855e+03
#> 11         1 Species11  1.084061e+04
#> 12         1 Species12  1.407900e+04
#> 13         1 Species13  8.522522e+03
#> 14         1 Species14  1.675796e+04
#> 15         1 Species15  1.590900e+04
#> 16         1 Species16  3.278354e+03
#> 17         1 Species17  1.749806e+04
#> 18         1 Species18  6.386705e+03
#> 19         1 Species19  7.009000e+03
#> 20         1 Species20  1.468238e+04
#> 21         1 Species21  1.331671e+04
#> 22         1 Species22  1.491477e+04
#> 23         1 Species23  1.269693e+04
#> 24         1 Species24  1.366900e+04
#> 25         1 Species25  8.997456e+03
#> 26         2  Species1 -1.000000e+00
#> 27         2  Species2 -9.176629e-01
#> 28         2  Species3  1.081476e-01
#> 29         2  Species4  7.364853e+01
#> 30         2  Species5  9.970582e+01
#> 31         2  Species6  1.018239e+01
#> 32         2  Species7  5.263158e-02
#> 33         2  Species8 -9.176629e-01
#> 34         2  Species9  8.002631e+01
#> 35         2 Species10  2.748031e+02
#> 36         2 Species11  1.463371e+01
#> 37         2 Species12 -9.176629e-01
#> 38         2 Species13  2.930800e+01
#> 39         2 Species14  1.092291e+01
#> 40         2 Species15  2.032623e+02
#> 41         2 Species16  7.956733e+00
#> 42         2 Species17  7.657933e+02
#> 43         2 Species18  7.421053e+00
#> 44         2 Species19  1.628852e+01
#> 45         2 Species20  2.242352e+01
#> 46         2 Species21  3.352576e+00
#> 47         2 Species22 -9.733285e-01
#> 48         2 Species23  2.961744e-01
#> 49         2 Species24  1.376494e+00
#> 50         2 Species25  6.632105e+02

# Cz indices
net_bip <- mat_to_net(comat, weight = TRUE)
clust_bip <- netclu_greedy(net_bip, bipartite = TRUE)
site_species_metrics(cluster_object = clust_bip, comat = comat, 
bipartite_link = net_bip, indices = "Cz")
#> Warning: Column 'Node1' has been renamed 'Sites'. If this column does not
#>             correspond to sites, rename it before running the function.
#> Warning: Column 'Node2' has been renamed 'Species'. If this column does not
#>             correspond to species, rename it before running the function.
#>         Node Bioregion Category         C          z
#> 1      Site1         5     site 0.8279773 -0.3333333
#> 2      Site2         4     site 0.8350000        NaN
#> 3      Site3         2     site 0.8200000  1.0690450
#> 4      Site4         5     site 0.8279773  1.1666667
#> 5      Site5         1     site 0.8429752        NaN
#> 6      Site6         6     site 0.8344671  1.4031215
#> 7      Site7         1     site 0.8279773        NaN
#> 8      Site8         5     site 0.8300000  1.1666667
#> 9      Site9         6     site 0.8393195  0.4677072
#> 10    Site10         5     site 0.8317580  1.1666667
#> 11    Site11         2     site 0.8347107  1.0690450
#> 12    Site12         3     site 0.8250000        NaN
#> 13    Site13         3     site 0.8253968        NaN
#> 14    Site14         7     site 0.8347107  1.0954451
#> 15    Site15         6     site 0.8388430  1.4031215
#> 16    Site16         1     site 0.8402778        NaN
#> 17    Site17         7     site 0.8148148  1.0954451
#> 18    Site18         3     site 0.8388430        NaN
#> 19    Site19         2     site 0.8350000  1.0690450
#> 20    Site20         1     site 0.8355388        NaN
#> 21  Species1         7  species 0.8365651 -0.7302967
#> 22  Species2         2  species 0.8125000 -0.8017837
#> 23  Species3         6  species 0.8395062 -0.4677072
#> 24  Species4         1  species 0.8395062        NaN
#> 25  Species6         2  species 0.8442907 -0.8017837
#> 26  Species7         5  species 0.8365651 -0.3333333
#> 27  Species8         6  species 0.8125000 -1.4031215
#> 28  Species9         6  species 0.8400000 -0.4677072
#> 29 Species10         3  species 0.8395062        NaN
#> 30 Species11         5  species 0.8304498 -0.3333333
#> 31 Species12         7  species 0.8046875 -0.7302967
#> 32 Species13         1  species 0.8271605        NaN
#> 33 Species14         2  species 0.8209877 -0.8017837
#> 34 Species15         3  species 0.8281250        NaN
#> 35 Species17         3  species 0.8333333        NaN
#> 36 Species18         5  species 0.8365651 -0.3333333
#> 37 Species19         6  species 0.8203125 -0.4677072
#> 38 Species20         5  species 0.8235294 -0.3333333
#> 39 Species21         1  species 0.8271605        NaN
#> 40 Species22         6  species 0.8333333 -0.4677072
#> 41 Species23         1  species 0.8177778        NaN
#> 42 Species24         7  species 0.8437500 -0.7302967
#> 43 Species25         2  species 0.8365651 -0.8017837
#> 44 Species16         4  species 0.8304498        NaN
#> 45  Species5         5  species 0.7928994 -1.8333333