Calculate contribution metrics of sites and species https://github.com/Farewe/biogeonetworks/blob/master/R/networkmetrics.R

This function calculates metrics that assess the contribution of a given species or site to its bioregion.

Usage

site_species_metrics(
  cluster_object,
  comat,
  indices = c("rho", "Cz"),
  bipartite_link = NULL
)

Arguments

cluster_object

a bioregion.clusters object or a data.frame or a list of data.frame containing multiple partitions. At least two partitions are required. If a list of data.frame is provided, they should all have the same number of rows (i.e., same items in the clustering for all partitions).

comat

a co-occurrence matrix with sites as rows and species as columns.

indices

a character specifying the contribution metric to compute. Available options are rho and Cz.

bipartite_link

NULL by default. Needed for Cz indices. A data.frame where each row represents the interaction between two nodes and an optional third column indicating the weight of the interaction.

Value

A list of data.frames if multiples indices are selected or a single data.frame with three columns if one index is selected. Each data.frame has three columns: the species, the bioregion, and the desired summary statistics.

Details

The \(\rho\) metric is derived from Lenormand2019bioregion. Its formula is the following: \(\rho_{ij} = (n_ij - ((n_i n_j)/n))/(sqrt(((n - n_j)/(n-1)) (1-(n_j/n)) ((n_i n_j)/n)))\)

with \(n\) the number of sites, \(n_i\) the number of sites in which species \(i\) is present, \(n_j\) the number of sites belonging to the bioregion \(j\), \(n_ij\) the number of occurrences of species \(i\) in sites belonging to the bioregion \(j\).

Cz metrics are derived from Guimera2005bioregion. Their respective formula are: \(C_i = 1 - \sum_{s=1}^{N_M}{{(\frac{k_is}{k_i}})^2}\)

where \(k_{is}\) is the number of links of node (species or site) \(i\) to nodes in bioregion \(s\), and \(k_i\) is the total degree of node \(i\). The participation coefficient of a node is therefore close to 1 if its links are uniformly distributed among all the bioregions and 0 if all its links are within its own bioregion.

And: \(z_i = \frac{k_i - \overline{k_{si}}}{\sigma_{k_{si}}}\)

where \(k_i\) is the number of links of node (species or site) \(i\) to other nodes in its bioregion \(s_i\), \(\overline{k_{si}}\) is the average of \(k\) over all the nodes in \(s_i\), and \(\sigma_{k_{si}}\) is the standard deviation of \(k\) in \(s_i\). The within-bioregion degree z-score measures how well-connected node \(i\) is to other nodes in the bioregion.

See also

Author

Pierre Denelle (pierre.denelle@gmail.com), Boris Leroy (leroy.boris@gmail.com) and Maxime Lenormand (maxime.lenormand@inrae.fr)

Examples

comat <- matrix(sample(1000, 50), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

comat <- matrix(sample(0:1000, size = 500, replace = TRUE, prob = 1/1:1001),
                20, 25)
rownames(comat) <- paste0("Site",1:20)
colnames(comat) <- paste0("Species",1:25)

dissim <- dissimilarity(comat, metric = "Simpson")
clust1 <- nhclu_kmeans(dissim, n_clust = 3, index = "Simpson")

net <- similarity(comat, metric = "Simpson")
com <- netclu_greedy(net)

site_species_metrics(cluster_object = clust1, comat = comat,
indices = "rho")
#>    Bioregion   Species         rho
#> 1          3  Species1 1154.963637
#> 2          3  Species2  427.164316
#> 3          3  Species3  771.921377
#> 4          3  Species4 1419.060606
#> 5          3  Species5 1465.698396
#> 6          3  Species6  227.679912
#> 7          3  Species7  432.069646
#> 8          3  Species8  313.536415
#> 9          3  Species9  678.023274
#> 10         3 Species10  668.151515
#> 11         3 Species11 1377.570009
#> 12         3 Species12  881.442055
#> 13         3 Species13  400.935977
#> 14         3 Species14  538.125888
#> 15         3 Species15  146.485422
#> 16         3 Species16  845.974876
#> 17         3 Species17  620.020353
#> 18         3 Species18  880.878788
#> 19         3 Species19  721.830957
#> 20         3 Species20  712.270018
#> 21         3 Species21  764.515152
#> 22         3 Species22  754.992612
#> 23         3 Species23  226.398472
#> 24         3 Species24  791.181818
#> 25         3 Species25  219.060606
#> 26         1  Species1  278.834764
#> 27         1  Species2  231.438587
#> 28         1  Species3  157.362235
#> 29         1  Species4  546.207316
#> 30         1  Species5  270.974476
#> 31         1  Species6  133.167922
#> 32         1  Species7  291.863586
#> 33         1  Species8   34.072893
#> 34         1  Species9  659.675952
#> 35         1 Species10  106.062641
#> 36         1 Species11  725.121640
#> 37         1 Species12   65.551970
#> 38         1 Species13  364.486142
#> 39         1 Species14    4.145007
#> 40         1 Species15  342.407197
#> 41         1 Species16   53.461051
#> 42         1 Species17   12.286985
#> 43         1 Species18   33.384430
#> 44         1 Species19  106.068056
#> 45         1 Species20   29.518981
#> 46         1 Species21  239.872056
#> 47         1 Species22  124.550963
#> 48         1 Species23  499.388685
#> 49         1 Species24  815.184618
#> 50         1 Species25  452.472802
#> 51         2  Species1  988.809364
#> 52         2  Species2   80.245428
#> 53         2  Species3  580.867594
#> 54         2  Species4  766.857304
#> 55         2  Species5  733.021506
#> 56         2  Species6 1198.065456
#> 57         2  Species7  410.651388
#> 58         2  Species8  203.830604
#> 59         2  Species9  311.055872
#> 60         2 Species10  782.767207
#> 61         2 Species11  453.640620
#> 62         2 Species12  134.626627
#> 63         2 Species13  536.871053
#> 64         2 Species14 1061.820312
#> 65         2 Species15 1294.270090
#> 66         2 Species16  702.816892
#> 67         2 Species17  758.094497
#> 68         2 Species18  199.404112
#> 69         2 Species19 1185.112229
#> 70         2 Species20  279.574935
#> 71         2 Species21  436.284884
#> 72         2 Species22  878.489968
#> 73         2 Species23  570.750609
#> 74         2 Species24 1336.078263
#> 75         2 Species25  410.946891

site_species_metrics(cluster_object = com, comat = comat, indices = "rho")
#>    Bioregion   Species           rho
#> 1          1  Species1  1.766984e+04
#> 2          1  Species2  5.426050e+03
#> 3          1  Species3  1.143400e+04
#> 4          1  Species4  1.984423e+04
#> 5          1  Species5  1.709266e+04
#> 6          1  Species6  1.174285e+04
#> 7          1  Species7  8.415259e+03
#> 8          1  Species8  3.428726e+03
#> 9          1  Species9  1.152066e+04
#> 10         1 Species10  1.116773e+04
#> 11         1 Species11  1.706064e+04
#> 12         1 Species12  5.839168e+03
#> 13         1 Species13  8.192539e+03
#> 14         1 Species14  1.237400e+04
#> 15         1 Species15  1.350400e+04
#> 16         1 Species16  1.224247e+04
#> 17         1 Species17  1.027400e+04
#> 18         1 Species18  7.042834e+03
#> 19         1 Species19  1.541318e+04
#> 20         1 Species20  4.986988e+03
#> 21         1 Species21  1.080525e+04
#> 22         1 Species22  1.334843e+04
#> 23         1 Species23  9.511520e+03
#> 24         1 Species24  2.188144e+04
#> 25         1 Species25  7.905437e+03
#> 26         2  Species1  1.507528e+02
#> 27         2  Species2 -8.885233e-01
#> 28         2  Species3 -9.176629e-01
#> 29         2  Species4  1.284737e+02
#> 30         2  Species5  3.537508e+02
#> 31         2  Species6  4.801860e+01
#> 32         2  Species7  5.731073e+00
#> 33         2  Species8  1.748812e+02
#> 34         2  Species9  1.125817e+02
#> 35         2 Species10  1.611053e+02
#> 36         2 Species11  4.127579e+02
#> 37         2 Species12  5.334102e+02
#> 38         2 Species13  3.407731e+02
#> 39         2 Species14 -9.176629e-01
#> 40         2 Species15 -9.176629e-01
#> 41         2 Species16 -9.459053e-01
#> 42         2 Species17  9.314279e+01
#> 43         2 Species18  3.221579e+02
#> 44         2 Species19 -8.885233e-01
#> 45         2 Species20  6.381791e+02
#> 46         2 Species21  5.263158e-02
#> 47         2 Species22  1.129522e+01
#> 48         2 Species23  4.618244e+00
#> 49         2 Species24  4.263158e+00
#> 50         2 Species25  2.157895e+00

# Cz indices
net_bip <- mat_to_net(comat, weight = TRUE)
clust_bip <- netclu_greedy(net_bip, bipartite = TRUE)
site_species_metrics(cluster_object = clust_bip, comat = comat, 
bipartite_link = net_bip, indices = "Cz")
#> Warning: Column 'Node1' has been renamed 'Sites'. If this column does not
#>             correspond to sites, rename it before running the function.
#> Warning: Column 'Node2' has been renamed 'Species'. If this column does not
#>             correspond to species, rename it before running the function.
#>         Node Bioregion Category         C          z
#> 1      Site1         4     site 0.8072562 -0.8017837
#> 2      Site2         4     site 0.8204159 -0.8017837
#> 3      Site3         4     site 0.8144044 -0.8017837
#> 4      Site4         5     site 0.8181818  1.1666667
#> 5      Site5         5     site 0.8194444  1.1666667
#> 6      Site6         4     site 0.8117914 -0.8017837
#> 7      Site7         2     site 0.7900000  1.9666868
#> 8      Site8         3     site 0.8163265  1.2421180
#> 9      Site9         5     site 0.8181818  1.1666667
#> 10    Site10         1     site 0.8057851  0.4281744
#> 11    Site11         1     site 0.8160000  0.4281744
#> 12    Site12         1     site 0.8024691 -2.1408721
#> 13    Site13         5     site 0.8050000 -0.3333333
#> 14    Site14         3     site 0.8057851 -0.2070197
#> 15    Site15         6     site 0.8140496  1.1547005
#> 16    Site16         1     site 0.7975207  0.4281744
#> 17    Site17         2     site 0.8088643  0.5175492
#> 18    Site18         1     site 0.8057851  0.4281744
#> 19    Site19         3     site 0.8209877  1.2421180
#> 20    Site20         1     site 0.8099174  0.4281744
#> 21  Species2         2  species 0.7822222 -0.9315885
#> 22  Species3         1  species 0.7500000  0.4281744
#> 23  Species4         4  species 0.8033241  1.0690450
#> 24  Species5         1  species 0.7839506  0.4281744
#> 25  Species6         3  species 0.8027682 -0.2070197
#> 26  Species7         5  species 0.7958478 -1.8333333
#> 27  Species8         1  species 0.7543253  0.4281744
#> 28  Species9         1  species 0.7962963 -2.1408721
#> 29 Species10         6  species 0.7811634 -0.5773503
#> 30 Species11         4  species 0.7337278  1.0690450
#> 31 Species12         2  species 0.7911111 -0.2070197
#> 32 Species13         1  species 0.7839506  0.4281744
#> 33 Species14         3  species 0.7968750 -0.2070197
#> 34 Species16         5  species 0.7889273 -0.3333333
#> 35 Species18         2  species 0.7922438 -0.2070197
#> 36 Species19         2  species 0.7555556 -0.9315885
#> 37 Species20         2  species 0.8024691 -0.2070197
#> 38 Species21         4  species 0.8033241  1.0690450
#> 39 Species23         5  species 0.7681661 -0.3333333
#> 40 Species24         5  species 0.7867036 -0.3333333
#> 41 Species25         3  species 0.8033241 -0.2070197
#> 42  Species1         3  species 0.8088889 -1.6561573
#> 43 Species15         6  species 0.7656250 -0.5773503
#> 44 Species17         5  species 0.7812500 -0.3333333
#> 45 Species22         1  species 0.7820069  0.4281744